Theoretical Direction of Maximum Horizontal Stress in PoR reference system

Models the direction of maximum horizontal stress \(\sigma_{Hmax}\) in the Euler pole (Pole of Rotation) coordinate reference system. When type of plate boundary is given, it also gives the deviation from the theoretically predicted azimuth of \(\sigma_{Hmax}\), the deviation, and the normalized \(\chi^2\) statistics.

Usage

PoR_shmax(df, euler, type = c("in", "out", "right", "left"))

Arguments

df

data.frame containing the coordinates of the point(s) (lat, lon), the orientation of \(\sigma_{Hmax}\) azi and its standard deviation unc (optional)

euler

data.frame containing the coordinates of the Euler pole for the plate boundary (lat, lon)

type

Character. Type of plate boundary (optional). Can be "out", "in", "right", or "left" for outward, inward, right-lateral, or left-lateral moving plate boundaries, respectively.

Value

Either a numeric vector of the azimuths in the transformed coordinate system, or a "data.frame" with

"azi.PoR"

the transformed azimuths,

"prd"

the predicted azimuths,

"dev"

the deviation, and

"nchisq"

the normalized \(\chi^2\) statistics.

Details

According to the theory, the azimuth of \(\sigma_{Hmax}\) in the pole of rotation reference system is approximate 0 (or 180), 45, 90, 135 degrees if the stress is sourced by an outward, sinistral, inward, or dextral moving plate boundary, respectively. directions of \(\sigma_{Hmax}\) with respect to the four plate boundary types.

See also

Examples

data("nuvel1")
# North America relative to Pacific plate:
euler <- subset(nuvel1, nuvel1$plate.rot == "na")

data("san_andreas")
res <- PoR_shmax(san_andreas, euler, type = "right")
head(res)
#>      azi.PoR prd         dev       nchisq
#> 1 133.099514 135   -1.900486 0.0004459072
#> 2 145.655205 135   10.655205 0.0140164678
#> 3   8.999536 135 -126.000464 0.3599938109
#> 4  38.553903 135  -96.446097 0.8618832902
#> 5  77.882633 135  -57.117367 0.4027646442
#> 6  12.393901 135 -122.606099 0.4066740573