Median and statistics on Pi-periodic Data

Calculate the mean, median, quartile, interquartile range, variance, deviation, and error of orientation data.

Usage

circular_quasi_median(x, quiet = TRUE)

circular_mean(x, quiet = TRUE)

circular_quasi_quartile(x, quiet = TRUE)

circular_quasi_IQR(x, quiet = TRUE)

circular_var(x, quiet = TRUE)

circular_mean_deviation(x, quiet = TRUE)

circular_median_deviation(x, quiet = TRUE)

circular_mean_error(x, quiet = TRUE)

Source

median(), quantile(), and IQR() are the equivalents for non-periodic data.

Arguments

x

Numeric vector (NA values will be removed).

quiet

logical. If false, a warning message is printed if there are NA values.

Value

Numeric vector

Details

Quasi median on the circle, quasi quartiles on a circle, quasi interquartile range on a circle.

References

Ratanaruamkarn, S., Niewiadomska-Bugaj, M., Wang, J.-C. (2009). A New Estimator of a Circular Median. Communications in Statistics - Simulation and Computation, 38(6), 1269--1291. doi: 10.1080/03610910902899950 .

Reiter, K., Heidbach, O., Schmitt, D., Haug, K., Ziegler, M., & Moeck, I. (2014). A revised crustal stress orientation database for Canada. Tectonophysics, 636, 111-124. doi: 10.1016/j.tecto.2014.08.006

Examples

x <- c(0, 45, 55, 40 + 180, 50 + 180)
circular_mean(x)
#> [1] 40.22302
circular_quasi_median(x)
#> [1] 55
circular_quasi_quartile(x)
#>       0%      25%      50%      75%     100% 
#>  0.00000 10.79908 45.00000 51.26613 50.00000 
circular_quasi_IQR(x)
#> [1] 40.46705
circular_var(x)
#> [1] 0.8
circular_mean_deviation(x)
#> [1] 145
circular_median_deviation(x)
#> [1] 5
circular_mean_error(x)
#> [1] 145