Compact smoothed stress field

Filter smoothed stress field containing a range of search radii or kernel half widths to find smallest wavelength (R) with the least circular sd. or dispersion for each coordinate, respectively.

Usage

compact_grid(x, type = c("stress", "dispersion"))

Arguments

x

output of stress2grid(), PoR_stress2grid(), or kernel_dispersion()

type

character. Type of the grid x. Either "stress" (when input is stress2grid() or PoR_stress2grid()) or "dispersion" (when input is kernel_dispersion()).

Value

sf object

See also

Examples

data("san_andreas")
res <- stress2grid(san_andreas)
compact_grid(res)
#> Simple feature collection with 40 features and 7 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -124.57 ymin: 24.33 xmax: -107.07 ymax: 39.33
#> Geodetic CRS:  WGS 84
#> First 10 features:
#>      R     lon   lat        azi        sd       mdr   N              geometry
#> 1  200 -107.07 24.33 177.589942 11.206266 0.8997897  28 POINT (-107.07 24.33)
#> 2  150 -107.07 34.33  23.617647  3.838594 0.7388223   3 POINT (-107.07 34.33)
#> 3  900 -107.07 36.83  22.539255 21.314908 0.8125780 157 POINT (-107.07 36.83)
#> 4  300 -107.07 39.33 109.014489  9.402220 0.7586065   5 POINT (-107.07 39.33)
#> 5   50 -109.57 24.33 162.095475 15.752508 0.8796419   6 POINT (-109.57 24.33)
#> 6  100 -109.57 26.83 177.478341  6.197443 0.8414019   4 POINT (-109.57 26.83)
#> 7  250 -109.57 29.33  26.287607  7.653565 0.8070561   3 POINT (-109.57 29.33)
#> 8  150 -109.57 31.83  24.081554  8.001669 0.6061889   4 POINT (-109.57 31.83)
#> 9  600 -109.57 34.33   9.440376 22.251337 0.7945751 115 POINT (-109.57 34.33)
#> 10 200 -109.57 36.83 124.358340 18.161864 0.8157291   5 POINT (-109.57 36.83)