Distance from plate boundary

Absolute distance of data points from the nearest plate boundary in degree

Usage

distance_from_pb(x, ep, pb, tangential = FALSE)

Arguments

x, pb

sf objects of the data points and the plate boundary geometries in the geographical coordinate system

ep

data.frame of the geographical coordinates of the Euler pole (lat, lon)

tangential

Logical. TRUE for tangential boundaries and FALSE (the default) for inward and outward boundaries.

Value

Numeric vector of the great circle distances in degree

Details

The distance to the plate boundary is the longitudinal or latitudinal difference between the data point and the plate boundary (along the closest latitude or longitude) for inward/outward or tangential plate boundaries, respectively.

Examples

data("nuvel1")
na_pa <- subset(nuvel1, nuvel1$plate.rot == "na")

data("PB2002")
san_andreas <- subset(PB2002, PB2002$PlateA %in% c("NA", "PA") & PB2002$PlateB %in% c("NA", "PA"))

data("wsm2016")
california <- subset(
  wsm2016,
  wsm2016$lat >= 23 & wsm2016$lat <= 40 &
    wsm2016$lon >= -126 & wsm2016$lon <= -108
)
california <- sf::st_set_crs(sf::st_as_sf(california, coords = c("lon", "lat")), "WGS84")

distance_from_pb(x = california, ep = na_pa, pb = san_andreas, tangential = TRUE)
#>    [1] 71.08406 74.67965 71.63611 71.37800 73.09744 75.37766 75.71422 74.11596
#>    [9] 75.15616 73.92107 72.49081 69.66631 70.38475 67.93790 69.52686 72.61742
#>   [17] 73.41923 73.18080 73.11238 73.28311 73.37605 72.97736 72.46613 71.83284
#>   [25] 71.85733 71.47498 72.10596 71.63319 72.09553 72.09190 71.85175 71.79433
#>   [33] 73.52654 71.92377 71.99356 75.32175 75.32175 74.72785 68.73800 69.13538
#>   [41] 68.91093 71.02164 73.23658 70.68649 71.26662 70.90738 70.90738 71.20926
#>   [49] 69.26616 69.55990 72.09662 72.09173 70.77632 69.48515 70.02578 70.45872
#>   [57] 70.96543 71.62197 71.62197 71.84858 71.90516 70.67821 71.98068 71.92397
#>   [65] 72.65480 72.48460 72.48460 72.29179 71.92383 71.86784 71.69641 71.78042
#>   [73] 71.75936 71.86449 72.03616 72.01804 71.85035 71.80132 71.33440 71.94130
#>   [81] 71.87832 72.05723 72.00755 74.72785 67.90240 68.73800 68.68355 69.13538
#>   [89] 68.96602 68.36733 68.85405 69.18497 68.91093 71.08455 70.25967 71.02164
#>   [97] 71.08916 69.46833 70.68649 70.80856 70.33846 71.26662 71.10550 69.14679
#>  [105] 71.20926 69.33587 69.26403 69.98943 69.55990 70.77632 71.01939 69.48515
#>  [113] 70.28171 70.15433 69.51264 71.05651 69.99589 70.02578 69.49310 70.45872
#>  [121] 71.23251 70.87476 70.93245 69.92135 71.93306 70.89510 69.70959 71.01061
#>  [129] 71.23261 71.33858 71.33508 71.41145 71.44931 71.53587 70.96543 71.58714
#>  [137] 71.64973 71.67827 71.57106 71.82289 71.44309 71.97965 71.92577 71.18857
#>  [145] 71.86448 71.54661 71.25977 71.90516 71.92040 70.67821 71.59495 71.60765
#>  [153] 71.52825 71.93396 73.13157 72.53402 72.78390 71.71603 72.85821 72.13343
#>  [161] 72.77183 72.56681 72.14245 72.15910 72.87736 72.12747 71.33250 71.27535
#>  [169] 71.20246 71.20065 71.89958 72.29179 73.37332 67.85570 70.66083 71.38564
#>  [177] 72.19973 71.61601 73.01034 72.59105 73.05097 73.05097 71.81852 70.96016
#>  [185] 70.89592 73.01501 70.89451 70.78975 72.02210 72.16199 72.22100 73.48101
#>  [193] 81.05996 76.44954 77.34073 79.64177 74.91464 75.79746 79.77185 76.56373
#>  [201] 68.92924 69.11839 74.67727 70.89196 71.34545 79.67102 76.59911 76.70178
#>  [209] 77.76031 72.62481 69.24156 69.19638 69.04572 69.12894 70.35170 74.03272
#>  [217] 74.19389 74.31670 74.36921 74.38927 74.39454 74.27394 74.29037 74.24756
#>  [225] 74.21311 74.07280 74.12931 74.21697 74.12878 74.07518 74.07265 74.30171
#>  [233] 70.82269 70.50923 69.72179 69.69877 68.66123 68.24181 68.23056 67.85532
#>  [241] 70.68164 71.12908 69.51793 69.30448 70.71394 74.60918 76.56378 76.49191
#>  [249] 74.02790 71.04925 72.00169 73.40935 69.50420 69.82464 71.75947 77.54083
#>  [257] 81.83079 81.10521 80.30443 80.33676 77.27300 78.66539 79.52674 79.85402
#>  [265] 76.41791 76.47797 81.16865 80.63240 81.03990 81.16501 75.31812 75.83826
#>  [273] 70.47814 68.88375 72.04385 73.90086 71.27272 70.08265 75.55004 73.01012
#>  [281] 73.08142 72.50282 73.47108 79.50880 78.97473 79.22557 79.17665 78.96603
#>  [289] 75.68218 75.21057 75.04534 76.61447 77.33509 76.31103 75.20480 77.92382
#>  [297] 73.35110 72.76979 72.94127 72.46613 73.06444 70.49006 70.45530 74.18977
#>  [305] 70.51095 73.85724 73.56261 72.88333 71.91668 69.06487 69.01510 69.03140
#>  [313] 71.28145 69.77090 70.16819 70.48433 70.78630 71.68498 71.13584 71.22555
#>  [321] 71.30165 71.66207 71.83974 71.74637 71.79294 72.00560 73.20146 72.56392
#>  [329] 72.65928 75.11264 78.49813 76.74701 75.58865 74.64766 76.99745 77.84254
#>  [337] 77.08052 76.77629 76.10260 77.94460 78.49049 77.53637 76.65923 79.43365
#>  [345] 72.73748 72.90849 73.07912 72.66763 72.83906 79.22013 79.22013 81.36534
#>  [353] 77.34073 76.25884 77.65668 77.91450 78.82599 78.82599 78.90354 81.49277
#>  [361] 79.41608 79.41608 79.52575 77.51368 79.73470 77.61027 76.99512 75.59898
#>  [369] 75.23448 80.66676 75.01904 75.01502 77.53340 78.56110 74.99372 75.57073
#>  [377] 75.28711 70.62098 74.03279 74.21697 74.25134 68.66650 70.02736 71.71623
#>  [385] 71.34886 76.26219 75.43089 76.18664 71.38564 69.27333 76.44965 71.05413
#>  [393] 75.73027 68.03250 72.57727 72.40553 80.84689 80.83007 80.36512 80.04171
#>  [401] 81.13390 71.18749 71.11600 71.18821 71.14975 71.08917 70.88116 70.60975
#>  [409] 68.97214 68.52595 69.83578 70.85858 72.41284 71.02984 81.63112 81.89268
#>  [417] 80.72262 77.33191 71.48729 71.46794 71.15588 71.13647 71.10960 71.21407
#>  [425] 71.17908 71.15187 71.29455 69.09527 74.01678 72.50603 72.69229 81.41496
#>  [433] 80.67689 81.55861 81.69391 79.76508 75.31777 75.68458 75.43808 73.52553
#>  [441] 82.21727 82.21836 80.90247 79.86687 74.36056 69.25987 75.79102 74.89816
#>  [449] 75.23860 75.09248 75.13070 75.07270 75.08794 74.97219 74.98743 75.00299
#>  [457] 75.01825 74.92969 75.02253 75.05311 74.93287 74.94846 74.96373 74.82038
#>  [465] 74.85092 74.91342 74.91743 74.86278 74.87847 74.82377 74.53076 74.59376
#>  [473] 74.70648 74.72251 74.72618 74.85233 74.68302 74.62808 74.64383 74.36495
#>  [481] 74.39621 74.64822 74.43124 74.58884 74.68886 74.75268 74.34431 74.35999
#>  [489] 74.37535 74.53379 74.53818 74.60185 74.25694 74.27229 74.28865 74.30402
#>  [497] 74.32006 74.49448 74.21722 74.26404 74.43931 74.45548 74.14173 74.17312
#>  [505] 74.26817 74.39553 74.45926 74.52381 74.17715 74.22437 74.39992 74.41612
#>  [513] 73.95852 74.53308 74.04167 74.05704 74.07309 74.08948 74.12096 74.13670
#>  [521] 74.15280 74.31258 74.32844 74.34464 74.36019 74.37674 73.96609 74.15693
#>  [529] 74.38091 74.41310 74.44533 73.96928 73.98533 74.01712 74.03320 74.04861
#>  [537] 74.06503 74.08047 74.27307 74.28862 73.78236 73.86506 73.89680 73.91286
#>  [545] 73.92926 73.94466 74.07224 74.08873 74.23354 74.24979 74.01279 74.10914
#>  [553] 73.82423 73.84030 73.85671 74.03243 74.04861 74.21026 74.24672 74.31208
#>  [561] 73.78407 73.79948 73.96054 73.97639 73.99259 74.00880 74.13779 73.86762
#>  [569] 74.06149 73.74287 73.87170 73.90375 73.93618 73.96830 73.98454 74.06589
#>  [577] 74.08219 74.09783 73.81516 73.83170 73.84724 73.91253 73.92844 73.94471
#>  [585] 74.00955 74.02619 74.05885 74.07485 73.59368 73.94906 74.08009 74.11218
#>  [593] 74.14531 73.77442 73.79099 73.80723 73.82348 73.84008 73.87195 73.88857
#>  [601] 73.90486 73.96980 74.01918 74.03488 73.69777 73.52383 73.71809 73.73435
#>  [609] 73.75094 73.76720 73.78348 73.79942 73.83202 73.84834 73.86939 73.90212
#>  [617] 74.00049 74.03369 73.45008 73.46656 73.48272 73.49888 73.67797 73.69425
#>  [625] 73.75909 73.79208 73.80841 73.40889 73.42539 73.71907 73.73575 73.75211
#>  [633] 73.76847 73.62471 73.78964 73.62964 73.71144 73.74456 73.53523 73.86494
#>  [641] 73.65502 73.67143 73.54435 73.57743 73.67656 73.61496 73.63139 73.57488
#>  [649] 73.59133 73.50082 73.51728 73.53409 73.04281 73.57231 73.46064 73.41530
#>  [657] 73.42044 73.43730 73.12501 73.25766 73.34649 73.36301 73.37954 73.39642
#>  [665] 72.80141 72.83495 72.86783 73.37278 73.39038 72.87544 73.41233 72.64774
#>  [673] 72.74703 72.76366 72.77996 73.33257 73.34951 72.65130 72.68384 72.78374
#>  [681] 72.88396 73.21966 73.25315 73.28703 72.58882 72.60508 72.68822 72.56240
#>  [689] 72.57939 72.59602 72.61233 72.62934 72.64600 72.66233 72.67938 72.69606
#>  [697] 73.21810 73.23508 73.27449 73.30891 72.53705 72.55337 72.57040 72.60413
#>  [705] 72.63719 72.44076 72.70842 72.77597 73.04650 73.11509 72.46065 72.47801
#>  [713] 72.49435 72.51547 73.22820 72.43527 72.46905 72.25478 72.36344 72.39728
#>  [721] 72.43082 72.45216 72.46863 72.27002 72.61006 72.88360 72.95225 72.42704
#>  [729] 72.44424 73.00976 72.96383 72.40162 72.96970 71.87960 72.22752 72.26126
#>  [737] 72.36477 72.39899 72.91732 72.09967 72.44313 72.51231 72.72036 72.79010
#>  [745] 72.17205 72.18931 72.20691 72.22419 72.24078 72.29302 72.30998 72.32731
#>  [753] 72.34465 72.10802 72.83676 72.19906 72.21638 72.23335 72.26803 72.28572
#>  [761] 72.17397 72.26047 72.66169 72.67963 72.69723 72.19653 72.26578 72.61513
#>  [769] 72.27080 72.62136 72.65626 72.06724 72.34616 72.48679 72.22925 72.24672
#>  [777] 72.56234 72.58032 72.59796 72.61561 72.63362 72.11212 72.18227 72.25185
#>  [785] 72.18734 72.20518 72.50429 72.52194 72.53960 72.55762 72.57460 72.59229
#>  [793] 72.48084 72.49852 72.51621 72.53426 72.41607 72.44040 72.45811 72.47583
#>  [801] 72.18003 72.25120 72.32216 71.99056 72.07316 72.14502 72.22941 71.99078
#>  [809] 72.02662 71.18160 71.80167 71.94699 70.28315 70.89352 74.96097 73.25396
#>  [817] 72.92932 72.69758 72.76800 72.53518 72.60705 72.37317 72.51782 75.47594
#>  [825] 75.27723 70.57132 70.59987 70.62833 71.22509 71.23230 70.00879 77.99863
#>  [833] 75.51702 81.79018 79.41331 79.41348 81.69991 81.66359 70.25345 77.55779
#>  [841] 69.89498 74.26222 72.12989 69.21387 74.99907 67.90830 71.90276 70.63729
#>  [849] 68.15049 69.34459 80.96793 81.01258 81.02072 81.00736 80.96225 80.97960
#>  [857] 80.96793 80.96561 80.77404 80.09531 81.96768 80.55233 80.77074 81.68039
#>  [865] 79.30719 79.54376 81.31106 79.52715 79.52792 78.42050 81.63669 80.09531
#>  [873] 80.14338 80.06364 77.72815 77.74261 80.10079 76.62410 80.03869 76.66464
#>  [881] 79.99083 76.59284 76.69885 75.41548 74.72894 71.50944 77.08004 75.61790
#>  [889] 77.35520 76.30929 75.54352 73.87398 74.66645 74.36726 74.08353 74.11380
#>  [897] 70.68768 71.03365 70.18881 70.55385 70.34955 70.33536 73.69137 69.85685
#>  [905] 69.37992 68.80613 68.38997 68.20927 71.54193 69.12971 69.55083 74.02267
#>  [913] 73.05791 74.62553 78.99207 69.95938 70.15678 73.07849 72.31483 80.69260
#>  [921] 80.57924 81.29604 78.61234 78.62050 78.63045 80.77613 79.30163 80.71227
#>  [929] 81.85407 78.65152 79.34317 81.84958 79.35841 79.56669 79.02939 79.74372
#>  [937] 79.11724 79.29471 79.54045 78.79670 78.71842 78.77664 78.68650 78.71591
#>  [945] 78.72403 81.93928 79.55337 78.73008 78.92956 80.68193 81.63653 79.10457
#>  [953] 79.41309 75.37968 75.38617 75.36992 75.37644 77.04424 76.96122 75.40560
#>  [961] 69.82532 81.11569 76.74438 73.25574 79.81391 79.84983 73.88223 81.81305
#>  [969] 81.53808 74.83088 80.88919 80.89815 80.93216 77.39697 77.98383 77.26463
#>  [977] 77.43167 75.43820 72.36539 72.36539 72.36539 81.96785 81.73032 80.16025
#>  [985] 77.88311 75.48390 75.46412 75.00341 74.99354 75.75228 75.48033 75.76196
#>  [993] 74.99026 74.99026 75.03295 80.78353 74.90800 81.57785 74.29182 75.55497
#> [1001] 77.02008 75.46138 81.49745 81.43695 81.43621 81.49396 81.36399 81.38190
#> [1009] 81.36549 81.37165 82.16838 80.76942 75.52902 80.27210 75.63430 79.47342
#> [1017] 75.65213 81.02734 80.89205 80.88286 80.89480 80.99584 71.08054 79.37100
#> [1025] 79.28950 77.94597 79.21640 81.09599 81.02365 69.93323 69.39509 79.52937
#> [1033] 76.92938 77.83999 81.75903 75.47383 81.15750 75.05387 75.00667 76.86099
#> [1041] 80.74728 80.81495 80.75316 81.96654 81.93627 81.02564 79.12486 70.28187
#> [1049] 80.92190 72.99029 81.56539 80.54534 80.38381 80.52217 71.68553 71.59744
#> [1057] 76.89251 81.66861 80.46236 75.40501 67.92707 73.02138 72.26735 76.83952
#> [1065] 76.71714 76.93331 73.19777 79.22849 75.55363 81.68930 80.41027 79.18928
#> [1073] 68.61779 76.64785 76.65106 76.66623 76.73939 76.71366 81.92742 76.93749
#> [1081] 81.46892 72.50588 73.57323 73.51126 73.48522 72.06444 73.52330 70.09131
#> [1089] 75.50149 72.51370 81.89426 70.76324 70.76324 76.98484 67.93484 67.94668
#> [1097] 67.87846 67.99251 67.92993 67.98355 67.87535 67.91032 67.98055 67.95013
#> [1105] 67.94659 67.95721 67.98104 67.94581 67.93863 67.93101 67.92432 67.94300
#> [1113] 67.91714 67.87176 67.96908 67.96817 68.02806 67.99852 67.96532 67.92721
#> [1121] 67.95292 67.97671 67.98875 67.96217 67.92190 67.99590 68.02657 68.02657
#> [1129] 67.96667 67.96997 67.93954 67.94555 72.49917 72.57992 72.51954 72.51641
#> [1137] 72.51641 72.51639 72.51639 72.57933 72.57933 67.93686