Normalized Chi-Squared Test

A quantitative comparison between the predicted and observed directions of \(\sigma_{Hmax}\) is obtained by the calculation of the average azimuth and by a normalized \(\chi^2\) test.

Usage

norm_chisq(obs, prd, unc)

Arguments

obs

Numeric vector containing the observed azimuth of \(\sigma_{Hmax}\), same length as prd

prd

data.frame containing the modeled azimuths of \(\sigma_{Hmax}\), i.e. the return object from model_shmax()

unc

Uncertainty of observed \(\sigma_{Hmax}\), either a numeric vector or a number

Value

Numeric vector

Details

The normalized \(\chi^2\) test is $$ {Norm} \chi^2_i = = \frac{ \sum^M_{i = 1} \left( \frac{\alpha_i - \alpha_{{predict}}}{\sigma_i} \right) ^2} {\sum^M_{i = 1} \left( \frac{90}{\sigma_i} \right) ^2 }$$ The value of the chi-squared test statistic is a number between 0 and 1 indicating the quality of the predicted \(\sigma_{Hmax}\) directions. Low values (\(\le 0.15\)) indicate good agreement, high values (\(> 0.7\)) indicate a systematic misfit between predicted and observed \(\sigma_{Hmax}\) directions.

References

Wdowinski, S., 1998, A theory of intraplate tectonics. Journal of Geophysical Research: Solid Earth, 103, 5037-5059, doi: 10.1029/97JB03390.

Examples

data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to
# Pacific plate
point <- data.frame(lat = 45, lon = 20)
prd <- model_shmax(point, euler)
norm_chisq(obs = c(50, 40, 42), prd$sc, unc = c(10, NA, 5))
#> [1] 0.001426232