Conversion between PoR to geographical coordinate system

Transform spherical objects from PoR to geographical coordinate system and vice versa.

Usage

PoR_to_geographical(x, ep)

geographical_to_PoR(x, ep)

Arguments

x

sf object of the data points in geographical or PoR coordinate system

ep

data.frame of the geographical coordinates of the Euler pole (lat, lon)

Value

sf object of the data points in the transformed geographical or PoR coordinate system

Details

The PoR coordinate reference system is oblique transformation of the geographical coordinate system with the Euler pole coordinates being the the translation factors.

Examples

data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate

data("wsm2016")
example.geo <- sf::st_set_crs(sf::st_as_sf(wsm2016[1:10, ], coords = c("lon", "lat")), "EPSG:4326")

example.por <- geographical_to_PoR(example.geo, euler)
PoR_to_geographical(example.por, euler)
#> Simple feature collection with 10 features and 58 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 127.3626 ymin: 3.92544 xmax: 136.2982 ymax: 15.82978
#> CRS:           +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> # A tibble: 10 × 59
#>    id       site    azi type  depth quality regime locality   country date      
#>  * <chr>    <chr> <dbl> <chr> <dbl> <chr>   <chr>  <chr>      <chr>   <date>    
#>  1 wsm00001 TU31    100 FMS    14   C       NF     NA         Turkey  1964-10-06
#>  2 wsm00002 TU28    105 FMS     0   C       SS     NA         Turkey  1953-03-18
#>  3 wsm00003 TU25     87 FMS     0   C       NF     NA         Turkey  1939-09-22
#>  4 wsm00004 TU13     38 FMS    38   C       TS     NA         Turkey  1966-03-07
#>  5 wsm00005 TU12    101 FMS    16   C       NF     centralTu… Turkey  1965-06-13
#>  6 wsm00006 TU60     82 FMS    15   C       SS     NA         Turkey  1983-10-21
#>  7 wsm00007 TU54    109 FMS     6.6 C       NF     NA         Turkey  1971-05-25
#>  8 wsm00008 TU53    100 FMS     9.6 C       NF     NA         Turkey  1970-04-19
#>  9 wsm00009 TU50    123 FMS     9.9 C       NF     NA         Turkey  1969-03-25
#> 10 wsm00010 TU49    115 FMS     8.8 C       NF     NA         Turkey  1969-03-23
#> # … with 49 more variables: time <chr>, number <dbl>, sd <dbl>, method <chr>,
#> #   s1az <dbl>, s1pl <dbl>, s2az <dbl>, s2pl <dbl>, s3az <dbl>, s3pl <dbl>,
#> #   mag_int_s1 <dbl>, slopes1 <dbl>, mag_int_s2 <dbl>, slopes2 <dbl>,
#> #   mag_int_s3 <dbl>, slopes3 <dbl>, pore_magin <dbl>, pore_slope <dbl>,
#> #   mag_type <chr>, eq_mag <dbl>, mom <dbl>, mom_exp <dbl>, eq_avg <chr>,
#> #   tot_len <dbl>, sd_weight <chr>, samp_int <dbl>, top <dbl>, bot <dbl>,
#> #   no_mag_ven <dbl>, young <dbl>, poisson <dbl>, rock <chr>, age <chr>, …