Conversion between spherical PoR to geographical coordinate system

Transformation from spherical PoR to geographical coordinate system and vice versa

Usage

geographical_to_PoR(x, PoR)

PoR_to_geographical(x, PoR)

Arguments

x

Can be either a "data.frame" containing lat and lon coordinates of a point in the geographical CRS or the lat.PoR, lon.PoR) of the point in the PoR CRS, a two-column matrix containing the lat and lon coordinates, a sf object, or a raster object.

PoR

Pole of Rotation. "data.frame" or object of class "euler.pole" containing the geographical coordinates of the Euler pole

Value

object of same type of x with the transformed coordinates. If x was a data.frame, transformed coordinates are named lat.PoR and lon.PoR for PoR CRS, or lat and lon for geographical CRS).

Examples

data("nuvel1")
por <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate
data("san_andreas")
san_andreas.por <- geographical_to_PoR(san_andreas, por)
head(san_andreas.por)
#> Simple feature collection with 6 features and 9 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -79.22036 ymin: 57.12526 xmax: -77.54563 ymax: 57.76312
#> Geodetic CRS:  unnamed
#>         id   lat     lon azi unc type depth quality regime
#> 1 wsm00954 33.65 -117.67   1  26  FMF  0.00       E      S
#> 2 wsm00955 33.85 -117.92   9  28  FMF  0.00       E      S
#> 3 wsm00956 34.05 -117.92   1  25  FMF  0.00       E      S
#> 4 wsm00957 34.05 -117.67 180  22  FMF  0.00       E      S
#> 5 wsm00958 34.30 -117.41  57  19   BO  2.62       B      S
#> 6 wsm00959 34.73 -117.78   4  40   HF  0.49       D      S
#>                     geometry
#> 1  POINT (-77.54563 57.1603)
#> 2 POINT (-78.07337 57.12526)
#> 3 POINT (-78.36079 57.25074)
#> 4 POINT (-78.12052 57.41224)
#> 5  POINT (-78.23381 57.7364)
#> 6 POINT (-79.22036 57.76312)
head(PoR_to_geographical(san_andreas.por, por))
#> Simple feature collection with 6 features and 9 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -117.92 ymin: 33.65 xmax: -117.41 ymax: 34.73
#> Geodetic CRS:  WGS 84
#>         id   lat     lon azi unc type depth quality regime
#> 1 wsm00954 33.65 -117.67   1  26  FMF  0.00       E      S
#> 2 wsm00955 33.85 -117.92   9  28  FMF  0.00       E      S
#> 3 wsm00956 34.05 -117.92   1  25  FMF  0.00       E      S
#> 4 wsm00957 34.05 -117.67 180  22  FMF  0.00       E      S
#> 5 wsm00958 34.30 -117.41  57  19   BO  2.62       B      S
#> 6 wsm00959 34.73 -117.78   4  40   HF  0.49       D      S
#>                geometry
#> 1 POINT (-117.67 33.65)
#> 2 POINT (-117.92 33.85)
#> 3 POINT (-117.92 34.05)
#> 4 POINT (-117.67 34.05)
#> 5  POINT (-117.41 34.3)
#> 6 POINT (-117.78 34.73)