Error of Model's Prediction

The maximum error in the model's predicted azimuth given the Pole of rotations uncertainty and distance of the data point to the pole.

Usage

prd_err(dist_PoR, sigma_PoR = 1)

Arguments

dist_PoR

Distance to Euler pole (great circle distance, in degree)

sigma_PoR

uncertainty of the position of the Pole of rotation (in degree).

Value

numeric vector. The maximum error for azimuths prediction (in degree)

References

Ramsay, J.A. Folding and fracturing of rocks. McGraw-Hill, New York, 1967.

See also

Examples

prd_err(67, 1)
#> [1] 0.9060055

# San Andreas example:
data("nuvel1")
por <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate
data("san_andreas")
d <- PoR_distance(san_andreas, por)
prd_err(d)
#>    [1] 1.769859 1.768023 1.774618 1.783184 1.800640 1.802095 1.776316 1.761912
#>    [9] 1.762615 1.755257 1.782133 1.761423 1.782107 1.780452 1.763312 1.759568
#>   [17] 1.765357 1.765840 1.824262 1.824262 1.815155 1.780276 1.777478 1.769883
#>   [25] 1.784044 1.978590 1.780871 1.774094 1.758692 1.758692 1.771977 1.767687
#>   [33] 1.774253 1.786006 1.785762 1.784634 1.738223 1.730387 1.735391 1.756827
#>   [41] 1.761458 1.761458 1.760530 1.739973 1.766817 1.770227 1.769974 1.762454
#>   [49] 1.762454 1.748638 1.767302 1.763609 1.764480 1.763722 1.761737 1.767809
#>   [57] 1.779657 1.768495 1.763244 1.758941 1.762846 1.767668 1.763633 1.781683
#>   [65] 1.767496 1.815155 1.796643 1.780276 1.768385 1.777478 1.770524 1.801586
#>   [73] 1.728522 1.777228 1.769883 1.771626 1.753501 1.784044 1.768506 1.773353
#>   [81] 1.780871 1.756819 1.734721 1.774094 1.769296 1.757515 1.771977 1.771434
#>   [89] 1.767860 1.729227 1.774253 1.784634 1.770955 1.738223 1.745781 1.782059
#>   [97] 1.774354 1.781138 1.731343 1.730387 1.738045 1.735391 1.742803 1.772870
#>  [105] 1.772280 1.728712 1.793972 1.716292 1.773067 1.771532 1.758818 1.761996
#>  [113] 1.762401 1.767334 1.756464 1.758895 1.756827 1.758388 1.763040 1.771965
#>  [121] 1.760973 1.784246 1.760469 1.767107 1.760802 1.719625 1.767712 1.743191
#>  [129] 1.721687 1.760530 1.768593 1.739973 1.773223 1.774037 1.772965 1.767359
#>  [137] 1.769474 1.764037 1.757867 1.747656 1.757458 1.745400 1.758311 1.766087
#>  [145] 1.745831 1.746723 1.759084 1.745934 1.725087 1.719195 1.721563 1.721805
#>  [153] 1.756396 1.748638 1.800500 1.769243 1.758427 1.822317 1.761700 1.789596
#>  [161] 1.749054 1.758395 1.758395 1.720707 1.702913 1.703682 1.758069 1.713185
#>  [169] 1.721614 1.735424 1.737924 1.739475 1.764153 1.802019 1.654084 1.977195
#>  [177] 1.754257 1.799664 1.645197 2.335817 1.795004 1.805558 1.967184 1.944338
#>  [185] 1.944534 1.971969 2.274501 1.759609 1.997183 1.950723 2.193475 2.139952
#>  [193] 2.225192 2.203774 2.152258 2.193227 2.158069 2.075703 2.345416 2.431821
#>  [201] 2.257723 2.162175 2.490279 1.776513 1.757720 1.766130 1.776316 1.798221
#>  [209] 1.812203 1.801532 1.785699 1.758235 1.824903 1.856353 1.740549 1.851067
#>  [217] 1.861486 1.846503 1.875480 1.839960 1.844643 1.837397 1.842464 1.851860
#>  [225] 1.834286 1.843037 1.854324 1.885725 1.770047 2.114948 2.051797 1.967097
#>  [233] 2.094476 2.148550 2.080530 2.048068 1.972185 2.060176 2.108089 1.966703
#>  [241] 1.860955 1.992702 1.736648 1.744839 1.753104 1.743009 1.751271 1.707970
#>  [249] 1.707970 1.654604 1.654084 1.810276 1.653014 1.738100 1.719033 1.719033
#>  [257] 1.716649 1.636144 1.703701 1.703701 1.702938 1.720548 1.696726 1.725363
#>  [265] 1.728891 1.716982 1.768313 1.648903 1.760512 1.734022 1.685773 2.076865
#>  [273] 2.117987 1.929178 1.898103 1.758427 1.857898 1.855115 1.857936 1.714148
#>  [281] 1.699972 1.693838 1.707326 1.731132 1.730515 1.751028 1.750630 1.757098
#>  [289] 1.757737 1.757973 1.753087 1.752987 1.753256 1.751222 1.763434 1.752362
#>  [297] 1.809406 1.731066 1.733170 1.753484 1.747721 1.721187 1.727075 1.746245
#>  [305] 1.869930 1.792856 1.750804 1.791798 1.749774 1.750112 1.853108 1.653066
#>  [313] 1.741844 1.651912 1.760855 1.879764 1.743782 1.744118 1.743782 1.863623
#>  [321] 1.747289 1.790189 1.920124 1.713254 1.744398 1.863890 1.792856 1.758840
#>  [329] 1.983456 1.747450 1.723824 1.863952 2.057417 1.854888 1.855268 1.737046
#>  [337] 1.744114 2.046711 1.744114 2.115783 1.766899 1.745718 2.261090 1.742604
#>  [345] 1.744151 1.811001 1.810498 1.809139 1.809818 1.742967 1.751532 1.805559
#>  [353] 1.808396 1.811526 1.749844 2.035223 1.876859 1.876859 1.878564 1.876859
#>  [361] 1.724509 1.803918 1.877558 1.765452 1.756795 1.748377 1.878257 1.877865
#>  [369] 1.741593 1.797754 1.858414 1.729298 1.854406 1.773377 1.726808 1.752493
#>  [377] 1.792504 1.981571 1.724806 1.762553 1.747429 1.736272 1.806944 1.856086
#>  [385] 1.741636 1.741927 1.853476 1.805451 1.798047 1.739280 1.741884 1.740926
#>  [393] 1.741507 1.741507 1.741507 1.741550 1.741507 1.783581 1.824587 1.805406
#>  [401] 1.861899 1.746321 1.747896 1.746107 1.762746 1.743590 1.763950 1.749952
#>  [409] 1.827993 1.825124 1.825124 1.748862 1.826210 1.808229 1.784239 1.773754
#>  [417] 1.773754 1.746991 1.758652 1.744816 1.856187 1.752095 1.741173 1.743399
#>  [425] 1.723095 1.788526 1.754153 1.800740 1.800337 1.877884 1.725627 1.743181
#>  [433] 1.737287 1.759127 1.746995 1.741184 1.806185 1.824113 1.732995 1.855549
#>  [441] 1.807932 1.857283 1.858121 1.971951 1.750035 1.747736 1.853286 1.742142
#>  [449] 1.747172 1.868463 1.814334 2.108870 1.748181 1.751993 1.748518 1.812853
#>  [457] 1.865044 1.816601 1.827376 1.816191 1.819738 1.817925 1.824976 1.823161
#>  [465] 1.821312 1.819501 1.826665 1.815765 1.812157 1.822913 1.747720 1.821067
#>  [473] 1.819259 1.826145 1.741086 1.822532 1.815287 1.811609 1.746943 1.811368
#>  [481] 1.783564 1.805556 1.809542 1.809301 1.839765 1.832539 1.816198 1.814446
#>  [489] 1.810713 1.796536 1.812403 1.812143 1.810329 1.838759 1.835201 1.806650
#>  [497] 1.827846 1.810068 1.795838 1.814493 1.788842 1.831118 1.829296 1.827510
#>  [505] 1.809803 1.806134 1.799115 1.834305 1.832519 1.830735 1.828952 1.827206
#>  [513] 1.807732 1.747817 1.832165 1.826863 1.807463 1.805736 1.837312 1.833784
#>  [521] 1.738025 1.823159 1.809053 1.802050 1.795105 1.830065 1.769637 1.824742
#>  [529] 1.805395 1.803673 1.770651 1.851139 1.787873 1.838513 1.836734 1.834994
#>  [537] 1.833218 1.829706 1.827899 1.826164 1.808633 1.806841 1.805121 1.803368
#>  [545] 1.801616 1.843640 1.822470 1.798018 1.794527 1.791041 1.839870 1.838131
#>  [553] 1.834584 1.746802 1.832848 1.831077 1.829308 1.827539 1.806560 1.804808
#>  [561] 1.857344 1.844766 1.861290 1.841217 1.839481 1.837710 1.835939 1.801250
#>  [569] 1.805326 1.802177 1.803433 1.821925 1.801663 1.803600 1.820166 1.809221
#>  [577] 1.804492 1.805062 1.802780 1.747172 1.803555 1.825249 1.814777 1.842553
#>  [585] 1.840820 1.839052 1.819813 1.743357 1.818094 1.800720 1.793666 2.228818
#>  [593] 1.802103 1.786774 1.840383 1.838617 1.821209 1.819421 1.817706 1.815991
#>  [601] 1.802106 1.827952 1.807145 1.838216 1.824275 1.820774 1.817349 1.813857
#>  [609] 1.812149 1.803518 1.801816 1.800081 1.823899 1.822153 1.820406 1.813539
#>  [617] 1.811764 1.810060 1.803186 1.801453 1.798060 1.798923 1.796845 1.796296
#>  [625] 1.844507 1.806432 1.792737 1.789325 1.785919 1.747687 1.821741 1.819999
#>  [633] 1.818293 1.816588 1.814850 1.811410 1.798293 1.809675 1.807976 1.801120
#>  [641] 1.796009 1.859042 1.795802 1.859424 1.794284 1.826692 1.842059 1.821327
#>  [649] 1.819624 1.817887 1.752792 1.816185 1.814485 1.812716 1.809321 1.807625
#>  [657] 1.802287 1.800250 1.801624 1.798909 1.801759 1.788700 1.797325 1.785308
#>  [665] 1.843305 1.841558 1.839847 1.838136 1.803907 1.819209 1.817511 1.810656
#>  [673] 1.807236 1.805545 1.803240 1.841089 1.804405 1.839347 1.811442 1.736804
#>  [681] 1.808567 1.806844 1.805156 1.803470 1.801851 1.815159 1.793491 1.798159
#>  [689] 1.811575 1.803045 1.755533 1.799649 1.754885 1.818156 1.784294 1.802653
#>  [697] 1.800974 1.810957 1.807560 1.797429 1.800582 1.754885 1.798908 1.798515
#>  [705] 1.796846 1.799784 1.798118 1.730462 1.796418 1.843397 1.789576 1.803198
#>  [713] 1.797717 1.799177 1.795655 1.793964 1.822104 1.808732 1.796872 1.795217
#>  [721] 1.793563 1.801522 1.791876 1.848431 1.845078 1.841691 1.788180 1.786534
#>  [729] 1.834424 1.781411 1.854215 1.844218 1.842514 1.840843 1.786140 1.784465
#>  [737] 1.850560 1.847218 1.837214 1.737271 1.827233 1.794157 1.859771 1.790834
#>  [745] 1.729569 1.787583 1.853588 1.851918 1.843617 1.849627 1.759305 1.848000
#>  [753] 1.846302 1.844638 1.916407 1.843012 1.764455 1.841315 1.839654 1.745694
#>  [761] 1.838029 1.836335 1.785309 1.783641 1.776952 1.773697 1.845685 1.859771
#>  [769] 1.844025 1.842404 1.839091 1.835743 1.737242 1.852005 1.866078 1.825572
#>  [777] 1.819014 1.792794 1.786265 1.846717 1.845065 1.843409 1.838149 1.769626
#>  [785] 1.842790 1.839493 1.857252 1.840212 1.836928 1.833680 2.101760 1.828464
#>  [793] 1.826826 1.737558 1.842795 1.810320 1.784458 1.778032 1.824603 1.720877
#>  [801] 1.823006 1.769815 1.771223 1.765144 1.820790 1.767782 1.867397 1.737468
#>  [809] 1.827915 1.743499 1.824703 1.815085 1.811848 1.802084 1.763959 1.833713
#>  [817] 1.801672 1.746119 1.795303 1.776196 1.769874 1.823762 1.822181 1.820567
#>  [825] 1.818987 1.804961 1.817371 1.812601 1.810955 1.809377 1.801624 1.807801
#>  [833] 1.754655 1.826652 1.759957 1.815170 1.813595 1.811952 1.808804 1.807199
#>  [841] 1.803625 1.811368 1.803454 1.767152 1.752788 1.765567 1.764015 1.806305
#>  [849] 1.799973 1.754535 1.754535 1.768463 1.796534 1.765121 1.761959 1.812016
#>  [857] 1.786840 1.774275 1.794387 1.899416 1.792830 1.899763 1.764638 1.763060
#>  [865] 1.761516 1.759973 1.758398 1.801962 1.795700 1.789412 1.792279 1.899763
#>  [873] 1.790693 1.755481 1.764153 1.762611 1.761071 1.754859 1.759498 1.757925
#>  [881] 1.756387 1.760556 1.759019 1.757483 1.755916 1.763419 1.758539 1.757006
#>  [889] 1.773712 1.747868 1.755474 1.778421 1.772203 1.867752 1.766027 1.783488
#>  [897] 1.770608 1.764508 1.751787 1.748140 1.748485 1.766483 1.763415 1.829889
#>  [905] 1.754674 1.742715 1.764762 1.720135 1.864035 1.796946 1.739055 1.753862
#>  [913] 1.757236 1.755480 1.749252 1.754237 1.799530 1.747531 1.741370 1.739589
#>  [921] 1.754212 1.727504 1.821172 1.805481 1.737046 1.747415 1.742604 1.807118
#>  [929] 1.760393 2.173214 1.752994 1.802016 1.754212 1.754885 1.755209 1.747465
#>  [937] 1.754212 2.172485 1.737870 1.754212 1.737405 1.737405 1.906855 1.753539
#>  [945] 1.752221 1.752892 1.801163 1.754587 1.754911 1.719726 1.804251 1.799261
#>  [953] 1.765138 1.764523 1.869448 1.742519 1.739008 1.713849 1.755157 1.801030
#>  [961] 1.886512 1.887628 1.886512 1.887559 1.888280 1.887885 1.887164 1.905856
#>  [969] 1.891355 1.887302 1.889001 1.747063 1.741222 1.764399 1.888932 1.889328
#>  [977] 1.750012 1.747140 1.889584 1.753539 1.749318 1.752299 1.741680 1.792596
#>  [985] 1.743782 1.742623 1.760444 1.889910 1.743772 1.855212 1.743772 1.743357
#>  [993] 1.739980 1.742938 1.720529 1.801563 1.797399 1.744053 1.759676 1.807917
#> [1001] 2.171061 1.902962 1.761809 1.799548 1.801777 1.878194 1.784846 1.743782
#> [1009] 1.743447 1.865352 1.744062 1.813533 1.858676 1.762714 1.807166 2.287433
#> [1017] 1.861976 1.726145 1.759628 1.755114 1.760115 1.691791 1.751867 1.785630
#> [1025] 1.801396 1.801377 1.760311 1.808975 1.802614 1.799561 1.759626 1.753137
#> [1033] 1.819204 2.151761 1.800171 1.750614 1.766604 1.750614 1.742433 1.742433
#> [1041] 1.743725 1.764499 1.889328 1.790249 1.761971 1.801604 2.165907 1.751329
#> [1049] 1.737000 1.746130 1.738158 1.797906 1.803858 1.747278 1.861205 1.868075
#> [1057] 1.747278 1.751427 1.799048 1.758926 1.755174 2.253569 1.753396 1.747137
#> [1065] 1.799102 1.816344 1.818489 1.787326 1.776522 1.746326 1.767536 1.769252
#> [1073] 1.724955 1.748738 1.683613 1.679275 1.681016 1.793970 1.902195 1.756856
#> [1081] 2.091546 1.800897 1.744818 1.743610 1.746130 1.742646 1.747096 1.737360
#> [1089] 1.682915 1.748209 1.753565 1.742853 1.760769 1.723420 1.746732 1.739508
#> [1097] 1.746119 1.804767 1.751442 1.746119 1.739608 2.365910 1.858072 1.858768
#> [1105] 1.750803 1.745282 1.807960 1.802724 2.112563 1.904503 1.903893 1.903091
#> [1113] 1.903091 1.903492 1.749218 1.744798 1.876147 1.800347 1.746106 1.714380
#> [1121] 1.863964 1.650603 1.720529 1.859178 1.872488 1.799679 1.747736 1.768219
#> [1129] 1.804398 1.826518 1.762602 1.738290 1.742099 2.081944 1.747766 1.743399
#> [1137] 1.693660 1.744013 1.670468 1.746334 1.845570 1.808884 1.891040 1.753187
#> [1145] 2.051145 1.807737 1.992511 1.756507 1.861053 1.808355 1.745013 1.742604
#> [1153] 1.856085 1.751065 1.741434 1.761899 1.748002 2.012590 1.838183 1.775932
#> [1161] 1.747985 1.936575 1.741901 1.793826 1.750966 2.044732 2.046216 1.775920
#> [1169] 1.851722 1.709567 1.782620 1.747172 1.770656 1.769971 1.769641 1.711984
#> [1177] 1.713234 1.767131 1.838183 1.771144 1.792848 1.777392 1.814193 1.746518
#> [1185] 1.746319 1.775642 1.742980 1.760286 1.799292 1.905429 1.746130 1.746130
#> [1193] 1.802079 1.864027 1.756666 1.720529 1.764149 2.277819 2.059349 2.055234
#> [1201] 1.813399 2.056664 2.058914 2.053741 1.752595 1.747172 2.002847 1.800708
#> [1209] 1.751325 1.751954 1.751021 1.751041 1.751360 1.751081 1.745264 1.741593
#> [1217] 1.861381 1.765673 1.769985 1.772425 1.772140 1.766905 1.776294 1.859217
#> [1225] 1.859217 1.755643 1.802195 1.741979 1.856362 1.855982 1.752246 1.712634
#> [1233] 1.738025 2.058475 1.737306 1.738402 1.863019 2.257290 1.755289 1.769003
#> [1241] 1.745651 1.807917 1.765109 1.767542 1.763345 1.764681 1.763345 1.809322
#> [1249] 1.765314 1.901755 2.050268 1.799574 1.751437 1.744053 1.760324 1.750908
#> [1257] 1.810694 1.745760 1.810016 1.808211 1.754377 1.763673 1.767076 1.798935
#> [1265] 1.756040 1.804024 1.889584 1.754301 1.761709 1.751179 1.755767 1.814478
#> [1273] 1.814128 1.822555 1.803135 1.759385 1.810432 1.810780 1.749203 1.799364
#> [1281] 2.023221 2.019987 1.802865 1.802865 1.800569 1.801973 2.023934 1.767543
#> [1289] 1.798590 1.811476 1.813414 2.148497 1.743813 1.751842 1.860848 1.814664
#> [1297] 1.744966 1.800684 1.812601 1.800277 1.800672 1.803697 1.798910 1.743893
#> [1305] 1.760429 1.859885 1.752377 1.755139 1.800596 1.742524 1.800155 1.893567
#> [1313] 1.799944 1.814608 1.762411 1.803488 1.751246 1.814908 1.751585 1.751246
#> [1321] 1.751246 1.751585 1.751301 1.750963 1.803466 1.999301 1.997985 1.751301
#> [1329] 1.879260 1.836421 1.836421 1.836766 1.796913 1.796913 1.799944 1.807661
#> [1337] 1.794177 1.750355 2.062861 1.857030 1.802428 1.765456 1.750068 1.801349
#> [1345] 1.820168 1.858842 1.861669 1.839375 1.813319 1.856671 1.770487 1.801792
#> [1353] 1.803480 1.751302 1.799599 1.753953 1.905269 1.894761 1.799919 1.755138
#> [1361] 1.765905 1.764617 1.766593 1.765261 1.755917 1.904938 1.729974 1.750355
#> [1369] 1.753080 1.755229 1.815454 2.016209 2.017240 1.741560 1.751098 1.868879
#> [1377] 1.746732 2.054393 1.729166 1.995571 1.806069 1.935002 1.819223 1.893640
#> [1385] 1.894038 1.741216 1.755490 1.741216 1.740839 1.811813 1.747846 1.766361
#> [1393] 1.783364 1.743022 1.855656 1.855982 1.856362 1.741173 1.772946 1.741507
#> [1401] 1.855430 1.856718 1.856107 1.985238 1.744059 1.777445 1.825843 1.825843
#> [1409] 1.826562 1.747265 1.760929 1.796392 2.020814 1.801904 1.742465 1.807223
#> [1417] 1.749133 1.782466 1.746257 1.745921 1.760728 1.971153 1.894363 1.725301
#> [1425] 1.749745 1.882763 1.880732 1.881193 1.881516 1.820314 1.908105 1.880086
#> [1433] 1.880340 1.751907 1.742245 1.801054 1.800940 1.761988 1.836077 1.753604
#> [1441] 1.751455 1.812614 1.740674 1.737692 1.812275 1.741808 1.759860 1.761116
#> [1449] 1.754587 1.801034 1.800545 1.754237 1.740535 1.737046 1.745822 1.753056
#> [1457] 1.755711 1.811476 1.758159 1.751634 1.741173 1.848366 1.747191 1.867050
#> [1465] 1.865891 2.041849 1.858382 1.858765 1.850504 1.811077 1.850883 2.055776
#> [1473] 1.943697 1.757477 1.760332 1.869871 1.763766 1.730192 1.873621 1.751907
#> [1481] 2.025033 1.749686 1.757037 1.745850 1.745850 1.801759 1.749155 1.754797
#> [1489] 1.747365 1.749662 1.755711 1.755711 1.719540 1.867852 1.755028 1.755357
#> [1497] 1.753613 1.741303 1.742476 1.800917 1.761314 1.810746 1.793149 1.793493
#> [1505] 1.793493 1.857115 1.857496 1.754908 1.743336 2.174620 1.732354 1.727695
#> [1513] 1.763522 1.745111 2.006631 1.756272 1.813383 1.759711 1.802067 1.745209
#> [1521] 1.772447 2.138241 1.761792 1.802301 1.804527 1.791788 1.753067 1.747429
#> [1529] 1.812936 1.744395 1.762309 1.749662 1.765469 1.917749 1.860307 1.800391
#> [1537] 1.740533 1.771099 1.745160 1.782587 1.772799 1.747647 1.793183 2.485027
#> [1545] 1.782742 1.860691 1.807072 1.761869 1.751710 1.750672 1.811842 1.800334
#> [1553] 1.816640 1.816655 1.740709 2.140916 1.739014 1.809703 1.809389 1.754142
#> [1561] 1.803480 1.764528 1.813639 1.862689 1.871991 1.744173 2.065677 2.009600
#> [1569] 1.855536 1.812818 2.185487 1.761133 2.372673 1.897939 1.811738 1.786654
#> [1577] 1.787408 1.817723 2.064587 1.990224 1.989880 1.991647 1.982306 1.718837
#> [1585] 1.988948 1.975124 1.740508 2.002718 1.754633 1.876731 1.939830 1.735294
#> [1593] 1.748682 1.802195 2.021112 2.022037 2.055561 1.846542 1.843281 1.844246
#> [1601] 1.842627 1.845240 1.845612 1.844307 1.844867 1.842621 1.845196 1.843705
#> [1609] 1.843429 1.848051 1.845502 1.841004 1.843836 1.842912 1.845518 1.846589
#> [1617] 1.846338 1.847303 1.845171 1.847242 1.846941 1.844311 1.845876 1.840431
#> [1625] 1.844684 1.846034 1.762952 1.844122 1.845253 1.846640 1.845667 1.839211
#> [1633] 1.844991 1.848562 1.845261 1.848426 1.835230 2.020002 1.742938 1.675639
#> [1641] 1.744871 1.912934 1.731185 1.992010 1.740056 1.748185 1.741680 1.750606
#> [1649] 1.755183 1.765818 1.812685 1.812685 1.767513 1.805829 1.767513 1.786966
#> [1657] 1.790856 1.745211 1.747335 1.808627 1.812870 2.146163 1.808298 1.843716
#> [1665] 1.860211 1.843695 1.742886 1.749713 1.751710 1.759588 1.749464 1.810098
#> [1673] 1.845626 1.862993 1.898065 1.882150 2.235486 2.252153 2.323304 2.208434
#> [1681] 2.250865 1.811516 1.829866 1.824471 1.823633 1.810284 1.824717 1.791671
#> [1689] 1.752748 1.719613 1.726186 1.855324 1.736567 1.744214 1.828556 1.766138
#> [1697] 1.762950 1.763315 1.785115 1.756518 1.781510 2.033445 1.867741 1.879861
#> [1705] 1.762605 1.879995 1.744501 1.875043 1.838924 1.937244 1.968640 2.268880
#> [1713] 1.875369 1.709301 1.715899 1.761975 1.739703 1.844828 1.819823 1.917263
#> [1721] 1.917330 1.912117 1.910690 1.918279 1.918893 1.916243 1.920186 1.915837
#> [1729] 1.920796 1.908678 1.921412 1.915430 1.915363 1.909486 1.912045 1.911103
#> [1737] 1.896838 1.914483 1.925651 1.909486 1.917939 1.917872 1.911103 1.912317
#> [1745] 1.919436 1.908275 1.909214 1.905857 1.911574 1.917872 1.907132 1.918279
#> [1753] 1.916717 1.905990 1.906056 1.907535 1.923598 1.921209 1.906795 1.908811
#> [1761] 1.910227 1.909618 1.915430 1.916377 1.910227 1.907535 1.782665 1.918279
#> [1769] 1.909082 1.921070 1.909956 1.915903 1.874527 1.840797 1.844288 1.843615
#> [1777] 1.918687 1.923052 1.846063 1.909956 1.921003 1.845501 1.922095 1.921277
#> [1785] 1.917531 1.919369 1.915770 1.915296 1.856242 1.853299 1.854432 1.912251
#> [1793] 1.751973 1.915770 1.911348 2.079499 1.753051 1.921412 1.745886 1.765140
#> [1801] 1.765140 1.907601 1.763833 1.767758 1.751127 1.917939 1.871635 1.917531
#> [1809] 1.911169 1.907535 1.906992 1.916717 1.769516 1.730863 1.729725 1.747792
#> [1817] 1.919028 1.821036 1.819999 1.820337 1.819645 1.820337 1.819999 2.158364
#> [1825] 1.819267 1.751733 1.751127 1.761302 1.747628 1.763483 1.920459 1.920662
#> [1833] 1.920186 1.907805 1.907738 1.905455 1.907335 1.908142 1.906596 1.905857
#> [1841] 1.915970 1.910297 1.751523 1.913804 1.905119 1.908882 1.913671 1.747432
#> [1849] 1.747432 2.367446 1.798061 1.751644 1.920662 1.880024 1.839411 1.839783
#> [1857] 1.737136 1.800889 1.801629 1.809737 1.800808 1.756859 1.906999 1.799624
#> [1865] 1.801165 1.827947 1.823622 1.802103 1.747865 1.815193 1.814478 1.753878
#> [1873] 1.740930 1.905924 1.817647 1.915770 1.860242 1.763977 1.962559 1.868043
#> [1881] 1.916923 1.915903 2.163262 1.860704 1.743190 1.821469 1.822555 1.823963
#> [1889] 1.823259 1.823642 1.822200 1.823286 1.824366 1.824004 1.824366 1.823292
#> [1897] 1.765708 1.742983 1.766671 1.862569 1.859637 1.861592 1.863269 1.861033
#> [1905] 1.922027 1.864600 1.862884 1.899841 1.963413 1.947672 2.032494 1.803680
#> [1913] 1.808749 1.809449 1.866142 1.870504 1.807340 1.743022 1.899442 1.921003
#> [1921] 1.757954 1.863060 1.865580 1.789136 1.861208 1.869939 1.912251 1.764139
#> [1929] 1.759794 1.764820 1.754186 1.773930 1.776392 1.777284 1.776814 1.778173
#> [1937] 2.097827 1.917872 1.886406 1.914278 1.805369 1.920662 1.898773 1.909890
#> [1945] 1.785273 1.919095 2.194740 1.781203 1.746159 1.751556 1.868772 1.747164
#> [1953] 1.735831 1.778650 1.860178 1.915363 1.759418 1.824730 1.806186 1.843685
#> [1961] 2.388186 1.759113 1.806655 1.917464 1.778451 1.778593 1.780318 1.808992
#> [1969] 1.920118 2.009494 1.837103 1.815594 1.867019 1.913194 1.786998 1.744897
#> [1977] 1.746256 1.762805 1.797802 1.747315 1.743782 1.744418 1.751634 1.888729
#> [1985] 1.889856 1.889856 1.889123 1.787674 1.746085 1.753613 1.752038 1.799852
#> [1993] 1.826142 1.826142 1.741414 1.718094 1.725613 1.895172 1.878140 1.751181
#> [2001] 1.915769 1.806792 1.809902 2.018846 1.747614 1.750386 1.743823 1.743876
#> [2009] 1.739538 1.764377 1.766971 1.774617 1.730754 1.864459 1.918078 2.097515
#> [2017] 1.804365 1.739379 1.794747 1.798387 1.752312 1.883357 1.725615 1.816283
#> [2025] 1.812708 1.813058 1.862198 1.821114 1.766734 1.844843 1.801345 1.776878
#> [2033] 1.813286 1.807847 1.818126 1.818459 1.818469 1.819195 1.806625 1.763930
#> [2041] 1.765290 1.775361 1.812156 1.748045 1.750232 1.768276 2.012624 2.007585
#> [2049] 1.881492 1.769383 2.252264 1.846529 1.774164 1.769301 1.767327 1.766670
#> [2057] 1.769668 1.770465 1.949025 1.780722 2.095386 1.906416 1.904624 1.864376
#> [2065] 1.752848 1.786978 1.843058 1.860963 1.736532 1.762894 1.876594 1.784937
#> [2073] 1.892424 1.891804 1.914211 1.744447 1.900991 1.922708 1.922708 2.080973
#> [2081] 1.753582 1.753137 1.746075 1.753100 1.753140 1.752451 1.746174 1.751550
#> [2089] 1.752468 1.751203 1.751951 1.753244 1.752250 1.752011 1.751156 1.753482
#> [2097] 1.751981 1.750464 1.746534 1.752253 1.751852 1.752572 1.753549 1.752871
#> [2105] 1.751103 1.752166 1.751871 1.752842 1.753864 1.750992 1.753386 1.752504
#> [2113] 1.752504 1.751729 1.751959 1.753239 1.844082 1.846357 1.845220 1.844986
#> [2121] 1.844986 1.844972 1.844972 1.846320 1.846320 1.751587 2.503156 1.912557
#> [2129] 1.912326 1.924132 1.912006 2.155572 2.045456 2.045456 2.045456 2.045456
#> [2137] 1.924223 1.762405 1.766287 1.762066 1.760812 1.861544 1.870719 1.880850
#> [2145] 1.896345 1.872040 1.721244 1.723316 1.877898 1.855435 1.857919 1.884011
#> [2153] 1.855139 1.857995 1.862620 1.752520 1.926514 1.694488 1.826891 1.760505
#> [2161] 1.815470 1.710480 1.798812 1.790838 1.842705 1.825090 1.814503 1.800475
#> [2169] 1.839072 1.886870 1.692347 1.842650 1.758043 1.723687 1.733471 1.851901
#> [2177] 1.846280 1.755799 1.714119 1.758887 1.704191 1.722006 1.671144 1.747995
#> [2185] 1.739224 1.762229 1.789819 1.714010 1.733503 1.873851 1.741008 1.757746
#> [2193] 1.872483 1.864112 1.895783 1.872008 1.801300 1.752421 1.786419 1.816434
#> [2201] 1.805565 1.744696 1.747348 1.732368 2.233624 1.741876 1.771825 1.857948
#> [2209] 1.777379 1.778812 1.716632 1.824325 1.840052 1.805608 1.908299 1.830066
#> [2217] 1.821519 1.827609 1.832214 2.065639 1.833684 1.768316 1.880761 1.747542
#> [2225] 1.858271 1.821166 1.744108 1.854327 1.876211 1.876211 1.744117 1.746420
#> [2233] 1.747172 2.063655 2.090636 1.855771 1.884759 1.880910 1.883635 1.884026
#> [2241] 1.886123 1.850943 1.769885 1.853804 1.747702 1.830118 1.744314 1.744314
#> [2249] 1.743357 1.699073 1.853804 1.840972 1.839510 1.901442 1.747685 1.747472
#> [2257] 1.856280 1.753115 1.762877 1.765434 1.758488 1.761075 1.753816 1.758931
#> [2265] 1.760983 1.761719 1.763650 1.762669 1.755033 1.760182 1.733962 1.761625
#> [2273] 1.764111 1.764449 1.762700 1.760646 1.760646 1.760402 1.866622 1.892292
#> [2281] 1.762439 1.742433 1.864439 1.757683 1.760817 1.760855 1.760014 1.746622
#> [2289] 1.743272 1.767143 1.856946 1.835745 1.744117 1.892686 1.844484 1.749066
#> [2297] 1.767449 1.859403 1.860482 1.857250 1.858875 1.858037 1.863192 1.860409
#> [2305] 1.859403 1.852077 1.862715 1.858420 1.862111 1.865432 1.866424 1.869151
#> [2313] 1.861177 1.863245 1.870536 1.866755 1.862787 1.802862 1.863172 1.737136
#> [2321] 1.838979 1.873544 1.747551 1.747500 1.747243 1.860243 1.861488 1.862184
#> [2329] 1.760924 1.857652 1.861873 1.742938 1.826934 1.781903 1.862881 1.741622
#> [2337] 1.855662 1.742796 1.760261 1.862954 1.675719 1.679140 1.862569 1.742646
#> [2345] 1.937286 1.760979 1.829491 1.741550 1.842975 1.842975 1.749978 1.831807
#> [2353] 1.842232 1.747049 1.747001 1.840087 1.905385 1.749066 1.743357 1.879195
#> [2361] 1.737781 1.738069 1.744397 1.744929 1.744929 1.867803 1.868358 1.869520
#> [2369] 1.841923 1.854828 1.857423 1.857099 1.869373 1.857862 1.857862 1.858186
#> [2377] 1.857862 1.737378 1.737046 1.737287 1.736955 1.747450 1.866278 1.729633
#> [2385] 1.716255 1.742433 1.723015 1.896073 1.752299 1.845700 1.867779 1.870419
#> [2393] 1.867537 1.748142 1.761461 1.744228 1.744228 1.855916 1.856621 1.845046
#> [2401] 1.742891 1.762122 1.759620 1.760614 1.802907 1.747289 1.723721 1.842975
#> [2409] 1.740294 1.743564 1.744929 1.903088 1.738158 1.799155 1.873533 1.752207
#> [2417] 1.843123 1.889222 1.892660 1.887978 1.890285 1.853149 1.882598 1.857309
#> [2425] 1.884606 1.739540 1.855900 1.855842 1.855081 1.854758 1.854378 1.857805
#> [2433] 1.854929 1.857805 1.858511 1.858129 1.858511 1.867451 1.842678 1.891076
#> [2441] 1.737692 1.859178 1.859942 1.746326 1.859673 1.859942 1.858796 1.885494
#> [2449] 1.838180 1.856846 1.858909 1.858528 1.753041 1.739893 1.761233 1.811427
#> [2457] 1.882361 1.746577 1.747102 1.740674 1.855497 1.867849 1.744117 1.855842
#> [2465] 1.856902 1.729130 1.729741 1.742056 1.738941 1.746991 1.944462 1.747432
#> [2473] 1.747096 1.747472 1.853860 1.743022 1.855592 1.855821 1.855435 1.855435
#> [2481] 1.856465 1.858203 1.857496 1.858203 1.798308 1.857171 1.860820 1.810287
#> [2489] 1.925701 1.730523 1.742142 1.715409 1.745726 1.856269 1.854850 1.860065
#> [2497] 1.746424 1.746696 1.891076 1.763790 1.757797 1.758053 1.757801 1.865363
#> [2505] 1.852234 1.747265 1.824264 1.828553 1.828187 1.741507 1.741550 1.744349
#> [2513] 1.842250 1.850543 1.810331 1.808490 1.768179 1.753604 1.740294 1.867243
#> [2521] 1.747751 1.742646 1.742646 1.859506 1.746732 1.751073 1.742402 1.803218
#> [2529] 1.855099 1.744249 1.744229 1.743985 1.748548 1.808992 1.796594 1.745172
#> [2537] 1.721244 1.721816 1.719363 1.721244 1.719330 1.718780 1.720141 1.718813
#> [2545] 1.718845 1.719136 1.719428 1.724029 1.718974 1.718359 1.628519 1.631921
#> [2553] 1.765381 1.738069 1.721441 1.761466 1.754628 1.723532 1.867220 1.709185
#> [2561] 1.709825 1.797126 1.797495 1.722152 1.738550 1.722185 1.742099 1.743022
#> [2569] 1.719396 1.743314 1.877571 1.877508 1.877444 1.877054 1.877508 1.879397
#> [2577] 1.737737 1.748384 1.744901 1.747955 1.747518 1.747294 1.791015 1.746888
#> [2585] 1.919310 1.724712 1.876979 1.876652 1.740294 1.747450 1.742895 1.742938
#> [2593] 1.747472 1.742142 1.752234 2.074678 2.073830 1.756382 1.747137 1.761055
#> [2601] 1.782283 1.853108 1.747518 1.763366 1.771042 1.761065 1.743893 1.747208
#> [2609] 1.863599 1.763279 1.773348 1.810416 1.811128 1.742142 1.830095 1.874717
#> [2617] 1.769472 1.743949 1.718845 1.755413 1.756433 1.872637 1.802267 1.793003