Apply Rolling Functions using Circular Statistics

A generic function for applying a function to rolling margins of an array.

Usage

roll_circstats(
  x,
  w = NULL,
  FUN,
  axial = TRUE,
  na.rm = TRUE,
  width = NULL,
  by.column = FALSE,
  partial = TRUE,
  fill = NA,
  ...
)

Arguments

x

numeric vector. Values in degrees.

w

(optional) Weights. A vector of positive numbers and of the same length as x.

FUN

the function to be applied

axial

logical. Whether the data are axial, i.e. pi-periodical (TRUE, the default) or directional, i.e. \(2 \pi\)-periodical (FALSE).

na.rm

logical value indicating whether NA values in x should be stripped before the computation proceeds.

width

integer specifying the window width (in numbers of observations) which is aligned to the original sample according to the align argument. If NULL, an optimal width is calculated.

by.column

logical. If TRUE, FUN is applied to each column separately.

partial

logical or numeric. If FALSE then FUN is only applied when all indexes of the rolling window are within the observed time range. If TRUE (default), then the subset of indexes that are in range are passed to FUN. A numeric argument to partial can be used to determine the minimal window size for partial computations. See below for more details.

fill

a three-component vector or list (recycled otherwise) providing filling values at the left/within/to the right of the data range. See the fill argument of zoo::na.fill() for details

...

optional arguments passed to zoo::rollapply()

Value

numeric vector with the results of the rolling function.

Note

If the rolling statistics are applied to values that are a function of distance it is recommended to sort the values first.

Examples

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")
data("san_andreas")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
distance <- distance_from_pb(
  x = san_andreas,
  PoR = PoR,
  pb = plate_boundary,
  tangential = TRUE
)
dat <- san_andreas[order(distance), ]
roll_circstats(dat$azi, w = 1 / dat$unc, circular_mean, width = 51)
#>   [1]  26.96646551  28.37786435  26.07195156  26.55918729  26.58980875
#>   [6]  25.91465583  26.80333041  27.70326453  27.07830097  28.71990527
#>  [11]  29.59838411  30.84780515  30.63759657  29.77013667  28.38941331
#>  [16]  28.17546240  28.71359825  28.71359825  28.37217999  27.35100144
#>  [21]  26.92638427  26.42172323  25.68748528  26.50837294  25.83974368
#>  [26]  25.42726308  24.19804262  23.74777375  23.62440569  23.01738436
#>  [31]  22.61293299  22.86985850  23.38588573  22.40179847  21.51112273
#>  [36]  21.33882014  20.82007356  20.72699968  20.78935261  21.16340797
#>  [41]  20.81904383  21.70585855  21.73468661  20.55864602  19.66862138
#>  [46]  18.81057534  17.95045649  16.82239944  17.52265225  17.46732037
#>  [51]  17.29381388  16.55437812  16.71383374  18.04774902  16.90564632
#>  [56]  16.69174068  16.71421005  16.16108615  16.28363170  16.63767677
#>  [61]  16.76513558  16.45636270  15.42334533  15.74280028  15.91597731
#>  [66]  17.12665379  17.48474189  17.14056365  17.35935090  17.13456678
#>  [71]  17.99013544  17.99013544  18.10244079  18.51376979  18.09546775
#>  [76]  18.84670842  19.69197062  20.67812133  21.62951628  21.54138508
#>  [81]  21.82028452  22.35811105  22.16981917  22.53630394  23.33543377
#>  [86]  24.25008100  23.90376945  24.27947218  24.28786478  24.14467732
#>  [91]  23.70650554  23.32500515  22.91121551  22.57339870  22.83425432
#>  [96]  23.54893463  24.19884772  23.92514092  24.29853830  23.86642511
#> [101]  23.10447830  22.57396151  23.13146734  21.70335076  21.52555426
#> [106]  21.43690915  21.89270864  22.22226604  22.84440679  22.57718381
#> [111]  22.48977977  21.34424454  20.47023634  20.79798510  19.69950328
#> [116]  19.01173865  17.87475071  17.85340765  17.10229258  16.26888871
#> [121]  16.05270156  15.43961334  15.94592596  15.26613230  14.57227462
#> [126]  14.27371915  13.88068933  13.55058635  12.79254939  12.62674338
#> [131]  12.60590409  12.05428759  11.49016567  11.89754458  11.29962880
#> [136]  11.05920826  10.33378055   9.79338719   9.34735513   8.93408436
#> [141]   9.05159455   8.18302882   7.86839140   6.88921060   6.16065538
#> [146]   5.47829902   5.23710101  16.20504274  16.27433970  19.26083061
#> [151]  19.12038531  19.16936084  23.06620806  23.22813619  23.40461875
#> [156]  22.77386334  22.74816619  22.16076161  21.95322520  21.56194453
#> [161]  20.93452982  20.84623474  20.89114257  21.24480093  20.81960874
#> [166]  20.91731080  20.93871874  21.33520416  20.72003502  20.94264198
#> [171]  21.34703223  21.17209716  20.93538743  20.30558281  20.37887051
#> [176]  20.74516954  20.51764526  20.34077824  20.00228952  19.96085307
#> [181]  19.62325326  19.62325326  19.49406301  19.73334097  19.61784990
#> [186]  19.56922499  19.32284190  19.38950073  19.62116938  19.39278819
#> [191]  19.81112711  19.41728343  19.45968167  19.83505934  19.78863269
#> [196]  19.79825410  20.20105560  20.23610308   9.88034277  13.39797198
#> [201]   9.23482443   9.19268364   9.08628454   6.23673351   5.47069905
#> [206]   6.60447387   6.81745371   6.78684877   6.75256716   6.51837788
#> [211]   6.02674028   6.60803976   6.52694531   6.35153725   6.24952790
#> [216]   6.31088473   6.58726044   6.87622116   6.47209443   6.43981421
#> [221]   6.05229194   5.75985316   5.83575222   5.71429295   5.63528012
#> [226]   5.58576773   5.35327971   5.27646456   6.40060893   6.56583724
#> [231]   6.65695040   6.48181466   6.55612047   6.56861906   6.93226008
#> [236]   6.93226008   6.91806124   6.87615517   6.75705636   6.50453281
#> [241]   6.46212990   6.28906539   6.51012555   6.76768289   6.38070301
#> [246]   6.93684942   7.08120529   7.23786719   7.38049737   7.80638169
#> [251]   5.05990086   5.05990086   5.53780870   5.59851806   3.28746555
#> [256]   3.40905786   2.00301572   2.30620917   2.55052332   2.55052332
#> [261]   2.89299017   3.47308703   3.33729475   3.63995660   3.78098785
#> [266]   3.99802485   3.96268374   3.44433740   3.33920761   3.32013102
#> [271]   3.33856930   3.65746625   3.83601844   3.96711733   4.74683100
#> [276]   5.43003080   5.35983728   4.78781105   5.35115433   4.10707732
#> [281]   4.08427290   4.12763093   4.28068318   4.08418863   3.97385335
#> [286]   3.10087818   2.61022131   2.61022131   2.36894679   2.16707980
#> [291]   2.29604201   2.93944125   3.29892639   3.70453463   4.15013615
#> [296]   4.30314380   3.96833182   4.28418037   4.23643578   4.89493196
#> [301]   4.87344696   5.03315567   5.37967617   4.91577217   4.80390894
#> [306]   5.22216627   5.08459059   4.38087124   3.86361112   3.65084957
#> [311]   3.51406216   3.32267853   2.29622519   1.38440383   1.36065987
#> [316]   1.01586490 179.87232663   0.04050628 179.97097503   0.45004957
#> [321]   0.37724298   0.06716522 179.45960688 179.33716876 178.79362811
#> [326] 177.94422304 177.29386854 177.43859505 178.76950417 178.53811255
#> [331] 178.07189822 178.29311202 178.71323783 178.46936110 178.35267096
#> [336] 178.84825603 178.29480339 178.00200876 178.00200876 179.18330973
#> [341] 179.98217458   0.53900585   0.07740335 179.52853999 178.20611299
#> [346] 177.88600039 177.71000023 177.31089038 176.59311935 176.56551135
#> [351] 175.86279894 175.21755859 175.17111178 175.13097574 174.74731606
#> [356] 174.93650822 174.55426276 174.06024068 175.05450081 174.87047572
#> [361] 174.49131917 174.39748513 174.50822813 174.76570833 175.03582809
#> [366] 175.25007971 175.20973248 176.36467303 176.36467303 176.43677246
#> [371] 176.05261342 175.62358555 176.68049994 177.73648471 177.29564116
#> [376] 177.55729658 177.43866653 177.04370207 177.22686524 176.76076608
#> [381] 176.09799300 176.34689029 175.91761905 175.21625417 175.50543789
#> [386] 175.55170002 175.55170002 176.49778035 177.07177781 176.40865052
#> [391] 175.66951269 175.40676393 175.62305838 175.75984348 176.76637393
#> [396] 178.27551230 178.59047086 178.83221391 179.05733548 179.48967454
#> [401] 178.73130892 178.83038268 179.26239202 179.11693371 178.15144419
#> [406] 178.91254577 178.82323492 178.86005919 179.92083652 179.19243911
#> [411] 179.91414977   0.97376240   1.69884283   1.34256492   2.11503821
#> [416]   2.74515980   1.76685185   3.25433886   1.32896397   1.57524539
#> [421]   2.32439215   3.19422579