Apply Rolling Functions using Circular Statistical Tests for Uniformity

A generic function for applying a function to rolling margins of an array.

Usage

roll_normchisq(
  obs,
  prd,
  unc = NULL,
  width,
  by.column = FALSE,
  partial = TRUE,
  fill = NA,
  ...
)

roll_rayleigh(
  obs,
  prd,
  unc = NULL,
  width = NULL,
  by.column = FALSE,
  partial = TRUE,
  fill = NA,
  ...
)

roll_dispersion(
  obs,
  prd,
  unc = NULL,
  width = NULL,
  by.column = FALSE,
  partial = TRUE,
  fill = NA,
  ...
)

Arguments

obs

Numeric vector containing the observed azimuth of \(\sigma_{Hmax}\), same length as prd

prd

Numeric vector containing the modeled azimuths of \(\sigma_{Hmax}\), i.e. the return object from model_shmax()

unc

Uncertainty of observed \(\sigma_{Hmax}\), either a numeric vector or a number

width

integer specifying the window width (in numbers of observations) which is aligned to the original sample according to the align argument. If NULL, an optimal width is calculated.

by.column

logical. If TRUE, FUN is applied to each column separately.

partial

logical or numeric. If FALSE then FUN is only applied when all indexes of the rolling window are within the observed time range. If TRUE (default), then the subset of indexes that are in range are passed to FUN. A numeric argument to partial can be used to determine the minimal window size for partial computations. See below for more details.

fill

a three-component vector or list (recycled otherwise) providing filling values at the left/within/to the right of the data range. See the fill argument of zoo::na.fill() for details

...

Arguments passed on to zoo::rollapply

Value

numeric vector with the test statistic of the rolling test.

numeric vector

Note

If the rolling functions are applied to values that are a function of distance it is recommended to sort the values first.

Examples

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")
data("san_andreas")
ep <- subset(nuvel1, nuvel1$plate.rot == "na")
distance <- distance_from_pb(
  x = san_andreas,
  euler = ep,
  pb = plate_boundary,
  tangential = TRUE
)
dat <- san_andreas[order(distance), ]
dat.PoR <- PoR_shmax(san_andreas, ep, "right")
roll_normchisq(dat.PoR$azi.PoR, 135, dat$unc, width = 51)
#>    [1] 0.030365725 0.030363120 0.031241592 0.026031701 0.024639812 0.031773432
#>    [7] 0.031897071 0.031903592 0.031865219 0.031837036 0.031706470 0.031809372
#>   [13] 0.046208204 0.046144602 0.046083454 0.044718631 0.044660376 0.044619904
#>   [19] 0.043965969 0.040287380 0.040238858 0.040211512 0.040163667 0.037286698
#>   [25] 0.037142490 0.034625310 0.034525097 0.034534470 0.033889260 0.031661265
#>   [31] 0.031604645 0.031556721 0.031630795 0.031645406 0.031252875 0.031234478
#>   [37] 0.030763075 0.029907703 0.028816181 0.028761617 0.031658048 0.030001654
#>   [43] 0.028736155 0.028687609 0.027115857 0.027086505 0.027088849 0.026992539
#>   [49] 0.027620766 0.027625734 0.027247895 0.027249433 0.027234578 0.025570280
#>   [55] 0.027175376 0.028004242 0.023471898 0.023265543 0.023244414 0.023248829
#>   [61] 0.023000482 0.023132450 0.023011425 0.012077014 0.012095748 0.017371351
#>   [67] 0.017550815 0.017579959 0.018334318 0.018486414 0.019115088 0.018701320
#>   [73] 0.028112480 0.028127164 0.029228374 0.027975228 0.029189518 0.029291482
#>   [79] 0.029333145 0.029653576 0.032311111 0.032316347 0.032313161 0.032353497
#>   [85] 0.032391887 0.036737417 0.037607428 0.034337255 0.034330191 0.035254161
#>   [91] 0.035566879 0.034034752 0.035561581 0.037239245 0.037254418 0.037417296
#>   [97] 0.037544450 0.037535804 0.037511261 0.037507615 0.035865351 0.036273204
#>  [103] 0.036297699 0.037689687 0.039666280 0.039579381 0.044740483 0.044734186
#>  [109] 0.044684589 0.044684018 0.044678615 0.044399517 0.044346376 0.044681688
#>  [115] 0.044600157 0.051685558 0.063076923 0.063077288 0.063093080 0.063537538
#>  [121] 0.062183269 0.063033178 0.063597839 0.053267993 0.053252812 0.053254506
#>  [127] 0.056260626 0.056260459 0.058240300 0.058366329 0.037282845 0.036729226
#>  [133] 0.036723223 0.036724507 0.036662844 0.035910435 0.032313773 0.030797317
#>  [139] 0.033672552 0.033702159 0.036743000 0.036747413 0.036307650 0.036365763
#>  [145] 0.034983263 0.025877177 0.025729778 0.024155846 0.024154701 0.023521128
#>  [151] 0.023521346 0.024106212 0.022755713 0.022486534 0.021330466 0.021008567
#>  [157] 0.020063553 0.015973709 0.015970923 0.015978268 0.015973302 0.015973445
#>  [163] 0.016108446 0.016113659 0.016216756 0.016305722 0.009817073 0.009769534
#>  [169] 0.009768706 0.009753709 0.009749699 0.006930219 0.006748160 0.006554135
#>  [175] 0.006570411 0.006837818 0.006836721 0.006841597 0.006843189 0.006873911
#>  [181] 0.006822448 0.008399837 0.008730896 0.008790576 0.008795216 0.008778337
#>  [187] 0.007448095 0.007553131 0.007678748 0.007729013 0.007708889 0.003412976
#>  [193] 0.003592728 0.005028695 0.004993731 0.006004450 0.011933973 0.011998226
#>  [199] 0.014563443 0.014608283 0.015194495 0.015197226 0.019496592 0.021308355
#>  [205] 0.022116130 0.022636581 0.022735146 0.047214291 0.047217809 0.062898609
#>  [211] 0.062947316 0.061685463 0.061685103 0.061961387 0.061962041 0.061942386
#>  [217] 0.061851501 0.060074437 0.060085137 0.060065302 0.059778483 0.059759979
#>  [223] 0.059736790 0.060457527 0.060437234 0.060342859 0.057893136 0.057527467
#>  [229] 0.057243157 0.057266127 0.057577295 0.057586530 0.054799824 0.052122901
#>  [235] 0.051917636 0.051923936 0.050991070 0.062911905 0.062809839 0.061123523
#>  [241] 0.061082358 0.061103176 0.061096698 0.061326353 0.063595317 0.063583212
#>  [247] 0.059561023 0.060045615 0.059968567 0.062014630 0.062034575 0.062043372
#>  [253] 0.062057771 0.059504374 0.057841954 0.058660839 0.058772502 0.059568043
#>  [259] 0.044635912 0.043831554 0.022425410 0.021774252 0.021801368 0.021853220
#>  [265] 0.023873891 0.025772414 0.025358061 0.025356000 0.026131662 0.026122101
#>  [271] 0.023292971 0.023473994 0.045818595 0.048760511 0.048627652 0.048407228
#>  [277] 0.048478333 0.050235141 0.050467152 0.050583591 0.050546846 0.051118680
#>  [283] 0.051125845 0.053890781 0.057358700 0.056787439 0.057294981 0.058611862
#>  [289] 0.058616410 0.057119143 0.058483464 0.058480100 0.055916486 0.055859018
#>  [295] 0.055852337 0.055861002 0.055867617 0.056073995 0.055031222 0.055074700
#>  [301] 0.055071985 0.055094515 0.055082718 0.055081489 0.055014215 0.056637932
#>  [307] 0.056642115 0.055452335 0.055485540 0.055114775 0.068230339 0.076052906
#>  [313] 0.069740432 0.069038254 0.073702230 0.073084375 0.069822966 0.070664071
#>  [319] 0.070696612 0.070810432 0.068278205 0.069115947 0.069109145 0.060399440
#>  [325] 0.058449875 0.057303274 0.057861886 0.057844925 0.057956575 0.057988218
#>  [331] 0.060161146 0.059567170 0.058933796 0.058935865 0.058788358 0.057722297
#>  [337] 0.059614794 0.059323671 0.058906635 0.059022643 0.060086264 0.060295693
#>  [343] 0.060301953 0.062481233 0.061522915 0.061517951 0.061963154 0.061962845
#>  [349] 0.064857355 0.064423703 0.064454209 0.063779024 0.062892228 0.078783756
#>  [355] 0.078776303 0.078784563 0.078648479 0.078847441 0.080719440 0.080680843
#>  [361] 0.081489098 0.071616893 0.071532901 0.082524274 0.088368249 0.084192686
#>  [367] 0.087662725 0.095875214 0.091370004 0.083064831 0.082831671 0.091459917
#>  [373] 0.088512146 0.096110835 0.093737901 0.092564208 0.099340498 0.099077299
#>  [379] 0.096134859 0.113639854 0.112706206 0.112937346 0.111162904 0.110486033
#>  [385] 0.112585046 0.112920980 0.131050844 0.129673126 0.126895071 0.129411058
#>  [391] 0.124232538 0.124228479 0.124263035 0.124295725 0.119685929 0.123790688
#>  [397] 0.123867801 0.122668157 0.122752226 0.117710318 0.123097933 0.123081728
#>  [403] 0.125931201 0.126522907 0.147132465 0.147231288 0.146783662 0.146746311
#>  [409] 0.146317417 0.143334100 0.143287796 0.143348054 0.143272393 0.143543273
#>  [415] 0.143427746 0.142727008 0.132188648 0.127926424 0.135714074 0.140474878
#>  [421] 0.152488295 0.154243583 0.154179338 0.167570772 0.160436040 0.162451860
#>  [427] 0.166056059 0.170418339 0.169594840 0.174726285 0.160648113 0.156073339
#>  [433] 0.157889946 0.155058303 0.156330723 0.156603556 0.158495888 0.158545261
#>  [439] 0.158601918 0.162417093 0.162495032 0.168386124 0.172212309 0.172260729
#>  [445] 0.172234018 0.178689008 0.179030183 0.178068879 0.180017504 0.179968281
#>  [451] 0.147883877 0.159889308 0.159170960 0.159639768 0.163059643 0.129540741
#>  [457] 0.129626624 0.129494534 0.130462467 0.133657746 0.135486739 0.135557437
#>  [463] 0.144184665 0.144241264 0.144255914 0.143965064 0.143975606 0.142582877
#>  [469] 0.144407071 0.139248385 0.139338348 0.139603372 0.134698359 0.133892753
#>  [475] 0.126871889 0.125914421 0.126364652 0.126779283 0.120786861 0.125100071
#>  [481] 0.125167570 0.128967319 0.129270581 0.127213303 0.132609986 0.133783693
#>  [487] 0.133324693 0.133291308 0.119828394 0.120155915 0.119091396 0.122777093
#>  [493] 0.122768761 0.117194628 0.117136102 0.110569288 0.106983397 0.104973359
#>  [499] 0.201598510 0.199578580 0.199572415 0.240790955 0.231532614 0.233091169
#>  [505] 0.229165102 0.229136503 0.232146252 0.229289918 0.229682234 0.231932831
#>  [511] 0.229204812 0.230340947 0.230228312 0.226181206 0.213201343 0.204725852
#>  [517] 0.206696973 0.209983252 0.221915661 0.225221053 0.223452835 0.222393638
#>  [523] 0.220617391 0.226802274 0.216625893 0.215675075 0.215625259 0.215625146
#>  [529] 0.216371532 0.222263000 0.220702135 0.218412558 0.216081313 0.221986600
#>  [535] 0.221994938 0.221830935 0.221800923 0.221896085 0.221880932 0.221909041
#>  [541] 0.217744241 0.217532658 0.216060115 0.216382680 0.220383002 0.218723195
#>  [547] 0.232505174 0.235950787 0.232538374 0.104023300 0.105694824 0.096207420
#>  [553] 0.098907896 0.099054341 0.099926847 0.101889960 0.105415347 0.102494409
#>  [559] 0.104047849 0.104842545 0.108061405 0.108473464 0.101410450 0.101482573
#>  [565] 0.101322109 0.103838723 0.104653878 0.100031073 0.092287336 0.092821864
#>  [571] 0.091520528 0.087593003 0.085340124 0.081324953 0.081204643 0.084627704
#>  [577] 0.070990412 0.070978279 0.070896481 0.066056203 0.064287291 0.062260868
#>  [583] 0.055129705 0.054923615 0.054963515 0.055014654 0.055071750 0.055884337
#>  [589] 0.055910253 0.056227793 0.056260861 0.057394588 0.057593549 0.057555016
#>  [595] 0.056076455 0.055249869 0.048815283 0.048907109 0.050845506 0.054137109
#>  [601] 0.054154943 0.054091038 0.062691282 0.054294261 0.055451607 0.052721873
#>  [607] 0.052692912 0.042987073 0.044767541 0.044695391 0.042071446 0.028445084
#>  [613] 0.027373322 0.018061065 0.018003528 0.018161106 0.018455608 0.018429781
#>  [619] 0.018418726 0.017698992 0.017352198 0.018054622 0.019864518 0.019800664
#>  [625] 0.018768778 0.018718210 0.019417878 0.019770215 0.019953120 0.019972966
#>  [631] 0.019868397 0.030363317 0.032064627 0.034602685 0.037867994 0.037878512
#>  [637] 0.036675515 0.033961001 0.029856553 0.030779632 0.029401528 0.049728581
#>  [643] 0.049668594 0.049407161 0.043512125 0.043639351 0.044204616 0.044253738
#>  [649] 0.044177846 0.043842324 0.043799620 0.043804771 0.043196306 0.042691358
#>  [655] 0.047922444 0.032694627 0.032543297 0.033836692 0.033865901 0.033831808
#>  [661] 0.034277866 0.034698014 0.035732031 0.035735659 0.035747338 0.035747048
#>  [667] 0.035710863 0.035677213 0.035042020 0.035038394 0.035530999 0.037628800
#>  [673] 0.037490587 0.037373466 0.037381540 0.037364638 0.036636617 0.040099209
#>  [679] 0.040061691 0.040298621 0.040393157 0.040356616 0.037398338 0.037325534
#>  [685] 0.049482167 0.048109152 0.048114849 0.048683612 0.050011626 0.051482649
#>  [691] 0.050910014 0.051341669 0.039979712 0.038413154 0.038376958 0.041046154
#>  [697] 0.039295131 0.039292625 0.039269873 0.039233373 0.039154915 0.038572238
#>  [703] 0.038207072 0.039089960 0.041061646 0.041092289 0.061387342 0.061457154
#>  [709] 0.058855343 0.058811841 0.085684577 0.085237179 0.086835974 0.079287643
#>  [715] 0.079123621 0.079135917 0.079164465 0.078928245 0.074383666 0.070342341
#>  [721] 0.070258223 0.070277514 0.068912886 0.068908185 0.053978457 0.053826269
#>  [727] 0.053127409 0.053162712 0.051204990 0.051200932 0.051048259 0.051591293
#>  [733] 0.051573144 0.052073476 0.052033249 0.042700498 0.043086723 0.043086209
#>  [739] 0.042591813 0.042131159 0.042164383 0.042613129 0.042899373 0.042756417
#>  [745] 0.043999419 0.043914562 0.044288281 0.045560538 0.045559452 0.045636314
#>  [751] 0.051565466 0.051591928 0.052448728 0.052813733 0.052512093 0.051254275
#>  [757] 0.051253360 0.051609200 0.051274997 0.051439957 0.050003821 0.034669908
#>  [763] 0.032461778 0.030901964 0.019573406 0.018154482 0.018262152 0.018253446
#>  [769] 0.018251924 0.018323188 0.020179680 0.020202444 0.019994004 0.020105073
#>  [775] 0.020097977 0.033038802 0.033282049 0.034675164 0.036058939 0.036135514
#>  [781] 0.036149730 0.036243197 0.036305002 0.033613360 0.033594592 0.034280015
#>  [787] 0.033608092 0.032449394 0.030400132 0.031490136 0.028064101 0.026490840
#>  [793] 0.027082022 0.025965338 0.026726876 0.026663256 0.026784364 0.027073882
#>  [799] 0.026631163 0.026642666 0.024883576 0.014429429 0.014452711 0.013890560
#>  [805] 0.013846404 0.013853858 0.013814732 0.011611745 0.011505189 0.011579118
#>  [811] 0.011753767 0.012088554 0.015517036 0.016553301 0.016222780 0.016303084
#>  [817] 0.016535968 0.016394264 0.016397299 0.016398820 0.016426279 0.015444067
#>  [823] 0.015311693 0.015411460 0.015726487 0.015639252 0.015399444 0.015328806
#>  [829] 0.015597116 0.015703421 0.015705576 0.015399607 0.015387396 0.015373577
#>  [835] 0.015741785 0.015757574 0.014812470 0.017630306 0.017475493 0.017805441
#>  [841] 0.017505776 0.018490741 0.018866127 0.018557646 0.018614973 0.018212188
#>  [847] 0.018209045 0.018202097 0.018120583 0.017912830 0.018023401 0.017049786
#>  [853] 0.017117999 0.017107988 0.017113800 0.017123622 0.017120522 0.016928475
#>  [859] 0.020913403 0.020479982 0.020276467 0.020148680 0.018921248 0.015798140
#>  [865] 0.015435622 0.015664374 0.015508082 0.015391598 0.015122041 0.015122645
#>  [871] 0.014884756 0.014879972 0.015903112 0.016090862 0.015907984 0.014586600
#>  [877] 0.014521333 0.022382500 0.023725404 0.023949545 0.023876874 0.023504317
#>  [883] 0.024619630 0.023919627 0.047906641 0.051463679 0.051485134 0.056143142
#>  [889] 0.064362942 0.065917214 0.065685035 0.066630845 0.067104284 0.066279706
#>  [895] 0.064275018 0.064570614 0.065551669 0.062691012 0.060978375 0.060635762
#>  [901] 0.061036618 0.062174154 0.062340211 0.062566329 0.062611255 0.062593498
#>  [907] 0.062603323 0.060776961 0.062183598 0.060687316 0.062641923 0.063793074
#>  [913] 0.064492207 0.069373435 0.069379699 0.071882588 0.069645816 0.070453575
#>  [919] 0.069890990 0.066523661 0.066965543 0.068890769 0.064338414 0.064347752
#>  [925] 0.063458652 0.059030167 0.062387500 0.063190442 0.059197285 0.059372795
#>  [931] 0.059386302 0.059757237 0.060636736 0.060775515 0.062486956 0.038362031
#>  [937] 0.033960228 0.033884396 0.032080520 0.031480252 0.032143255 0.032478947
#>  [943] 0.030199909 0.030203552 0.030518579 0.031119888 0.032537364 0.030205823
#>  [949] 0.031092838 0.032341517 0.034199760 0.034023897 0.034039627 0.034567701
#>  [955] 0.034576119 0.034559476 0.034674864 0.034827733 0.036602750 0.035692616
#>  [961] 0.036745039 0.036706064 0.036072364 0.036072233 0.036063649 0.035376450
#>  [967] 0.033364034 0.035072583 0.035444349 0.036293755 0.040674628 0.041218898
#>  [973] 0.041412083 0.048019234 0.048519402 0.050162720 0.069007087 0.068585325
#>  [979] 0.069206089 0.070415881 0.067847386 0.068921507 0.069271614 0.070355897
#>  [985] 0.071870295 0.074127477 0.076051167 0.076166043 0.079327952 0.080082382
#>  [991] 0.078290244 0.078382844 0.089650470 0.230208381 0.232132623 0.234332217
#>  [997] 0.239993014 0.229853727 0.201150009 0.195385157 0.196268487 0.184922979
#> [1003] 0.186370891 0.186533214 0.177268098 0.178911652 0.180004332 0.179146777
#> [1009] 0.175843484 0.171995518 0.172460187 0.167626462 0.166187087 0.162107008
#> [1015] 0.153581630 0.154108845 0.237528627 0.243461661 0.239409514 0.228778480
#> [1021] 0.234460950 0.230013081 0.222580420 0.201298754 0.193903775 0.180174729
#> [1027] 0.178434797 0.178257198 0.175774006 0.171421297 0.166850597 0.166539555
#> [1033] 0.163185400 0.127243141 0.123564593 0.118969713 0.112772606 0.109284520
#> [1039] 0.109357865 0.100344157 0.098348890 0.094420042 0.087604712 0.083383930
#> [1045] 0.093565231 0.087371657 0.084361669 0.078942127 0.078454190 0.081069071
#> [1051] 0.082102922 0.081095305 0.073443478 0.072341737 0.072321916 0.070898871
#> [1057] 0.068045875 0.068321627 0.068527686 0.069263227 0.069732092 0.069436516
#> [1063] 0.069476497 0.069569224 0.070915215 0.075030163 0.075336234 0.075086907
#> [1069] 0.071447011 0.071985404 0.070464322 0.065864174 0.066768248 0.066517410
#> [1075] 0.072909100 0.071383099 0.074143396 0.066452590 0.066698146 0.067830217
#> [1081] 0.068792508 0.069247459
roll_rayleigh(dat.PoR$azi.PoR, prd = 135, unc = dat$unc, width = 51)
#>    [1]  0.86144976  0.86141046  0.85663842  0.86807344  0.87381094  0.84896609
#>    [7]  0.84785386  0.84764553  0.84881558  0.84965139  0.85198332  0.84910239
#>   [13]  0.79528121  0.79702846  0.79867662  0.80584853  0.80737869  0.80836606
#>   [19]  0.81295778  0.82442906  0.82571301  0.82636005  0.82759768  0.83658311
#>   [25]  0.83662925  0.84491788  0.84762703  0.84734078  0.85074834  0.85352359
#>   [31]  0.85478594  0.85622309  0.85396306  0.85349512  0.85533488  0.85539413
#>   [37]  0.85706271  0.86174778  0.86485858  0.86449561  0.85046737  0.85615548
#>   [43]  0.86045541  0.86202111  0.86853305  0.86893183  0.86885525  0.87175856
#>   [49]  0.86894433  0.86884233  0.87159441  0.87147971  0.87196161  0.87926956
#>   [55]  0.87597700  0.87392760  0.88957569  0.89164508  0.89226583  0.89213221
#>   [61]  0.89357806  0.89055448  0.89288397  0.93325538  0.93269398  0.92465327
#>   [67]  0.92366042  0.92243798  0.91717956  0.91402418  0.91190351  0.91414139
#>   [73]  0.87482357  0.87427330  0.86951581  0.87594763  0.87081745  0.86901300
#>   [79]  0.86793691  0.86533247  0.85679039  0.85656519  0.85669855  0.85525588
#>   [85]  0.85372137  0.82775706  0.81804739  0.82955162  0.82986542  0.82547878
#>   [91]  0.82449936  0.83396760  0.82642343  0.81843551  0.81803329  0.81232873
#>   [97]  0.80785654  0.80823913  0.80840611  0.80856752  0.81735832  0.81312497
#>  [103]  0.81211671  0.79909047  0.78344204  0.78430790  0.76822400  0.76835376
#>  [109]  0.76931423  0.76933956  0.76958297  0.77145309  0.77368995  0.77137000
#>  [115]  0.77336102  0.75052568  0.74125087  0.74123802  0.74073153  0.74092201
#>  [121]  0.74287978  0.73563971  0.73230820  0.77124663  0.77149570  0.77139625
#>  [127]  0.76102348  0.76102957  0.76230466  0.76094454  0.79303884  0.79780870
#>  [133]  0.79808964  0.79821605  0.80070935  0.80773329  0.83000835  0.84465075
#>  [139]  0.83448818  0.83317602  0.82226520  0.82205754  0.82424297  0.82173892
#>  [145]  0.82877774  0.84886725  0.85424422  0.86462080  0.86468614  0.86921093
#>  [151]  0.86919858  0.86513483  0.87216763  0.87579041  0.88871993  0.89517127
#>  [157]  0.89966834  0.91696578  0.91712032  0.91691668  0.91719816  0.91719000
#>  [163]  0.91541836  0.91512199  0.91295347  0.90896000  0.93513455  0.93474861
#>  [169]  0.93479622  0.93559593  0.93582537  0.94675042  0.95028988  0.95027911
#>  [175]  0.94939351  0.94707306  0.94717604  0.94690591  0.94681610  0.94640802
#>  [181]  0.94831168  0.94279940  0.93949338  0.93783899  0.93760687  0.93837082
#>  [187]  0.94261202  0.93784303  0.93603747  0.93343563  0.93455112  0.95072272
#>  [193]  0.94897805  0.94602223  0.94779862  0.93543009  0.91464432  0.91257157
#>  [199]  0.90286136  0.90082524  0.89627727  0.89614594  0.86916812  0.85978303
#>  [205]  0.85508589  0.85050541  0.84947610  0.77433839  0.77417128  0.74166336
#>  [211]  0.73860226  0.74420397  0.74422463  0.73895307  0.73891687  0.74021456
#>  [217]  0.74456338  0.75365320  0.75338805  0.75039716  0.75390992  0.75387443
#>  [223]  0.75471976  0.75226819  0.75332401  0.75506149  0.76276945  0.76417332
#>  [229]  0.76427246  0.76304973  0.76028191  0.75954869  0.76919149  0.77952351
#>  [235]  0.78240128  0.78230895  0.78706458  0.76762695  0.77182857  0.77871186
#>  [241]  0.78057326  0.77974881  0.78021026  0.77880322  0.77321199  0.77380436
#>  [247]  0.78956479  0.79215104  0.79391304  0.79126155  0.79043335  0.79001516
#>  [253]  0.78933223  0.80535331  0.81140575  0.81053405  0.80604425  0.80166703
#>  [259]  0.85168711  0.85543999  0.89271350  0.89828800  0.89889646  0.89762697
#>  [265]  0.89452056  0.87688512  0.87976380  0.87957956  0.87549643  0.87589398
#>  [271]  0.88749940  0.88453943  0.84038183  0.82135684  0.82037562  0.81871266
#>  [277]  0.81714021  0.81260205  0.81180568  0.81302386  0.81435557  0.80497782
#>  [283]  0.80466301  0.79564335  0.78183389  0.78181519  0.77658125  0.77025372
#>  [289]  0.76990599  0.77673746  0.76700104  0.76718331  0.77576896  0.77671163
#>  [295]  0.77706695  0.77828619  0.77777208  0.77685834  0.78148661  0.78072028
#>  [301]  0.78087002  0.77991402  0.78055391  0.78061803  0.78329683  0.77220398
#>  [307]  0.77197829  0.77653866  0.77544358  0.77615232  0.74110999  0.73153987
#>  [313]  0.73402566  0.73827884  0.72163548  0.72899504  0.74951046  0.74065762
#>  [319]  0.74023033  0.73682518  0.74643011  0.73825983  0.73905812  0.76183096
#>  [325]  0.77504838  0.77808541  0.77776503  0.77879785  0.77802096  0.77803968
#>  [331]  0.76579553  0.76997406  0.77891100  0.77940159  0.78155181  0.78913521
#>  [337]  0.77461067  0.77814356  0.78082306  0.77650234  0.76950805  0.77149918
#>  [343]  0.77111290  0.76361186  0.76756397  0.76787405  0.76016799  0.75995287
#>  [349]  0.74351416  0.74222670  0.74048000  0.74519373  0.75174484  0.69811295
#>  [355]  0.69815729  0.69818083  0.70038478  0.69418447  0.68769778  0.68974598
#>  [361]  0.68697298  0.71082990  0.71211523  0.69504992  0.66802446  0.68422373
#>  [367]  0.67263588  0.65636910  0.67499113  0.69350239  0.69825235  0.67257443
#>  [373]  0.68551396  0.64813207  0.66035001  0.66197941  0.64905033  0.64986174
#>  [379]  0.65661827  0.60795973  0.61246640  0.61991797  0.62182750  0.61855411
#>  [385]  0.60538976  0.59691827  0.54088254  0.55088619  0.56174378  0.55111159
#>  [391]  0.57088235  0.57101261  0.56976954  0.56863680  0.58253604  0.57214261
#>  [397]  0.56959528  0.57893968  0.57856806  0.59908483  0.59514475  0.59565670
#>  [403]  0.58312969  0.58060050  0.56134739  0.55959520  0.55680444  0.55806352
#>  [409]  0.56588709  0.57007858  0.57035005  0.56896464  0.57142395  0.56944057
#>  [415]  0.57320554  0.58615408  0.60071517  0.60993316  0.58089701  0.56863640
#>  [421]  0.54940714  0.53478446  0.53610883  0.49308009  0.51158977  0.50002567
#>  [427]  0.49107375  0.46996003  0.47434152  0.46313724  0.49758208  0.50790039
#>  [433]  0.49414165  0.49096435  0.48560232  0.48519332  0.48369266  0.48702735
#>  [439]  0.48609733  0.46508321  0.46287519  0.44497371  0.42355151  0.42193350
#>  [445]  0.42335737  0.40204033  0.39510295  0.39576743  0.38121789  0.38187109
#>  [451]  0.40674175  0.36441451  0.36836789  0.36385455  0.34952946  0.41450515
#>  [457]  0.41166531  0.41521688  0.40958141  0.39009204  0.37966814  0.37727259
#>  [463]  0.35431283  0.35192762  0.35123198  0.35492509  0.35559344  0.36094094
#>  [469]  0.35392736  0.36723370  0.36521352  0.36038780  0.38386964  0.39091185
#>  [475]  0.41629222  0.41781726  0.41419298  0.40945502  0.43700682  0.41073346
#>  [481]  0.40819751  0.39352628  0.39183469  0.40424804  0.38776947  0.38086740
#>  [487]  0.38922541  0.39056248  0.42478507  0.42398365  0.42639870  0.41144472
#>  [493]  0.41179937  0.44143808  0.44366559  0.46500500  0.48481705  0.50141512
#>  [499]  0.38798668  0.40593771  0.40626641  0.37623335  0.42104992  0.42023399
#>  [505]  0.43983398  0.44105249  0.44634041  0.46138624  0.45946219  0.46723256
#>  [511]  0.48776460  0.47229325  0.47489964  0.50566943  0.51881000  0.52006756
#>  [517]  0.49851453  0.47382787  0.45790650  0.45080213  0.44523776  0.44050016
#>  [523]  0.43542126  0.41870676  0.43332072  0.42638879  0.42834627  0.42835028
#>  [529]  0.42253797  0.40188355  0.41471561  0.41328630  0.43288766  0.41907992
#>  [535]  0.41870972  0.42004378  0.41944437  0.41611365  0.41676446  0.41562258
#>  [541]  0.42985639  0.43590274  0.45218809  0.44399692  0.43984860  0.43339927
#>  [547]  0.41320622  0.40768826  0.41300136  0.54605938  0.53605837  0.55874273
#>  [553]  0.54914391  0.54831601  0.54188241  0.53262330  0.51881827  0.52642334
#>  [559]  0.51566552  0.50421357  0.49234002  0.48783493  0.51422459  0.51218769
#>  [565]  0.51708025  0.51532282  0.51457866  0.53260254  0.56150471  0.56077184
#>  [571]  0.56797618  0.58194645  0.59345696  0.60583269  0.60886550  0.60650633
#>  [577]  0.64522369  0.64562621  0.64802933  0.67002104  0.68760909  0.70201514
#>  [583]  0.72807055  0.73036945  0.73050856  0.72891672  0.72822327  0.72324984
#>  [589]  0.72271270  0.71529526  0.71437057  0.70843800  0.70322006  0.70440881
#>  [595]  0.71877584  0.72452727  0.74653107  0.74361394  0.72739238  0.71233666
#>  [601]  0.71181870  0.71377099  0.68228578  0.70429797  0.69805138  0.70558519
#>  [607]  0.70650852  0.73640288  0.72727256  0.72990720  0.75514179  0.79872194
#>  [613]  0.81292828  0.82796891  0.83027942  0.82629486  0.82111126  0.82171927
#>  [619]  0.82331978  0.83208470  0.83380220  0.82715578  0.82641216  0.82719402
#>  [625]  0.83449869  0.83669356  0.82799428  0.81759027  0.81344767  0.81298125
#>  [631]  0.81371323  0.79600798  0.78449303  0.77267418  0.75629579  0.75600375
#>  [637]  0.76569054  0.77738012  0.80298609  0.80855857  0.82385351  0.78257578
#>  [643]  0.78441456  0.79134402  0.80475060  0.79969065  0.79676643  0.79475372
#>  [649]  0.79695970  0.80561115  0.80695139  0.80714618  0.80822407  0.81748859
#>  [655]  0.80616289  0.82858450  0.83386976  0.81679481  0.81548418  0.81614303
#>  [661]  0.80646048  0.80279613  0.79646314  0.79629573  0.79576657  0.79578006
#>  [667]  0.79722904  0.79853008  0.80289386  0.80304355  0.79867114  0.78204194
#>  [673]  0.78415094  0.78411894  0.78375257  0.78452737  0.78886028  0.77439811
#>  [679]  0.77607021  0.77560960  0.77740405  0.77984380  0.79208125  0.79446560
#>  [685]  0.75705653  0.76911048  0.76884176  0.76552548  0.76023760  0.75366995
#>  [691]  0.75826726  0.75693013  0.79308636  0.80202418  0.80268956  0.79581326
#>  [697]  0.80857059  0.80868259  0.80997570  0.81168943  0.81281201  0.81800539
#>  [703]  0.82152261  0.81532306  0.80136577  0.80019750  0.77882388  0.77826770
#>  [709]  0.79792529  0.79841466  0.73816283  0.74448090  0.73162825  0.72700457
#>  [715]  0.72449493  0.72339972  0.72265279  0.72514883  0.73426577  0.74212578
#>  [721]  0.74410826  0.74370081  0.75518343  0.75527427  0.77745834  0.77936067
#>  [727]  0.78362566  0.78681674  0.79972922  0.79994214  0.80303997  0.80416848
#>  [733]  0.80458421  0.80096271  0.80127740  0.83600407  0.83357662  0.83360275
#>  [739]  0.83589097  0.83606983  0.83545555  0.83286192  0.83012089  0.83039893
#>  [745]  0.82307935  0.82383944  0.82231779  0.81499577  0.81505237  0.81341764
#>  [751]  0.79007097  0.78972450  0.77762974  0.77292199  0.77706023  0.78915724
#>  [757]  0.78977623  0.78263567  0.78549439  0.78463893  0.79241341  0.84647015
#>  [763]  0.85693180  0.86866802  0.89170677  0.90138443  0.90023538  0.90065230
#>  [769]  0.90070303  0.90041848  0.89357632  0.89307286  0.89503379  0.89410268
#>  [775]  0.89424356  0.87831490  0.87729633  0.87288304  0.87039409  0.86957603
#>  [781]  0.86930314  0.86706328  0.86618721  0.87244109  0.87265331  0.86390706
#>  [787]  0.86430430  0.86807728  0.87417194  0.86837021  0.87585669  0.88101810
#>  [793]  0.87793427  0.88195543  0.87811131  0.88009121  0.87913898  0.87732946
#>  [799]  0.87934107  0.87913611  0.88371090  0.91652306  0.91590401  0.92657787
#>  [805]  0.92863246  0.92825779  0.92916487  0.93397270  0.93844413  0.93759561
#>  [811]  0.93669324  0.93466566  0.92071317  0.91606511  0.91859109  0.91742139
#>  [817]  0.91521119  0.91730587  0.91721768  0.91717154  0.91601765  0.92053998
#>  [823]  0.92166781  0.91991812  0.91816230  0.91912436  0.92082670  0.92159751
#>  [829]  0.92053943  0.91998321  0.91979188  0.92179136  0.92272296  0.92317898
#>  [835]  0.91988029  0.91902897  0.92711162  0.92120518  0.92188534  0.91969019
#>  [841]  0.92375790  0.91774654  0.91499062  0.91830874  0.91847358  0.92229609
#>  [847]  0.92230814  0.92241074  0.92306910  0.92538631  0.92405325  0.92778308
#>  [853]  0.92656674  0.92686610  0.92654846  0.92637399  0.92664475  0.92775292
#>  [859]  0.91620180  0.91820885  0.91904490  0.92076014  0.92419305  0.93620605
#>  [865]  0.93793111  0.93792698  0.93907804  0.94032689  0.94274849  0.94231303
#>  [871]  0.94310665  0.94331113  0.93946419  0.93733780  0.93901688  0.94367099
#>  [877]  0.94362924  0.91735897  0.90926460  0.90958858  0.90990336  0.91191822
#>  [883]  0.90560628  0.90828668  0.84648868  0.82887077  0.82801445  0.80875661
#>  [889]  0.80126856  0.79855566  0.79891438  0.79286639  0.79343289  0.79681265
#>  [895]  0.80113488  0.79943539  0.79201115  0.79513749  0.79938881  0.80017324
#>  [901]  0.79614625  0.79418757  0.79206598  0.79154658  0.79063327  0.79086246
#>  [907]  0.79030430  0.79568586  0.78723288  0.79355919  0.78499643  0.77829269
#>  [913]  0.77429293  0.76423432  0.76358816  0.75279576  0.75916909  0.75516478
#>  [919]  0.75610588  0.76069296  0.75908367  0.75318602  0.76368966  0.76342516
#>  [925]  0.76632413  0.76604082  0.75440641  0.74997697  0.76506842  0.76069549
#>  [931]  0.76033171  0.75761216  0.75291985  0.75314244  0.74600472  0.80883457
#>  [937]  0.82679868  0.82933859  0.84364529  0.83953701  0.82840801  0.82679315
#>  [943]  0.83833162  0.83816162  0.83540203  0.83175280  0.81493313  0.81097961
#>  [949]  0.79789111  0.78361724  0.76173676  0.75791177  0.75718349  0.74790953
#>  [955]  0.74758804  0.74803163  0.74352880  0.73802377  0.71872663  0.72340423
#>  [961]  0.70611209  0.70533720  0.70547062  0.70505437  0.70545793  0.71505068
#>  [967]  0.72083662  0.70001690  0.68949848  0.67498166  0.63953380  0.62368264
#>  [973]  0.61648505  0.56313679  0.55167654  0.51699151  0.43082624  0.41923716
#>  [979]  0.40682159  0.39032760  0.41169383  0.39670436  0.38774591  0.37158912
#>  [985]  0.35280115  0.32883412  0.30304154  0.29242468  0.26053187  0.24400899
#>  [991]  0.25597902  0.26583309  0.20038506  0.04563141  0.03268658  0.01767255
#>  [997] -0.01011261  0.00921501  0.07048647  0.09368467  0.09074370  0.10261374
#> [1003]  0.09594780  0.09400870  0.11915615  0.10801985  0.10304032  0.11054771
#> [1009]  0.13084045  0.14759272  0.14441024  0.16785032  0.17755640  0.19466019
#> [1015]  0.21695819  0.21285981  0.13010325  0.10497767  0.12292785  0.15011774
#> [1021]  0.13301176  0.15357367  0.17828971  0.21978813  0.24045621  0.28128984
#> [1027]  0.28422307  0.28586613  0.29868961  0.32130895  0.34313387  0.34132177
#> [1033]  0.35423667  0.39860622  0.41982504  0.44177045  0.46285737  0.47861591
#> [1039]  0.48259344  0.51484063  0.53111466  0.54028683  0.56278177  0.58155318
#> [1045]  0.54740729  0.56980993  0.58554062  0.61094606  0.61655316  0.60849692
#> [1051]  0.60173548  0.60675879  0.62424505  0.63150075  0.63172955  0.64167040
#> [1057]  0.65656207  0.65298762  0.65023882  0.64413521  0.64004803  0.64149909
#> [1063]  0.64045382  0.63892541  0.63159531  0.61736599  0.61441522  0.61668182
#> [1069]  0.63645824  0.63166481  0.63686729  0.66050738  0.65251145  0.65353083
#> [1075]  0.63111794  0.64071884  0.62645636  0.66349735  0.65990141  0.64971188
#> [1081]  0.64071778  0.63381920