Rose Diagram

Rose Diagram

Usage

rose_baseplot(
  start = -90,
  clockwise = TRUE,
  main = NULL,
  labels = TRUE,
  at = seq(0, 360 - 45, 45),
  cborder = TRUE,
  ...
)

Arguments

start

The starting direction for measurement of angles, that is, the spatial direction which corresponds to a measured angle of zero. Either a character string giving a compass direction ("N" for north, "S" for south, "E" for east, or "W" for west) or a number giving the angle from the the horizontal (East) axis to the starting direction. For example, if clockwise=FALSE, then start=90 and start="N" are equivalent. The default is to measure angles anti-clockwise from the horizontal axis (East direction).

clockwise

Logical. Whether angles increase in the clockwise direction (clockwise=TRUE, the default) or anti-clockwise, counter-clockwise direction (FALSE).

main

Character string specifying the title of the plot.

labels

Either a logical value indicating whether to plot labels next to the tick marks, or a vector of labels for the tick marks.

at

Optional vector of angles at which tick marks should be plotted. Set at=numeric(0) to suppress tick marks.

cborder

logical. Border of rose plot.

...

optional arguments passed to plot.default()

Value

none