Quick analysis of a stress data set

Returns the converted azimuths, distances to the plate boundary, statistics of the model, and some plots.

Usage

stress_analysis(
  x,
  PoR,
  type = c("none", "in", "out", "right", "left"),
  pb,
  plot = TRUE,
  ...
)

Arguments

x

data.frame or sf object containing the coordinates of the point(s) (lat, lon), the direction of \(\sigma_{Hmax}\) azi and its standard deviation unc (optional)

PoR

Pole of Rotation. data.frame or object of class "euler.pole" containing the geographical coordinates of the Euler pole

type

Character. Type of plate boundary (optional). Can be "out", "in", "right", or "left" for outward, inward, right-lateral, or left-lateral moving plate boundaries, respectively. If "none" (the default), only the PoR-equivalent azimuth is returned.

pb

(optional) sf object of the plate boundary geometries in the geographical coordinate system

plot

(logical). Whether to produce a plot additional to output.

...

optional arguments to distance_from_pb()

Value

list containing the following values:

results

data.frame showing the the coordinate and azimuth conversions (lat.PoR, lon.PoR, and azi.PoR), the predicted azimuths (prd), deviation angle from predicted (dev), circular distance (cdist), misfit to predicted stress direction (nchisq) and, if given, distance to tested plate boundary (distance)

stats

array with circular (weighted) mean, circular standard deviation, circular variance, circular median, skewness, kurtosis, the 95% confidence angle, circular dispersion, and the normalized Chi-squared test statistic

test

list containing the test results of the (weighted) Rayleigh test against the uniform distribution about the predicted orientation.

See also

Examples

# \donttest{
data("nuvel1")
na_pa <- subset(nuvel1, nuvel1$plate.rot == "na")

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")

data("san_andreas")
stress_analysis(san_andreas, na_pa, type = "right", plate_boundary, plot = TRUE)
#> Warning: longer object length is not a multiple of shorter object length
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#> Reject Null Hypothesis
#> Error in dplyr::mutate(dplyr::arrange(data.frame(azi, distance, prd, unc,     regime = factor(regime, levels = c("U", "N", "NS", "S", "TS",         "T"))), distance), nchisq_i = (deviation_norm(azi, prd)/unc)^2/(90/unc)^2,     cdist = circular_distance(azi, prd), roll_mean = roll_circstats(azi,         w = 1/unc, FUN = circular_mean, width = width), roll_sd = roll_circstats(azi,         w = 1/unc, FUN = circular_sd, width = width)/2, roll_nchisq = roll_normchisq(azi,         prd, unc, width = width), roll_disp = roll_dispersion(azi,         prd, w = 1/unc, width = width)): ℹ In argument: `cdist = circular_distance(azi, prd)`.
#> Caused by error:
#> ! `cdist` must be size 422 or 1, not 407.
# }