Example crustal stress dataset

Subsets of the World Stress Map (WSM) compilation of information on the crustal present-day stress field (Version 1.1. 2019).

Usage

data('san_andreas')

data('tibet')

data('iceland')

Format

A sf object / data.frame with 10 columns. Each row represents a different in-situ stress measurement:

id

Measurement identifier

lat

latitude in degrees

lon

longitude in degrees

azi

SHmax azimuth in degrees

unc

MMeasruement standard eviation (in degrees)

type

Type of measurement

depth

Depth in km

quality

WSM quality rank

regime

Stress regime

An object of class sf (inherits from data.frame) with 2624 rows and 10 columns.

An object of class sf (inherits from data.frame) with 1235 rows and 10 columns.

An object of class sf (inherits from data.frame) with 643 rows and 10 columns.

Source

https://www.world-stress-map.org/

Details

'san_andreas"

contains 407 stress data adjacent to the San Andreas Fault to be tested against a tangentially displaced plate boundary.

"tibet"

contains 947 stress data from the Himalaya and Tibetan plateau to be tested against an inward-moving displaced plate boundary.

'iceland

contains 201 stress data from Iceland to be tested against a outward-moving displaced plate boundary.

References

Heidbach, Oliver; Rajabi, Mojtaba; Di Giacomo, Domenico; Harris, James; Lammers, Steffi; Morawietz, Sophia; Pierdominici, Simona; Reiter, Karsten; von Specht, Sebastian; Storchak, Dmitry; Ziegler, Moritz O. (2025): World Stress Map Database Release 2025. GFZ Data Services. doi:10.5880/WSM.2025.001

See also

Examples

data("san_andreas")
head(san_andreas)
#> Simple feature collection with 6 features and 9 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -117.92 ymin: 33.65 xmax: -117.41 ymax: 34.73
#> Geodetic CRS:  WGS 84
#>         id   lat     lon azi unc type depth quality regime
#> 1 wsm00954 33.65 -117.67   1  26  FMF  0.00       E      S
#> 2 wsm00955 33.85 -117.92   9  28  FMF  0.00       E      S
#> 3 wsm00956 34.05 -117.92   1  25  FMF  0.00       E      S
#> 4 wsm00957 34.05 -117.67 180  22  FMF  0.00       E      S
#> 5 wsm00958 34.30 -117.41  57  19   BO  2.62       B      S
#> 6 wsm00959 34.73 -117.78   4  40   HF  0.49       D      S
#>                geometry
#> 1 POINT (-117.67 33.65)
#> 2 POINT (-117.92 33.85)
#> 3 POINT (-117.92 34.05)
#> 4 POINT (-117.67 34.05)
#> 5  POINT (-117.41 34.3)
#> 6 POINT (-117.78 34.73)

data("tibet")
head(tibet)
#> Simple feature collection with 6 features and 9 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: 74.9 ymin: 37.2 xmax: 102.167 ymax: 44.961
#> Geodetic CRS:  WGS 84
#>         id    lat     lon azi unc type depth quality regime
#> 1 wsm02008 37.200  74.900  31   7   BO  3.15       C   <NA>
#> 2 wsm02039 40.500  78.500 171  90  FMA 15.00     Xru      T
#> 3 wsm02086 44.961  85.008  80  25   BO  2.83       C   <NA>
#> 4 wsm04371 38.667 102.167  32  90   OC  0.00       E   <NA>
#> 5 wsm04372 38.667 102.167 167  90   OC  0.00       E   <NA>
#> 6 wsm04373 38.667 102.167 155  90   OC  0.00       E   <NA>
#>                 geometry
#> 1      POINT (74.9 37.2)
#> 2      POINT (78.5 40.5)
#> 3  POINT (85.008 44.961)
#> 4 POINT (102.167 38.667)
#> 5 POINT (102.167 38.667)
#> 6 POINT (102.167 38.667)

data("iceland")
head(iceland)
#> Simple feature collection with 6 features and 9 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -16.99 ymin: 64.38 xmax: -15.16 ymax: 66.33
#> Geodetic CRS:  WGS 84
#>         id   lat    lon azi unc type depth quality regime             geometry
#> 1 wsm00187 64.38 -15.16 167  25  FMS  26.0       C      N POINT (-15.16 64.38)
#> 2 wsm00188 66.33 -16.29 172  40  FMS  33.0       D      S POINT (-16.29 66.33)
#> 3 wsm00190 66.20 -16.70 174  25  FMS   0.0       C      S   POINT (-16.7 66.2)
#> 4 wsm00191 64.90 -16.94 164  25  FMS  10.0       C      T  POINT (-16.94 64.9)
#> 5 wsm00192 64.57 -16.95 140  90  FMS  13.8       E      N POINT (-16.95 64.57)
#> 6 wsm00193 64.66 -16.99 129  90  FMS  14.5       E      N POINT (-16.99 64.66)