Theoretical Plate Tectonic Stress Paths

Construct \(\sigma_\text{Hmax}\) lines that are following small circles, great circles, or loxodromes of an Euler pole for the relative plate motion.

Usage

eulerpole_paths(
  x,
  type = c("sc", "gc", "ld"),
  n = 10,
  angle,
  cw,
  returnclass = c("sf", "sp")
)

eulerpole_smallcircles(x, n = 10, returnclass = c("sf", "sp"))

eulerpole_greatcircles(x, n = 10, returnclass = c("sf", "sp"))

eulerpole_loxodromes(x, n = 10, angle = 45, cw, returnclass = c("sf", "sp"))

Arguments

x

data.frame containing coordinates of Euler pole in lat, lon, and rotation angle (optional)

type

Character string specifying the type of curves to export. Either "sm" for small circles (default), "gc" for great circles, or "ld" for loxodromes.

n

Number of equally spaced curves

angle

Direction of loxodromes; default = 45

cw

logical. Sense of loxodromes: TRUE for clockwise loxodromes (right-lateral displaced plate boundaries). FALSE for counterclockwise loxodromes (left-lateral displaced plate boundaries).

returnclass

"sf" (default) for simple features or "sp" for spatial objects

Value

sf or SpatialLinesDataFrame

Details

Maximum horizontal stress can be aligned to three types of curves related to relative plate motion:

Small circles

Lines that have a constant distance to the Euler pole. If x contains angle, output additionally gives absolute velocity on small circle (degree/Myr -> km/Myr).

Great circles

Paths of the the shortest distance between the Euler pole and its antipodal position.

Loxodromes

Lines of constant bearing, i.e. curves cutting small circles at a constant angle.

Author

Tobias Stephan

Examples

data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate

eulerpole_smallcircles(euler)
#> Simple feature collection with 11 features and 2 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -90 xmax: 180 ymax: 90
#> CRS:           +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> First 10 features:
#>     n.small_circle n.abs_vel                       geometry
#> 0                0   0.00000 LINESTRING (-180 -90, -162 ...
#> 18              18  26.90136 LINESTRING (-180 -72, -162 ...
#> 36              36  51.16943 LINESTRING (-180 -54, -162 ...
#> 54              54  70.42867 LINESTRING (-180 -36, -162 ...
#> 72              72  82.79387 LINESTRING (-180 -18, -162 ...
#> 90              90  87.05463 LINESTRING (-180 0, -162 0,...
#> 108            108  82.79387 LINESTRING (-180 18, -162 1...
#> 126            126  70.42867 LINESTRING (-180 36, -162 3...
#> 144            144  51.16943 LINESTRING (-180 54, -162 5...
#> 162            162  26.90136 LINESTRING (-180 72, -162 7...
eulerpole_greatcircles(euler)
#> Simple feature collection with 11 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -90 xmax: 180 ymax: 90
#> CRS:           +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> First 10 features:
#>         n                       geometry
#> -180 -180 LINESTRING (-180 -90, -180 ...
#> -144 -144 LINESTRING (144 -90, 144 -8...
#> -108 -108 LINESTRING (108 -90, 108 -8...
#> -72   -72 LINESTRING (72 -90, 72 -89,...
#> -36   -36 LINESTRING (36 -90, 36 -89,...
#> 0       0 LINESTRING (0 -90, 0 -89, 0...
#> 36     36 LINESTRING (-36 -90, -36 -8...
#> 72     72 LINESTRING (-72 -90, -72 -8...
#> 108   108 LINESTRING (-108 -90, -108 ...
#> 144   144 LINESTRING (-144 -90, -144 ...
eulerpole_loxodromes(x = euler, angle = 45, n = 10, cw = FALSE)
#> Simple feature collection with 15 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -89.80256 xmax: 180 ymax: 89.80256
#> CRS:           +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> First 10 features:
#>         n                       geometry
#> -252 -252 LINESTRING (162.1974 -89.80...
#> -216 -216 LINESTRING (126.1974 -89.80...
#> -180 -180 LINESTRING (90.19744 -89.80...
#> -144 -144 LINESTRING (54.19744 -89.80...
#> -108 -108 LINESTRING (18.19744 -89.80...
#> -72   -72 LINESTRING (-17.80256 -89.8...
#> -36   -36 LINESTRING (-53.80256 -89.8...
#> 0       0 LINESTRING (-89.80256 -89.8...
#> 36     36 LINESTRING (-125.8026 -89.8...
#> 72     72 LINESTRING (-161.8026 -89.8...
eulerpole_loxodromes(x = euler, angle = 30, cw = TRUE)
#> Simple feature collection with 13 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -89.20062 xmax: 180 ymax: 89.20062
#> CRS:           +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> First 10 features:
#>         n                       geometry
#> -216 -216 LINESTRING (180 62.35383, 1...
#> -180 -180 LINESTRING (180 0, 179.5 0....
#> -144 -144 LINESTRING (180 -62.35383, ...
#> -108 -108 LINESTRING (159.5 -89.20062...
#> -72   -72 LINESTRING (123.5 -89.20062...
#> -36   -36 LINESTRING (87.5 -89.20062,...
#> 0       0 LINESTRING (51.5 -89.20062,...
#> 36     36 LINESTRING (15.5 -89.20062,...
#> 72     72 LINESTRING (-20.5 -89.20062...
#> 108   108 LINESTRING (-56.5 -89.20062...