SHmax direction resulting from multiple plate boundaries

Calculates a \(\sigma_{Hmax}\) direction at given coordinates, sourced by multiple plate boundaries. This first-order approximation is the circular mean of the superimposed theoretical directions, weighted by the rotation rates of the underlying PoRs.

Usage

superimposed_shmax(df, PoRs, types, absolute = TRUE, PoR_weighting = NULL)

Arguments

df

data.frame containing the coordinates of the point(s) (lat, lon), and the direction of \(\sigma_{Hmax}\) azi (in degrees)

PoRs

multirow data.frame or "euler.pole" object that must contain lat, lon and angle

types

character vector with length equal to number of rows in PoRs. Type of plate boundary. Must be "out", "in", "right", or "left" for outward, inward, right-lateral, or left-lateral moving plate boundaries, respectively.

absolute

logical. Whether the resultant azimuth should be weighted using the absolute rotation at the points or the angular rotation of the PoRs.

PoR_weighting

(optional) numeric vector with length equal to number of rows in PoRs. Extra weightings for the used PoRs.

Value

two column vector. azi is the resultant azimuth in degrees / geographical CRS), R is the resultant length.

See also

Examples

data(san_andreas)
data(nuvel1)
pors <- subset(nuvel1, plate.rot %in% c("eu", "na"))
superimposed_shmax(san_andreas, pors, types = c("in", "right"), PoR_weighting = c(2, 1))
#>              azi         R
#>    [1,] 153.9732 0.3574903
#>    [2,] 147.7526 0.3618659
#>    [3,] 147.9943 0.3618863
#>    [4,] 148.5271 0.3612084
#>    [5,] 149.3690 0.3554973
#>    [6,] 149.0230 0.3597045
#>    [7,] 149.7781 0.3542771
#>    [8,] 149.8176 0.3548953
#>    [9,] 149.9187 0.3547016
#>   [10,] 149.7272 0.3575109
#>   [11,] 151.1476 0.3474427
#>   [12,] 151.1916 0.3477400
#>   [13,] 150.3815 0.3583289
#>   [14,] 152.1214 0.3467730
#>   [15,] 152.9389 0.3412762
#>   [16,] 150.8507 0.3585717
#>   [17,] 152.5910 0.3528076
#>   [18,] 152.6290 0.3531297
#>   [19,] 151.2603 0.3626948
#>   [20,] 152.6868 0.3580640
#>   [21,] 152.8787 0.3587323
#>   [22,] 148.6211 0.3505213
#>   [23,] 147.9607 0.3596111
#>   [24,] 147.6519 0.3628086
#>   [25,] 147.9563 0.3603081
#>   [26,] 147.9052 0.3607820
#>   [27,] 147.9494 0.3613394
#>   [28,] 148.2939 0.3604398
#>   [29,] 148.2390 0.3612594
#>   [30,] 148.3866 0.3608855
#>   [31,] 149.3239 0.3552460
#>   [32,] 148.4090 0.3634924
#>   [33,] 149.5352 0.3540822
#>   [34,] 148.7725 0.3606254
#>   [35,] 148.9783 0.3594641
#>   [36,] 148.7643 0.3613161
#>   [37,] 150.0229 0.3512969
#>   [38,] 149.7383 0.3536575
#>   [39,] 149.6829 0.3541265
#>   [40,] 149.0140 0.3603998
#>   [41,] 149.4440 0.3570666
#>   [42,] 148.6832 0.3641636
#>   [43,] 149.6499 0.3562852
#>   [44,] 150.0320 0.3537690
#>   [45,] 149.6394 0.3570011
#>   [46,] 149.3639 0.3592940
#>   [47,] 149.7947 0.3563327
#>   [48,] 149.7387 0.3568014
#>   [49,] 149.6833 0.3572558
#>   [50,] 149.4633 0.3590807
#>   [51,] 149.0799 0.3625976
#>   [52,] 150.1206 0.3543046
#>   [53,] 150.0640 0.3547612
#>   [54,] 149.7825 0.3570580
#>   [55,] 149.5071 0.3593298
#>   [56,] 150.2222 0.3541058
#>   [57,] 150.1643 0.3545708
#>   [58,] 149.8829 0.3568537
#>   [59,] 149.9264 0.3571127
#>   [60,] 150.0255 0.3569225
#>   [61,] 149.9141 0.3578211
#>   [62,] 150.5769 0.3534587
#>   [63,] 150.0566 0.3578965
#>   [64,] 150.1552 0.3577120
#>   [65,] 150.0994 0.3581571
#>   [66,] 150.1420 0.3584181
#>   [67,] 150.5034 0.3570815
#>   [68,] 152.2669 0.3450945
#>   [69,] 150.4100 0.3593015
#>   [70,] 152.4882 0.3444485
#>   [71,] 152.7344 0.3430304
#>   [72,] 151.9772 0.3484327
#>   [73,] 152.9813 0.3416177
#>   [74,] 153.0234 0.3419594
#>   [75,] 152.8287 0.3433111
#>   [76,] 152.6361 0.3446597
#>   [77,] 153.1093 0.3441704
#>   [78,] 150.7769 0.3617311
#>   [79,] 150.8160 0.3620073
#>   [80,] 153.0401 0.3479904
#>   [81,] 152.7855 0.3497133
#>   [82,] 153.0149 0.3487506
#>   [83,] 152.9510 0.3491882
#>   [84,] 152.8252 0.3500343
#>   [85,] 152.7392 0.3512113
#>   [86,] 151.3758 0.3618687
#>   [87,] 152.6054 0.3538667
#>   [88,] 151.5297 0.3613294
#>   [89,] 151.4721 0.3617377
#>   [90,] 151.3569 0.3625664
#>   [91,] 151.2998 0.3629735
#>   [92,] 151.6052 0.3619113
#>   [93,] 152.1840 0.3581686
#>   [94,] 150.7189 0.3641447
#>   [95,] 151.0990 0.3635883
#>   [96,] 151.7437 0.3768023
#>   [97,] 152.7171 0.3577641
#>   [98,] 153.8721 0.3578254
#>   [99,] 153.6868 0.3588994
#>  [100,] 153.5062 0.3595631
#>  [101,] 147.7642 0.3598052
#>  [102,] 147.5568 0.3624065
#>  [103,] 147.8573 0.3605722
#>  [104,] 148.0087 0.3598238
#>  [105,] 148.2613 0.3584836
#>  [106,] 148.0405 0.3621067
#>  [107,] 147.8291 0.3649775
#>  [108,] 148.4395 0.3604098
#>  [109,] 149.2994 0.3531928
#>  [110,] 149.1892 0.3541422
#>  [111,] 149.2311 0.3547500
#>  [112,] 149.1757 0.3552327
#>  [113,] 148.5837 0.3603965
#>  [114,] 148.3185 0.3630553
#>  [115,] 148.6290 0.3606289
#>  [116,] 148.6745 0.3608486
#>  [117,] 149.2637 0.3560758
#>  [118,] 148.8301 0.3598136
#>  [119,] 148.6184 0.3616607
#>  [120,] 148.7195 0.3610821
#>  [121,] 149.9196 0.3515080
#>  [122,] 149.8635 0.3519813
#>  [123,] 149.6373 0.3538637
#>  [124,] 149.5810 0.3543430
#>  [125,] 148.8709 0.3603915
#>  [126,] 148.7084 0.3621112
#>  [127,] 149.8516 0.3527125
#>  [128,] 150.3582 0.3492010
#>  [129,] 150.2415 0.3501533
#>  [130,] 150.0684 0.3515698
#>  [131,] 149.5599 0.3557845
#>  [132,] 150.4035 0.3494824
#>  [133,] 149.6056 0.3560299
#>  [134,] 148.8948 0.3623482
#>  [135,] 149.7612 0.3553684
#>  [136,] 150.0954 0.3529377
#>  [137,] 149.4237 0.3584834
#>  [138,] 150.1340 0.3535662
#>  [139,] 150.0760 0.3540343
#>  [140,] 149.4078 0.3595412
#>  [141,] 149.2995 0.3604418
#>  [142,] 149.6223 0.3580694
#>  [143,] 150.1786 0.3538381
#>  [144,] 150.0076 0.3552172
#>  [145,] 149.9503 0.3556819
#>  [146,] 149.3433 0.3606865
#>  [147,] 149.2901 0.3611385
#>  [148,] 150.9825 0.3480829
#>  [149,] 150.2792 0.3536496
#>  [150,] 150.1077 0.3550258
#>  [151,] 149.9947 0.3559479
#>  [152,] 149.8273 0.3573058
#>  [153,] 149.7096 0.3585760
#>  [154,] 149.2741 0.3621742
#>  [155,] 151.0268 0.3483766
#>  [156,] 150.3227 0.3539193
#>  [157,] 150.2657 0.3543740
#>  [158,] 150.0393 0.3562006
#>  [159,] 149.8708 0.3575633
#>  [160,] 150.0827 0.3564630
#>  [161,] 149.9697 0.3573721
#>  [162,] 149.7962 0.3590843
#>  [163,] 150.0136 0.3576363
#>  [164,] 149.9014 0.3585399
#>  [165,] 150.4468 0.3550932
#>  [166,] 150.3320 0.3559999
#>  [167,] 149.9937 0.3586969
#>  [168,] 150.1978 0.3579745
#>  [169,] 150.2410 0.3582418
#>  [170,] 150.0668 0.3599181
#>  [171,] 150.3967 0.3576130
#>  [172,] 152.3936 0.3441797
#>  [173,] 152.3517 0.3457386
#>  [174,] 150.7855 0.3572958
#>  [175,] 150.6708 0.3581779
#>  [176,] 150.5574 0.3590445
#>  [177,] 152.9583 0.3424035
#>  [178,] 152.8927 0.3428621
#>  [179,] 152.6994 0.3442118
#>  [180,] 152.5724 0.3451019
#>  [181,] 152.5082 0.3455582
#>  [182,] 150.6473 0.3592360
#>  [183,] 150.5900 0.3596777
#>  [184,] 152.9347 0.3432004
#>  [185,] 152.8701 0.3436422
#>  [186,] 153.3146 0.3409029
#>  [187,] 153.1718 0.3421967
#>  [188,] 153.1051 0.3426468
#>  [189,] 153.1467 0.3429889
#>  [190,] 153.0162 0.3438812
#>  [191,] 153.7970 0.3389332
#>  [192,] 152.8164 0.3470937
#>  [193,] 152.7940 0.3478546
#>  [194,] 152.7580 0.3489973
#>  [195,] 153.6357 0.3437225
#>  [196,] 153.0784 0.3483268
#>  [197,] 152.6642 0.3520153
#>  [198,] 151.4623 0.3604066
#>  [199,] 151.7936 0.3586142
#>  [200,] 151.3185 0.3622778
#>  [201,] 152.9918 0.3509803
#>  [202,] 151.5679 0.3616226
#>  [203,] 151.5102 0.3620295
#>  [204,] 151.3944 0.3628505
#>  [205,] 151.4902 0.3627218
#>  [206,] 152.4777 0.3578640
#>  [207,] 151.8615 0.3626066
#>  [208,] 150.8739 0.3652306
#>  [209,] 151.0258 0.3663159
#>  [210,] 151.4011 0.3658170
#>  [211,] 146.9730 0.3634298
#>  [212,] 147.1603 0.3655257
#>  [213,] 153.9200 0.3576237
#>  [214,] 153.7473 0.3588154
#>  [215,] 153.8014 0.3586730
#>  [216,] 147.3613 0.3629576
#>  [217,] 147.6573 0.3617883
#>  [218,] 147.9095 0.3600872
#>  [219,] 148.1576 0.3594337
#>  [220,] 147.8962 0.3621455
#>  [221,] 147.9425 0.3623639
#>  [222,] 148.9943 0.3538940
#>  [223,] 148.3464 0.3599745
#>  [224,] 148.3409 0.3606577
#>  [225,] 149.0201 0.3559368
#>  [226,] 148.4852 0.3606397
#>  [227,] 149.4435 0.3535803
#>  [228,] 149.4889 0.3538358
#>  [229,] 149.5968 0.3532453
#>  [230,] 149.7490 0.3529263
#>  [231,] 148.8185 0.3608515
#>  [232,] 148.4985 0.3639447
#>  [233,] 150.0802 0.3508285
#>  [234,] 148.9702 0.3601644
#>  [235,] 150.7193 0.3459658
#>  [236,] 150.1250 0.3510984
#>  [237,] 149.0676 0.3599453
#>  [238,] 149.0583 0.3606397
#>  [239,] 150.3904 0.3502303
#>  [240,] 149.2112 0.3599670
#>  [241,] 149.8065 0.3556178
#>  [242,] 149.6948 0.3565452
#>  [243,] 149.3096 0.3597460
#>  [244,] 149.1858 0.3617002
#>  [245,] 149.8393 0.3565952
#>  [246,] 150.4010 0.3523629
#>  [247,] 150.0513 0.3554803
#>  [248,] 149.5999 0.3594737
#>  [249,] 149.1663 0.3630709
#>  [250,] 151.0865 0.3479146
#>  [251,] 150.9673 0.3488380
#>  [252,] 151.1305 0.3482102
#>  [253,] 150.1124 0.3574499
#>  [254,] 150.3544 0.3573468
#>  [255,] 150.2982 0.3577901
#>  [256,] 150.4388 0.3578797
#>  [257,] 151.4238 0.3506618
#>  [258,] 150.5945 0.3572720
#>  [259,] 150.4815 0.3581512
#>  [260,] 150.4242 0.3585989
#>  [261,] 152.5526 0.3439883
#>  [262,] 150.4516 0.3595662
#>  [263,] 152.8624 0.3421271
#>  [264,] 152.6132 0.3454407
#>  [265,] 152.0400 0.3498133
#>  [266,] 151.0830 0.3568311
#>  [267,] 153.0407 0.3430930
#>  [268,] 152.8986 0.3443806
#>  [269,] 153.2400 0.3432831
#>  [270,] 152.9809 0.3450476
#>  [271,] 152.7771 0.3467596
#>  [272,] 151.2444 0.3579753
#>  [273,] 152.7530 0.3475255
#>  [274,] 150.8820 0.3612284
#>  [275,] 153.1412 0.3464068
#>  [276,] 152.3614 0.3523310
#>  [277,] 152.1171 0.3540347
#>  [278,] 151.1718 0.3608026
#>  [279,] 153.2969 0.3462717
#>  [280,] 153.1671 0.3471333
#>  [281,] 151.3567 0.3608815
#>  [282,] 151.2981 0.3613024
#>  [283,] 152.9142 0.3503217
#>  [284,] 151.3947 0.3611680
#>  [285,] 152.2736 0.3552701
#>  [286,] 152.5675 0.3535460
#>  [287,] 151.7326 0.3607416
#>  [288,] 151.5476 0.3623168
#>  [289,] 152.8685 0.3546851
#>  [290,] 152.4300 0.3592956
#>  [291,] 155.6243 0.3537033
#>  [292,] 147.6365 0.3654986
#>  [293,] 150.9530 0.3689913
#>  [294,] 148.0619 0.3590026
#>  [295,] 148.3308 0.3616976
#>  [296,] 148.1178 0.3642261
#>  [297,] 149.3886 0.3540540
#>  [298,] 149.2774 0.3550046
#>  [299,] 149.2223 0.3554720
#>  [300,] 148.7275 0.3603904
#>  [301,] 149.7087 0.3523026
#>  [302,] 150.2734 0.3482720
#>  [303,] 148.2985 0.3650835
#>  [304,] 149.5252 0.3548078
#>  [305,] 149.4643 0.3556372
#>  [306,] 148.6031 0.3630294
#>  [307,] 149.7958 0.3531828
#>  [308,] 149.6611 0.3555703
#>  [309,] 149.7054 0.3558271
#>  [310,] 150.4940 0.3500322
#>  [311,] 149.2563 0.3602016
#>  [312,] 149.2453 0.3609044
#>  [313,] 150.3608 0.3545530
#>  [314,] 150.2961 0.3553670
#>  [315,] 149.4572 0.3624618
#>  [316,] 151.9514 0.3454977
#>  [317,] 150.3523 0.3597421
#>  [318,] 150.2336 0.3609623
#>  [319,] 152.4253 0.3448991
#>  [320,] 150.3947 0.3600098
#>  [321,] 152.6585 0.3438697
#>  [322,] 152.7643 0.3437543
#>  [323,] 152.9989 0.3427528
#>  [324,] 150.3832 0.3627127
#>  [325,] 152.8401 0.3463204
#>  [326,] 153.4770 0.3423148
#>  [327,] 153.2694 0.3455432
#>  [328,] 150.8479 0.3626255
#>  [329,] 153.1032 0.3475592
#>  [330,] 151.2408 0.3617143
#>  [331,] 151.2795 0.3620025
#>  [332,] 151.4920 0.3610411
#>  [333,] 151.4332 0.3614593
#>  [334,] 152.7414 0.3526557
#>  [335,] 152.4950 0.3543317
#>  [336,] 152.6681 0.3534434
#>  [337,] 151.4135 0.3621541
#>  [338,] 151.4528 0.3624360
#>  [339,] 152.4261 0.3565167
#>  [340,] 155.3828 0.3413728
#>  [341,] 152.6706 0.3604484
#>  [342,] 152.1961 0.3635963
#>  [343,] 149.8747 0.3694423
#>  [344,] 152.6621 0.3580445
#>  [345,] 148.7751 0.3558187
#>  [346,] 149.1216 0.3557050
#>  [347,] 149.3319 0.3545323
#>  [348,] 148.9244 0.3599348
#>  [349,] 150.1437 0.3499770
#>  [350,] 149.6166 0.3553140
#>  [351,] 150.0202 0.3544970
#>  [352,] 150.8855 0.3478808
#>  [353,] 148.9730 0.3634974
#>  [354,] 150.9502 0.3542449
#>  [355,] 150.5379 0.3577119
#>  [356,] 153.0460 0.3411659
#>  [357,] 152.7408 0.3445394
#>  [358,] 152.2878 0.3480468
#>  [359,] 151.0135 0.3596829
#>  [360,] 153.2326 0.3466982
#>  [361,] 152.8490 0.3492746
#>  [362,] 153.4789 0.3453664
#>  [363,] 150.9262 0.3631696
#>  [364,] 151.5875 0.3609207
#>  [365,] 151.9463 0.3598068
#>  [366,] 152.2373 0.3594829
#>  [367,] 151.2524 0.3646926
#>  [368,] 147.7077 0.3613140
#>  [369,] 147.8065 0.3610467
#>  [370,] 152.5219 0.3432688
#>  [371,] 152.5508 0.3584848
#>  [372,] 154.0181 0.3543493
#>  [373,] 153.3981 0.3556080
#>  [374,] 150.1837 0.3506214
#>  [375,] 150.4139 0.3613172
#>  [376,] 154.3017 0.3543467
#>  [377,] 150.2861 0.3504268
#>  [378,] 152.7509 0.3594987
#>  [379,] 161.0655 0.3426522
#>  [380,] 153.2213 0.3561708
#>  [381,] 154.3678 0.3602910
#>  [382,] 153.1357 0.3582573
#>  [383,] 143.3386 0.3287226
#>  [384,] 149.9083 0.3522460
#>  [385,] 152.2018 0.3510680
#>  [386,] 159.6193 0.3527281
#>  [387,] 154.4073 0.3543035
#>  [388,] 152.7613 0.3504657
#>  [389,] 154.2280 0.3550632
#>  [390,] 154.4357 0.3541474
#>  [391,] 158.6121 0.3183638
#>  [392,] 151.4032 0.3591236
#>  [393,] 149.8067 0.3524493
#>  [394,] 150.4479 0.3497595
#>  [395,] 152.7501 0.3589461
#>  [396,] 152.7192 0.3588747
#>  [397,] 153.0995 0.3583763
#>  [398,] 152.7189 0.3586901
#>  [399,] 152.5386 0.3581347
#>  [400,] 153.2149 0.3467078
#>  [401,] 142.3802 0.4074306
#>  [402,] 154.8927 0.3097433
#>  [403,] 153.0167 0.3228820
#>  [404,] 151.4790 0.3033773
#>  [405,] 149.0266 0.2321662
#>  [406,] 149.8786 0.3563480
#>  [407,] 150.2668 0.3572543
#>  [408,] 151.2656 0.3576998
#>  [409,] 152.5317 0.3600561
#>  [410,] 155.7845 0.3355727
#>  [411,] 151.4098 0.3602338
#>  [412,] 145.2874 0.3127545
#>  [413,] 139.6026 0.4108290
#>  [414,] 139.6026 0.4108290
#>  [415,] 135.9025 0.4367475
#>  [416,] 142.3802 0.4074306
#>  [417,] 141.8814 0.4095520
#>  [418,] 141.9472 0.3959075
#>  [419,] 140.3087 0.4059468
#>  [420,] 140.3087 0.4059468
#>  [421,] 140.1681 0.4069420
#>  [422,] 135.6861 0.4404734
#>  [423,] 139.2593 0.4130575
#>  [424,] 139.2593 0.4130575
#>  [425,] 139.0730 0.4140404
#>  [426,] 138.7042 0.4167072
#>  [427,] 137.0637 0.4316748
#>  [428,] 142.5678 0.3940172
#>  [429,] 140.6019 0.4097756
#>  [430,] 156.0165 0.3349083
#>  [431,] 157.1794 0.3585936
#>  [432,] 157.9568 0.3580742
#>  [433,] 149.1023 0.3563641
#>  [434,] 156.7157 0.3586681
#>  [435,] 153.5876 0.3652901
#>  [436,] 152.7126 0.3557091
#>  [437,] 157.2920 0.3498315
#>  [438,] 153.3909 0.3554678
#>  [439,] 153.0959 0.3658052
#>  [440,] 153.1204 0.3583310
#>  [441,] 157.5500 0.3501602
#>  [442,] 153.4230 0.3579666
#>  [443,] 152.8018 0.3577638
#>  [444,] 151.0143 0.3632662
#>  [445,] 153.1599 0.3602152
#>  [446,] 152.7412 0.3574123
#>  [447,] 153.1008 0.3580706
#>  [448,] 153.0746 0.3662502
#>  [449,] 151.9099 0.3636017
#>  [450,] 146.1487 0.3698792
#>  [451,] 152.8672 0.3702795
#>  [452,] 152.5339 0.3586379
#>  [453,] 153.0442 0.3579063
#>  [454,] 150.9857 0.3634122
#>  [455,] 151.6627 0.3652825
#>  [456,] 155.1973 0.3579716
#>  [457,] 153.1238 0.3589006
#>  [458,] 138.5109 0.4252697
#>  [459,] 150.9823 0.2703009
#>  [460,] 140.4724 0.3620506
#>  [461,] 152.5392 0.3583195
#>  [462,] 151.5087 0.3607391
#>  [463,] 152.9188 0.3666041
#>  [464,] 158.6996 0.3498087
#>  [465,] 154.5547 0.3521144
#>  [466,] 154.6563 0.3607390
#>  [467,] 153.1933 0.3539149
#>  [468,] 146.8347 0.3242329
#>  [469,] 147.2965 0.3618389
#>  [470,] 152.8784 0.3583619
#>  [471,] 158.9290 0.3472029
#>  [472,] 157.3220 0.3589896
#>  [473,] 154.8529 0.3605866
#>  [474,] 155.1558 0.3505063
#>  [475,] 144.9905 0.3735879
#>  [476,] 150.5302 0.3637357
#>  [477,] 158.0897 0.3484420
#>  [478,] 158.3107 0.3497998
#>  [479,] 155.2406 0.3511097
#>  [480,] 157.5921 0.3508389
#>  [481,] 154.5955 0.3537496
#>  [482,] 156.8887 0.3505192
#>  [483,] 153.7689 0.3576743
#>  [484,] 153.0882 0.3658641
#>  [485,] 151.6971 0.3635301
#>  [486,] 154.0799 0.3546476
#>  [487,] 154.7146 0.3534498
#>  [488,] 153.6561 0.3571907
#>  [489,] 152.9928 0.3660079
#>  [490,] 158.6121 0.3183638
#>  [491,] 153.0273 0.3428248
#>  [492,] 155.4104 0.3561835
#>  [493,] 154.2146 0.3549347
#>  [494,] 153.6163 0.3563276
#>  [495,] 153.3332 0.3588813
#>  [496,] 152.6696 0.3580874
#>  [497,] 151.0970 0.3573813
#>  [498,] 153.0252 0.3427877
#>  [499,] 153.0273 0.3428248
#>  [500,] 154.3529 0.3567362
#>  [501,] 148.2931 0.3603177
#>  [502,] 155.3789 0.3605307
#>  [503,] 153.6417 0.3572925
#>  [504,] 150.5992 0.3565777
#>  [505,] 158.2239 0.3497400
#>  [506,] 156.2700 0.3494717
#>  [507,] 150.8402 0.3635269
#>  [508,] 151.9363 0.3654624
#>  [509,] 152.9266 0.3427804
#>  [510,] 152.7005 0.3587213
#>  [511,] 157.3500 0.3499729
#>  [512,] 157.2003 0.3499470
#>  [513,] 153.2255 0.3586190
#>  [514,] 153.5151 0.3554443
#>  [515,] 152.0248 0.3679936
#>  [516,] 153.1375 0.3692707
#>  [517,] 153.2595 0.3689238
#>  [518,] 149.8508 0.3650235
#>  [519,] 158.8948 0.3420349
#>  [520,] 153.7601 0.2582494
#>  [521,] 142.3389 0.3171314
#>  [522,] 151.7771 0.2801050
#>  [523,] 154.2462 0.3386256
#>  [524,] 146.4720 0.3198816
#>  [525,] 147.9144 0.3660576
#>  [526,] 150.3453 0.3499501
#>  [527,] 151.8964 0.3634655
#>  [528,] 150.8479 0.3638037
#>  [529,] 153.4860 0.3582398
#>  [530,] 153.4481 0.3668279
#>  [531,] 150.5536 0.3623262
#>  [532,] 153.1866 0.3602415
#>  [533,] 152.7423 0.3587000
#>  [534,] 152.8036 0.3577421
#>  [535,] 152.5284 0.3604044
#>  [536,] 153.8109 0.3673560
#>  [537,] 152.7752 0.3593352
#>  [538,] 152.6250 0.3557359
#>  [539,] 154.2092 0.3573221
#>  [540,] 151.9379 0.3635251
#>  [541,] 151.8667 0.3580805
#>  [542,] 156.0999 0.3353281
#>  [543,] 156.0159 0.3349017
#>  [544,] 153.0273 0.3428248
#>  [545,] 157.9885 0.3485824
#>  [546,] 152.5824 0.3583060
#>  [547,] 152.6915 0.3617154
#>  [548,] 153.0273 0.3428248
#>  [549,] 154.6822 0.3513523
#>  [550,] 157.5412 0.3529683
#>  [551,] 155.3425 0.3609884
#>  [552,] 153.4496 0.3612485
#>  [553,] 155.4498 0.3609514
#>  [554,] 152.5628 0.3595178
#>  [555,] 151.5324 0.3610624
#>  [556,] 152.7242 0.3545464
#>  [557,] 154.4918 0.3568768
#>  [558,] 153.7948 0.3646766
#>  [559,] 151.7381 0.3654095
#>  [560,] 153.0437 0.3429588
#>  [561,] 152.9266 0.3427804
#>  [562,] 156.6593 0.3304695
#>  [563,] 150.5433 0.3266284
#>  [564,] 155.9269 0.3277749
#>  [565,] 157.0336 0.3217120
#>  [566,] 155.9564 0.2980417
#>  [567,] 152.3433 0.2818392
#>  [568,] 151.9786 0.2762808
#>  [569,] 152.5848 0.3131230
#>  [570,] 151.5503 0.2992208
#>  [571,] 153.2588 0.3094446
#>  [572,] 156.8461 0.3020387
#>  [573,] 159.3761 0.3299252
#>  [574,] 159.0130 0.3213739
#>  [575,] 159.5714 0.3530812
#>  [576,] 159.3434 0.3323312
#>  [577,] 153.0685 0.3403275
#>  [578,] 152.8370 0.3238467
#>  [579,] 153.1759 0.3521197
#>  [580,] 152.5947 0.3595058
#>  [581,] 160.7158 0.3457522
#>  [582,] 155.0300 0.3536710
#>  [583,] 150.5810 0.3643194
#>  [584,] 136.3963 0.4399657
#>  [585,] 151.5452 0.2512678
#>  [586,] 149.8861 0.2728013
#>  [587,] 148.3101 0.2540056
#>  [588,] 153.0812 0.3400288
#>  [589,] 157.2104 0.3545050
#>  [590,] 157.4285 0.3532378
#>  [591,] 157.7672 0.3524636
#>  [592,] 157.6138 0.3525952
#>  [593,] 154.8901 0.3599995
#>  [594,] 149.8940 0.3520956
#>  [595,] 149.0889 0.3605479
#>  [596,] 148.9327 0.3600850
#>  [597,] 148.8089 0.3607863
#>  [598,] 148.7918 0.3605760
#>  [599,] 148.7688 0.3608156
#>  [600,] 148.9968 0.3600952
#>  [601,] 148.9833 0.3599163
#>  [602,] 149.0417 0.3600181
#>  [603,] 149.0942 0.3600135
#>  [604,] 149.2359 0.3610962
#>  [605,] 149.2009 0.3603233
#>  [606,] 149.0523 0.3605764
#>  [607,] 149.2351 0.3598081
#>  [608,] 149.3285 0.3596202
#>  [609,] 149.3623 0.3591614
#>  [610,] 148.4321 0.3684665
#>  [611,] 154.6317 0.3551727
#>  [612,] 155.0603 0.3557886
#>  [613,] 156.6882 0.3521695
#>  [614,] 156.9077 0.3505087
#>  [615,] 160.3609 0.3432615
#>  [616,] 160.9650 0.3439902
#>  [617,] 154.8206 0.3554919
#>  [618,] 153.1698 0.3661448
#>  [619,] 156.6242 0.3557737
#>  [620,] 156.9968 0.3554987
#>  [621,] 154.4937 0.3583871
#>  [622,] 148.4090 0.3613815
#>  [623,] 147.2239 0.3244358
#>  [624,] 146.9651 0.3318558
#>  [625,] 152.6594 0.3143264
#>  [626,] 158.2848 0.3162141
#>  [627,] 155.0491 0.3300081
#>  [628,] 152.6650 0.3286071
#>  [629,] 155.8338 0.3633603
#>  [630,] 155.3902 0.3629756
#>  [631,] 152.2158 0.3666849
#>  [632,] 142.7507 0.3885472
#>  [633,] 135.1027 0.4371205
#>  [634,] 136.3573 0.4309758
#>  [635,] 137.8318 0.4149458
#>  [636,] 137.7815 0.4147838
#>  [637,] 143.1894 0.3873973
#>  [638,] 140.7614 0.3979377
#>  [639,] 139.2374 0.4057505
#>  [640,] 138.6367 0.4099774
#>  [641,] 145.0222 0.3739359
#>  [642,] 144.9777 0.3729241
#>  [643,] 136.2871 0.4251004
#>  [644,] 137.2396 0.4192291
#>  [645,] 136.5014 0.4255663
#>  [646,] 136.2962 0.4246908
#>  [647,] 146.5300 0.3761182
#>  [648,] 156.2020 0.3449910
#>  [649,] 154.9866 0.3298709
#>  [650,] 157.8897 0.3157319
#>  [651,] 154.8875 0.3641458
#>  [652,] 154.1819 0.3297168
#>  [653,] 141.4301 0.3236813
#>  [654,] 142.4041 0.3278021
#>  [655,] 142.3589 0.3132990
#>  [656,] 142.5111 0.3251552
#>  [657,] 152.5078 0.2846517
#>  [658,] 156.2732 0.3441161
#>  [659,] 149.0676 0.3609777
#>  [660,] 157.9276 0.3508250
#>  [661,] 157.9011 0.3517586
#>  [662,] 155.9891 0.3579775
#>  [663,] 157.5896 0.3320140
#>  [664,] 156.5775 0.3363340
#>  [665,] 155.3738 0.3325136
#>  [666,] 155.6296 0.3381068
#>  [667,] 154.4410 0.3404303
#>  [668,] 152.8648 0.3438344
#>  [669,] 147.6306 0.3615951
#>  [670,] 143.3738 0.3273748
#>  [671,] 147.0316 0.3214997
#>  [672,] 149.7327 0.3132611
#>  [673,] 146.8309 0.3159652
#>  [674,] 147.4754 0.3353252
#>  [675,] 145.1450 0.3631089
#>  [676,] 153.3617 0.2625764
#>  [677,] 149.0523 0.3605764
#>  [678,] 148.9828 0.3608523
#>  [679,] 160.0472 0.3395209
#>  [680,] 154.0678 0.2582351
#>  [681,] 152.1446 0.2580215
#>  [682,] 157.0626 0.3553591
#>  [683,] 148.1812 0.3109761
#>  [684,] 158.1967 0.3169013
#>  [685,] 151.7011 0.2796279
#>  [686,] 153.5992 0.4046046
#>  [687,] 152.8128 0.3441989
#>  [688,] 152.9508 0.3458969
#>  [689,] 136.7973 0.4286481
#>  [690,] 136.8219 0.4290172
#>  [691,] 137.7099 0.4166070
#>  [692,] 138.2942 0.4124488
#>  [693,] 136.3656 0.4225199
#>  [694,] 155.8581 0.3582635
#>  [695,] 156.2280 0.3384724
#>  [696,] 155.9950 0.3378427
#>  [697,] 135.4540 0.4336387
#>  [698,] 134.9933 0.4371963
#>  [699,] 137.1075 0.4176910
#>  [700,] 142.9969 0.3902369
#>  [701,] 160.2602 0.3316954
#>  [702,] 150.2373 0.3474399
#>  [703,] 154.6887 0.3240336
#>  [704,] 152.6892 0.3434131
#>  [705,] 135.8375 0.4302436
#>  [706,] 137.1463 0.4210189
#>  [707,] 135.5796 0.4337160
#>  [708,] 135.3442 0.4336973
#>  [709,] 138.7286 0.4126548
#>  [710,] 147.1206 0.3651348
#>  [711,] 146.4675 0.3669780
#>  [712,] 146.7858 0.3679766
#>  [713,] 150.3617 0.3567083
#>  [714,] 134.4257 0.4422762
#>  [715,] 134.4277 0.4430439
#>  [716,] 136.7598 0.4222303
#>  [717,] 138.5767 0.4121956
#>  [718,] 154.7047 0.3584597
#>  [719,] 154.6626 0.3584440
#>  [720,] 154.5606 0.3590556
#>  [721,] 156.3936 0.3305583
#>  [722,] 156.3875 0.3304905
#>  [723,] 155.1345 0.3627675
#>  [724,] 141.8820 0.3940663
#>  [725,] 145.3563 0.3941142
#>  [726,] 135.1738 0.4362199
#>  [727,] 139.0470 0.4247836
#>  [728,] 139.0375 0.4253320
#>  [729,] 135.3326 0.4346842
#>  [730,] 135.3958 0.4349778
#>  [731,] 154.6774 0.3636998
#>  [732,] 142.6734 0.3900083
#>  [733,] 157.1161 0.3557363
#>  [734,] 147.9741 0.3586736
#>  [735,] 159.6126 0.3531594
#>  [736,] 152.6516 0.3588664
#>  [737,] 136.5968 0.4289455
#>  [738,] 136.5222 0.4290799
#>  [739,] 136.5128 0.4284680
#>  [740,] 136.5309 0.4290731
#>  [741,] 136.6174 0.4273751
#>  [742,] 136.5851 0.4278680
#>  [743,] 136.6017 0.4282448
#>  [744,] 136.6050 0.4283200
#>  [745,] 136.9798 0.4216026
#>  [746,] 138.1597 0.4155005
#>  [747,] 134.8612 0.4381836
#>  [748,] 137.3697 0.4193395
#>  [749,] 136.9878 0.4213513
#>  [750,] 135.3674 0.4341186
#>  [751,] 139.6309 0.4035024
#>  [752,] 139.1205 0.4101777
#>  [753,] 136.0216 0.4288975
#>  [754,] 139.1695 0.4090244
#>  [755,] 139.1610 0.4093826
#>  [756,] 141.1456 0.3974859
#>  [757,] 135.4431 0.4344743
#>  [758,] 138.1597 0.4155005
#>  [759,] 138.0738 0.4161092
#>  [760,] 138.2196 0.4148935
#>  [761,] 141.7638 0.4101916
#>  [762,] 141.7474 0.4100771
#>  [763,] 138.1510 0.4155013
#>  [764,] 144.5712 0.3771095
#>  [765,] 138.2607 0.4148051
#>  [766,] 144.1815 0.3848017
#>  [767,] 138.3437 0.4143705
#>  [768,] 144.3975 0.3821791
#>  [769,] 144.2762 0.3813771
#>  [770,] 146.9515 0.3655271
#>  [771,] 137.1248 0.4206085
#>  [772,] 137.3201 0.4198106
#>  [773,] 136.0562 0.4272106
#>  [774,] 140.8195 0.3987612
#>  [775,] 140.7933 0.3987131
#>  [776,] 136.9509 0.4242027
#>  [777,] 139.6249 0.4041499
#>  [778,] 137.0858 0.4212233
#>  [779,] 135.0623 0.4377579
#>  [780,] 140.7567 0.3988806
#>  [781,] 139.5499 0.4046068
#>  [782,] 135.0720 0.4390657
#>  [783,] 139.4886 0.4063161
#>  [784,] 139.0911 0.4098633
#>  [785,] 140.1033 0.4017604
#>  [786,] 138.7855 0.4113758
#>  [787,] 139.9505 0.4024610
#>  [788,] 139.1326 0.4098319
#>  [789,] 140.3621 0.4055647
#>  [790,] 140.4873 0.4052383
#>  [791,] 140.3926 0.4055447
#>  [792,] 140.5403 0.4050259
#>  [793,] 140.4837 0.4055242
#>  [794,] 140.4749 0.4053949
#>  [795,] 134.9119 0.4381841
#>  [796,] 139.1076 0.4100750
#>  [797,] 140.4633 0.4054962
#>  [798,] 140.2667 0.4013635
#>  [799,] 135.4424 0.4354022
#>  [800,] 139.9702 0.4024559
#>  [801,] 139.3678 0.4079949
#>  [802,] 147.0139 0.3653748
#>  [803,] 146.9937 0.3655658
#>  [804,] 147.0320 0.3653130
#>  [805,] 147.0240 0.3652793
#>  [806,] 143.5366 0.3870903
#>  [807,] 143.4223 0.3930911
#>  [808,] 146.9213 0.3663615
#>  [809,] 159.4375 0.3269499
#>  [810,] 136.3560 0.4282060
#>  [811,] 144.1288 0.3832903
#>  [812,] 150.4448 0.3612745
#>  [813,] 138.6617 0.4121171
#>  [814,] 138.6067 0.4120222
#>  [815,] 150.4722 0.3471778
#>  [816,] 135.1338 0.4370059
#>  [817,] 135.6095 0.4382233
#>  [818,] 148.2378 0.3586068
#>  [819,] 136.7731 0.4232000
#>  [820,] 136.7466 0.4245091
#>  [821,] 136.6830 0.4253527
#>  [822,] 142.8831 0.3907488
#>  [823,] 141.9733 0.3919296
#>  [824,] 143.1740 0.3881297
#>  [825,] 142.8866 0.3892546
#>  [826,] 146.8932 0.3659715
#>  [827,] 153.5282 0.3398819
#>  [828,] 153.5282 0.3398819
#>  [829,] 153.5282 0.3398819
#>  [830,] 134.8623 0.4388335
#>  [831,] 135.2787 0.4367442
#>  [832,] 138.0611 0.4151665
#>  [833,] 142.0835 0.3931537
#>  [834,] 146.8728 0.3650639
#>  [835,] 146.8132 0.3667342
#>  [836,] 147.6008 0.3650128
#>  [837,] 147.6062 0.3651777
#>  [838,] 146.2618 0.3689881
#>  [839,] 146.7515 0.3674323
#>  [840,] 146.2553 0.3688312
#>  [841,] 147.6164 0.3650839
#>  [842,] 147.6164 0.3650839
#>  [843,] 147.5338 0.3654115
#>  [844,] 136.9582 0.4221888
#>  [845,] 147.7332 0.3651962
#>  [846,] 135.5446 0.4347997
#>  [847,] 148.9336 0.3606644
#>  [848,] 146.6062 0.3680608
#>  [849,] 143.5529 0.3875707
#>  [850,] 146.9556 0.3641634
#>  [851,] 135.6913 0.4313974
#>  [852,] 135.7978 0.4307466
#>  [853,] 135.7971 0.4312830
#>  [854,] 135.7041 0.4294571
#>  [855,] 135.9284 0.4294553
#>  [856,] 135.8942 0.4302622
#>  [857,] 135.9237 0.4299239
#>  [858,] 135.9118 0.4302442
#>  [859,] 134.5185 0.4431087
#>  [860,] 136.9751 0.4228311
#>  [861,] 146.6618 0.3677452
#>  [862,] 137.8346 0.4184376
#>  [863,] 151.5595 0.2846775
#>  [864,] 139.2541 0.4089652
#>  [865,] 146.4388 0.3684726
#>  [866,] 136.5359 0.4238535
#>  [867,] 136.7676 0.4232848
#>  [868,] 136.7786 0.4237632
#>  [869,] 136.7708 0.4224276
#>  [870,] 136.5893 0.4238139
#>  [871,] 156.9758 0.3275368
#>  [872,] 139.5155 0.4041945
#>  [873,] 139.6267 0.4049106
#>  [874,] 141.9312 0.3949650
#>  [875,] 139.7618 0.4039575
#>  [876,] 136.3975 0.4269173
#>  [877,] 136.5317 0.4251759
#>  [878,] 155.7571 0.3577757
#>  [879,] 159.8471 0.3303652
#>  [880,] 138.9995 0.4176908
#>  [881,] 143.9862 0.3801344
#>  [882,] 142.1020 0.3945271
#>  [883,] 135.2285 0.4358540
#>  [884,] 146.7477 0.3676869
#>  [885,] 136.2954 0.4266869
#>  [886,] 147.7674 0.3607789
#>  [887,] 147.5778 0.3653299
#>  [888,] 144.0695 0.3805865
#>  [889,] 136.9702 0.4272952
#>  [890,] 136.8454 0.4291555
#>  [891,] 136.9593 0.4274579
#>  [892,] 134.8627 0.4378906
#>  [893,] 134.9164 0.4376687
#>  [894,] 136.5219 0.4260278
#>  [895,] 139.4634 0.4246797
#>  [896,] 155.2613 0.3573267
#>  [897,] 136.6623 0.4304300
#>  [898,] 148.7825 0.3910652
#>  [899,] 135.5671 0.4341787
#>  [900,] 137.3762 0.4197537
#>  [901,] 137.6656 0.4176387
#>  [902,] 137.4359 0.4179691
#>  [903,] 155.9922 0.3262102
#>  [904,] 152.3762 0.3675560
#>  [905,] 143.8524 0.3847349
#>  [906,] 135.3877 0.4343179
#>  [907,] 137.5702 0.4151876
#>  [908,] 146.2744 0.3786461
#>  [909,] 159.6156 0.3528896
#>  [910,] 148.7356 0.3911839
#>  [911,] 153.2198 0.3453655
#>  [912,] 144.0764 0.3811874
#>  [913,] 144.2788 0.3806829
#>  [914,] 143.9247 0.3814879
#>  [915,] 150.4960 0.3617628
#>  [916,] 141.2952 0.3468934
#>  [917,] 145.5100 0.3898619
#>  [918,] 135.3513 0.4345119
#>  [919,] 137.5985 0.4195055
#>  [920,] 139.8227 0.4031452
#>  [921,] 158.4293 0.3529163
#>  [922,] 144.3972 0.3803074
#>  [923,] 144.3884 0.3804005
#>  [924,] 144.4001 0.3796077
#>  [925,] 144.2271 0.3811261
#>  [926,] 144.3057 0.3801704
#>  [927,] 134.9316 0.4372822
#>  [928,] 146.5508 0.3238118
#>  [929,] 135.7439 0.4304198
#>  [930,] 155.3914 0.3589612
#>  [931,] 146.3817 0.3738767
#>  [932,] 154.1821 0.3294964
#>  [933,] 134.9928 0.4387935
#>  [934,] 159.6823 0.3522836
#>  [935,] 159.6478 0.3523898
#>  [936,] 159.5670 0.3539026
#>  [937,] 159.5606 0.3524425
#>  [938,] 159.6784 0.3523776
#>  [939,] 159.5591 0.3525776
#>  [940,] 159.5747 0.3538777
#>  [941,] 159.6681 0.3527140
#>  [942,] 159.5644 0.3525721
#>  [943,] 159.5845 0.3528275
#>  [944,] 159.6131 0.3526520
#>  [945,] 159.6306 0.3523785
#>  [946,] 159.5578 0.3526179
#>  [947,] 159.6160 0.3526392
#>  [948,] 159.6046 0.3528259
#>  [949,] 159.6868 0.3522984
#>  [950,] 159.6544 0.3526358
#>  [951,] 159.6232 0.3529593
#>  [952,] 159.5919 0.3537947
#>  [953,] 159.5794 0.3526096
#>  [954,] 159.5696 0.3526952
#>  [955,] 159.4790 0.3525930
#>  [956,] 159.5626 0.3523480
#>  [957,] 159.6044 0.3524668
#>  [958,] 159.6232 0.3528307
#>  [959,] 159.6074 0.3526167
#>  [960,] 159.5536 0.3527049
#>  [961,] 159.5600 0.3524965
#>  [962,] 159.6396 0.3522451
#>  [963,] 159.6300 0.3528528
#>  [964,] 159.5627 0.3523820
#>  [965,] 159.4798 0.3526067
#>  [966,] 159.4798 0.3526067
#>  [967,] 159.5693 0.3527222
#>  [968,] 159.6258 0.3526485
#>  [969,] 159.6523 0.3523706
#>  [970,] 152.6491 0.3553789
#>  [971,] 150.6735 0.3798815
#>  [972,] 151.9181 0.3636640
#>  [973,] 150.6021 0.3565961
#>  [974,] 150.1915 0.2824525
#>  [975,] 153.0355 0.3673426
#>  [976,] 153.8145 0.3780821
#>  [977,] 153.0934 0.3710422
#>  [978,] 147.5928 0.3834436
#>  [979,] 148.6837 0.2662357
#>  [980,] 153.5085 0.3661439
#>  [981,] 152.6304 0.3554118
#>  [982,] 155.3050 0.3608119
#>  [983,] 156.4246 0.3579740
#>  [984,] 152.9054 0.3621475
#>  [985,] 152.9332 0.3654786
#>  [986,] 151.5936 0.3612188
#>  [987,] 156.4430 0.3579187
#>  [988,] 153.2026 0.3582706
#>  [989,] 150.5227 0.3643993
#>  [990,] 153.2638 0.3592331
#>  [991,] 152.9266 0.3427804
#>  [992,] 151.3661 0.3544673
#>  [993,] 147.2965 0.3618389
#>  [994,] 148.2931 0.3603177
#>  [995,] 158.9290 0.3472029
#>  [996,] 158.0897 0.3484420
#>  [997,] 158.3107 0.3497998
#>  [998,] 154.6822 0.3513523
#>  [999,] 154.1192 0.3150238
#> [1000,] 155.2406 0.3511097
#> [1001,] 154.0181 0.3543493
#> [1002,] 154.3529 0.3567362
#> [1003,] 154.0799 0.3546476
#> [1004,] 157.5836 0.3508818
#> [1005,] 157.2003 0.3499470
#> [1006,] 152.7197 0.3545099
#> [1007,] 155.1558 0.3505063
#> [1008,] 156.4430 0.3579187
#> [1009,] 155.3050 0.3608119
#> [1010,] 154.6563 0.3607390
#> [1011,] 154.2092 0.3573221
#> [1012,] 153.1204 0.3583310
#> [1013,] 152.5947 0.3595058
#> [1014,] 154.3678 0.3602910
#> [1015,] 152.5752 0.3583743
#> [1016,] 151.4166 0.3602319
#> [1017,] 151.5936 0.3612188
#> [1018,] 151.6971 0.3635301
#> [1019,] 152.7435 0.3587103
#> [1020,] 153.6663 0.3570887
#> [1021,] 152.7117 0.3553329
#> [1022,] 160.2844 0.3419420
#> [1023,] 150.5227 0.3643993
#> [1024,] 139.9549 0.3671944
#> [1025,] 147.2140 0.3741711
#> [1026,] 151.4963 0.2533425
#> [1027,] 158.9723 0.3441427
#> [1028,] 150.5740 0.3262443
#> [1029,] 158.9122 0.3135782
#> [1030,] 157.8279 0.3149726
#> [1031,] 139.8690 0.3684728
#> [1032,] 152.6017 0.2835856
#> [1033,] 155.8042 0.3492171
#> [1034,] 159.6280 0.3374824
#> [1035,] 155.3104 0.3400288
#> [1036,] 150.6394 0.3260711
#> [1037,] 144.2302 0.3328927
#> [1038,] 144.3262 0.3484702
#> [1039,] 146.8159 0.3730136
#> [1040,] 147.1047 0.3734368
#> [1041,] 147.1744 0.3722822
#> [1042,] 151.6322 0.3027065
#> [1043,] 151.5713 0.2951049
#> [1044,] 155.9779 0.3448431
#> [1045,] 156.5666 0.3199680
#> [1046,] 156.6357 0.3260226
#> [1047,] 146.2469 0.3681369
#> [1048,] 156.1018 0.3433714
#> [1049,] 159.0779 0.3540632
#> [1050,] 156.8946 0.3558370
#> [1051,] 147.3728 0.3760481
#> [1052,] 148.3399 0.2842077
#> [1053,] 153.0810 0.2646600
#> [1054,] 153.2959 0.2642742
#> [1055,] 149.8371 0.3565820
#> [1056,] 148.6011 0.3568065
#> [1057,] 149.0946 0.3563224
#> [1058,] 149.6201 0.3549987
#> [1059,] 149.5784 0.3549194
#> [1060,] 155.4818 0.3486591
#> [1061,] 154.0024 0.3584351
#> [1062,] 154.7841 0.3635903
#> [1063,] 154.2951 0.3630343
#> [1064,] 154.8736 0.3602016
#> [1065,] 155.0668 0.3584553
#> [1066,] 150.4423 0.3519183
#> [1067,] 156.4437 0.3524014
#> [1068,] 157.2493 0.3521582
#> [1069,] 158.3291 0.3511230
#> [1070,] 159.7447 0.3456787
#> [1071,] 159.9900 0.3462997
#> [1072,] 155.5009 0.3338073
#> [1073,] 157.7057 0.3517924
#> [1074,] 156.4777 0.3566397
#> [1075,] 149.9625 0.3513173
#> [1076,] 153.6872 0.3239119
#> [1077,] 150.7141 0.3255717
#> [1078,] 143.2990 0.3007475
#> [1079,] 154.9507 0.3653530
#> [1080,] 154.7585 0.3643397
#> [1081,] 151.5040 0.3655759
#> [1082,] 146.7088 0.3188623