Azimuth conversion from PoR to geographical coordinate reference system

Conversion of PoR azimuths into geographical azimuths

Usage

PoR2Geo_azimuth(x, PoR)

Arguments

x

data.frame containing the PoR equivalent azimuths (azi.PoR), and either the geographical coordinates of the point(s) or the PoR-equivalent coordinates.

PoR

data.frame containing the geographical location of the Euler pole (lat, lon)

Value

numeric vector of transformed azimuths (in degrees)

References

Stephan, T., Enkelmann, E., and Kroner, U. "Analyzing the horizontal orientation of the crustal stress adjacent to plate boundaries". Sci Rep 13. 15590 (2023). doi:10.1038/s41598-023-42433-2 .

See also

Examples

data("nuvel1")
# North America relative to Pacific plate:
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
data("san_andreas")
san_andreas$azi.PoR <- PoR_shmax(san_andreas, PoR)

# convert back to geo CRS
PoR2Geo_azimuth(san_andreas, PoR)
#>   [1]  50  54  24  41  30  27  29  58  32  47  34  17  16   0 168   1  27 178
#>  [19] 145 129 152 103 165  37  18  21   8  47  34   2 175 167 168 156 157   8
#>  [37] 173 174  24  40  17   8 177 178   8  37  18  51  60  73 160  37  55  15
#>  [55]  11  20 177  44 128  38  31   1  85  16 176  48 175 172 157 175 170 171
#>  [73] 172 168 172  33  21 177  18 171 175 161 173  21  28  18  40  30  35 110
#>  [91]  35 178 171  42  20  24   8 134 163 174 175  19 169 171  14   6  10 105
#> [109] 173  26  56  18  19 164 158 177  17   0   3 168  12  23 174   7  14  53
#> [127]  54  36  21  16  40  13  14  22  43  18  16  12 172 171 178 144 179  16
#> [145]   8 171 170   4 176   6   6   7   6  14  15 164   7 170 173 179  17  29
#> [163]  28 179   9  18  30 141 139 140 150 149 149 160 161 161 174 171 172 175
#> [181] 172 169 169 173 174 175 175 174 169 175 177   1   1   4   7  11   8   8
#> [199]  11  15  16  31 178 168  11 174  18  27 172  44  22   4 162 169  41 178
#> [217] 167 176 168 175 158 178   1 152 172 167 176 165 168 174 164 175   6   1
#> [235]  14 178 178 173 153  14  13 178  37 172 172 176  28 175 172  24   0  14
#> [253]   1 177 176  20 179 170  14 178   6 168   5 171 172  13 178 172   3  26
#> [271] 171 173 174 171 156 176 174  11   4 176 160 171 174 179 177   5 175   5
#> [289] 168 165 171  21 163 176  27 177 174 172  12   7 176 171 175 167 171 171
#> [307] 174 159 179 175   7 163 178 178 169 159 175  12   7   2 166 178 167 174
#> [325] 168 173   6  17 156  33  25 174 176 168  44  30  14 175 178 162  15  61
#> [343]  47  30  40  24   3  16  38   2 172 167  23  26  15   7   8  19 174  30
#> [361]   5   3 125 127 146  58  55  27 172  27 175  41  25  48  38  33  21  35
#> [379]  38  18  23  22  19  25  24  17  42  30 175  38  24  28  97  30  13  18
#> [397] 177  12  17  45  14   8  94  85  31  13  26