Spatial interpolation of SHmax in PoR coordinate reference system

The data is transformed into the PoR system before the interpolation. The interpolation grid is returned in geographical coordinates and azimuths.

Usage

PoR_stress2grid(
  x,
  PoR,
  grid = NULL,
  PoR_grid = TRUE,
  lon_range = NULL,
  lat_range = NULL,
  gridsize = 2.5,
  ...
)

PoR_stress2grid_stats(
  x,
  PoR,
  grid = NULL,
  PoR_grid = TRUE,
  lon_range = NULL,
  lat_range = NULL,
  gridsize = 2.5,
  ...
)

Arguments

x

sf object containing

azi

SHmax in degree

unc

Uncertainties of SHmax in degree

type

Methods used for the determination of the orientation of SHmax

PoR

Pole of Rotation. "data.frame" or object of class "euler.pole" containing the geographical coordinates of the Euler pole

grid

(optional) Point object of class sf.

PoR_grid

logical. Whether the grid should be generated based on the coordinate range in the PoR (TRUE, the default) CRS or the geographical CRS (FALSE). Is ignored if grid is specified.

lon_range, lat_range

(optional) numeric vector specifying the minimum and maximum longitudes and latitudes (are ignored if "grid" is specified).

gridsize

Numeric. Target spacing of the regular grid in decimal degree. Default is 2.5 (is ignored if grid is specified)

...

Arguments passed to stress2grid()

Value

sf object containing

lon,lat

longitude and latitude in geographical CRS (in degrees)

lon.PoR,lat.PoR

longitude and latitude in PoR CRS (in degrees)

azi

geographical mean SHmax in degree

azi.PoR

PoR mean SHmax in degree

sd

Standard deviation of SHmax in degrees

R

Search radius in km

mdr

Mean distance of datapoints per search radius

N

Number of data points in search radius

Details

Stress field and wavelength analysis in PoR system and back-transformed

See also

Examples

data("san_andreas")
data("nuvel1")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
PoR_stress2grid(san_andreas, PoR)
#> Simple feature collection with 979 features and 10 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -125.0802 ymin: 21.44743 xmax: -106.3859 ymax: 41.36318
#> Geodetic CRS:  WGS 84
#> # A tibble: 979 × 11
#>    lon.PoR lat.PoR azi.PoR    sd     R     N   mdr             geometry
#>  *   <dbl>   <dbl>   <dbl> <dbl> <dbl> <int> <dbl>          <POINT [°]>
#>  1   -84.8    52.6    160.  24.6   450    46 0.922 (-125.0802 34.07892)
#>  2   -84.8    52.6    156.  22.8   500    85 0.887 (-125.0802 34.07892)
#>  3   -84.8    52.6    157.  21.1   550   187 0.891 (-125.0802 34.07892)
#>  4   -84.8    52.6    150.  21.1   600   298 0.867 (-125.0802 34.07892)
#>  5   -84.8    52.6    143.  21.4   650   385 0.837 (-125.0802 34.07892)
#>  6   -84.8    52.6    142.  19.7   700   483 0.817 (-125.0802 34.07892)
#>  7   -84.8    52.6    141.  18.7   750   566 0.791 (-125.0802 34.07892)
#>  8   -84.8    52.6    139.  18.4   800   703 0.786 (-125.0802 34.07892)
#>  9   -84.8    52.6    139.  18.3   850   772 0.760 (-125.0802 34.07892)
#> 10   -84.8    52.6    139.  18.1   900   818 0.732 (-125.0802 34.07892)
#> # ℹ 969 more rows
#> # ℹ 3 more variables: lat <dbl>, lon <dbl>, azi <dbl>

if (FALSE) { # \dontrun{
PoR_stress2grid_stats(san_andreas, PoR)
} # }