Conversion between PoR to geographical coordinate system

Transform spherical objects from PoR to geographical coordinate system and vice versa.

Usage

PoR_to_geographical(x, ep)

geographical_to_PoR(x, ep)

Arguments

x

sf, SpatRast, or Raster* object of the data points in geographical or PoR coordinate system

ep

data.frame of the geographical coordinates of the Euler pole (lat, lon)

Value

sf or SpatRast object of the data points in the transformed geographical or PoR coordinate system

Details

The PoR coordinate reference system is oblique transformation of the geographical coordinate system with the Euler pole coordinates being the translation factors.

Examples

data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate
data("san_andreas")
san_andreas.por <- geographical_to_PoR(san_andreas, euler)
PoR_to_geographical(san_andreas.por, euler)
#> Simple feature collection with 1082 features and 9 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -72 ymin: -156.92 xmax: -54.18 ymax: -140.06
#> Geodetic CRS:  WGS 84
#> Warning: bounding box has potentially an invalid value range for longlat data
#> Warning: bounding box has potentially an invalid value range for longlat data
#> # A tibble: 1,082 × 10
#>    id         lat   lon   azi   unc type  depth quality regime
#>  * <chr>    <dbl> <dbl> <dbl> <dbl> <chr> <dbl> <fct>   <chr> 
#>  1 wsm00967  35.2 -119.     6   3   BO     1.91 D       NA    
#>  2 wsm14489  32.8 -120.    16   4.3 BO     1.6  D       NA    
#>  3 wsm01115  34.4 -119.    61   5   BO     1.32 D       NA    
#>  4 wsm01130  34.3 -117.    90   5   HFM    2.09 D       S     
#>  5 wsm19348  38   -111.   133   5   FMA    5    D       N     
#>  6 wsm06768  34.7 -120.    65   6   BOT    1.7  D       TS    
#>  7 wsm14285  34.5 -122.    90   9.2 BO     4.2  D       NA    
#>  8 wsm06770  34.6 -121.    36  13   BOT    1.7  D       TS    
#>  9 wsm14481  31.3 -118.    13  15   BO     2.21 D       NA    
#> 10 wsm19347  38.4 -110.   102  15   FMA    0    D       S     
#> # … with 1,072 more rows, and 1 more variable: geometry <POINT [°]>