Error of Model's Prediction

The maximum error in the model's predicted azimuth given the Pole of rotations uncertainty and distance of the data point to the pole.

Usage

prd_err(dist_PoR, sigma_PoR = 1)

Arguments

dist_PoR

Distance to Euler pole (great circle distance, in degree)

sigma_PoR

uncertainty of the position of the Pole of rotation (in degree).

Value

numeric vector. The maximum error for azimuths prediction (in degree)

References

Ramsay, J.A. Folding and fracturing of rocks. McGraw-Hill, New York, 1967.

See also

Examples

prd_err(67, 1)
#> [1] 0.9060055

# San Andreas example:
data("nuvel1")
por <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to Pacific plate
data("san_andreas")
d <- PoR_distance(san_andreas, por)
prd_err(d)
#>    [1] 1.875505 1.965582 1.894943 1.916741 2.045546 2.175192 2.209802 2.007877
#>    [9] 2.096593 2.018212 2.016872 1.862409 1.905727 1.750007 1.850889 1.857785
#>   [17] 1.769859 1.768023 1.774618 1.783184 1.800640 1.802095 1.776316 1.761912
#>   [25] 1.762615 1.755257 1.782133 1.761423 1.782107 1.780452 1.763312 1.759568
#>   [33] 1.947893 1.765357 1.765840 1.824262 1.824262 1.815155 1.780276 1.777478
#>   [41] 1.769883 1.784044 1.978590 1.780871 1.774094 1.758692 1.758692 1.771977
#>   [49] 1.767687 1.774253 1.786006 1.785762 1.784634 1.738223 1.730387 1.735391
#>   [57] 1.756827 1.761458 1.761458 1.760530 1.739973 1.766817 1.770227 1.769974
#>   [65] 1.762454 1.762454 1.748638 1.767302 1.763609 1.764480 1.763722 1.761737
#>   [73] 1.767809 1.779657 1.768495 1.763244 1.758941 1.762846 1.767668 1.763633
#>   [81] 1.781683 1.767496 1.815155 1.796643 1.780276 1.768385 1.777478 1.770524
#>   [89] 1.801586 1.728522 1.777228 1.769883 1.771626 1.753501 1.784044 1.768506
#>   [97] 1.773353 1.780871 1.756819 1.734721 1.774094 1.769296 1.757515 1.771977
#>  [105] 1.771434 1.767860 1.729227 1.774253 1.784634 1.770955 1.738223 1.745781
#>  [113] 1.782059 1.774354 1.781138 1.731343 1.730387 1.738045 1.735391 1.742803
#>  [121] 1.772870 1.772280 1.728712 1.793972 1.716292 1.773067 1.771532 1.758818
#>  [129] 1.761996 1.762401 1.767334 1.756464 1.758895 1.756827 1.758388 1.763040
#>  [137] 1.771965 1.760973 1.784246 1.760469 1.767107 1.760802 1.719625 1.767712
#>  [145] 1.743191 1.721687 1.760530 1.768593 1.739973 1.773223 1.774037 1.772965
#>  [153] 1.767359 1.769474 1.764037 1.757867 1.747656 1.757458 1.745400 1.758311
#>  [161] 1.766087 1.745831 1.746723 1.759084 1.745934 1.725087 1.719195 1.721563
#>  [169] 1.721805 1.756396 1.748638 1.800500 1.782618 1.769243 1.758427 1.822317
#>  [177] 1.761700 1.789596 1.749054 1.758395 1.758395 1.720707 1.702913 1.703682
#>  [185] 1.758069 1.713185 1.721614 1.735424 1.737924 1.739475 1.764153 1.614144
#>  [193] 1.802019 1.654084 1.977195 1.754257 1.799664 1.645197 2.335817 1.795004
#>  [201] 1.805558 1.967184 1.944338 1.944534 1.971969 2.345506 2.264876 2.394754
#>  [209] 1.859227 1.751417 1.756697 1.759007 1.759037 1.737496 1.835636 1.800754
#>  [217] 1.805859 1.803892 1.805362 1.804379 1.804773 1.806000 1.804493 1.803686
#>  [225] 1.795362 1.800260 1.801174 1.802628 1.802232 1.804287 1.766332 1.764573
#>  [233] 1.758058 1.765888 1.773014 1.795126 1.790512 1.761418 1.719310 1.748068
#>  [241] 1.747498 1.748919 1.807127 2.038624 2.002278 2.000512 1.935308 1.891320
#>  [249] 1.924844 1.714002 1.718683 1.725783 1.763546 1.682073 1.678617 1.737152
#>  [257] 1.739949 1.759792 1.757719 1.752624 1.743797 1.799241 1.806270 1.722097
#>  [265] 1.728431 1.711057 1.724608 1.755487 1.805261 1.892617 1.940339 1.715998
#>  [273] 2.274501 1.759609 1.997183 1.901457 1.950723 2.193475 2.139952 2.225192
#>  [281] 2.203774 2.152258 2.193227 2.158069 2.075703 2.345416 2.431821 2.257723
#>  [289] 2.162175 2.490279 1.776513 1.757720 1.766130 1.776316 1.798221 1.809217
#>  [297] 1.786480 1.798666 1.812203 1.801532 1.785699 1.758235 1.766417 1.761917
#>  [305] 1.824903 1.856353 1.740549 1.851067 1.861486 1.846503 1.875480 1.839960
#>  [313] 1.844643 1.837397 1.842464 1.851860 1.834286 1.843037 1.854324 1.885725
#>  [321] 1.841440 1.770047 1.819518 2.114948 2.059164 2.051797 1.967097 2.094476
#>  [329] 2.148550 2.080530 2.048068 1.972185 2.060176 2.108089 1.966703 1.860955
#>  [337] 1.992702 1.736648 1.744839 1.753104 1.743009 1.751271 1.707970 1.707970
#>  [345] 1.654604 1.654084 1.810276 1.653014 1.738100 1.719033 1.719033 1.716649
#>  [353] 1.636144 1.703701 1.703701 1.702938 1.720548 1.696726 1.725363 1.728891
#>  [361] 1.716982 1.768313 1.648903 1.760512 1.765974 1.734022 1.685773 2.076865
#>  [369] 2.258505 2.117987 1.807579 1.801174 1.800721 1.813778 1.929178 1.898103
#>  [377] 2.270118 2.300910 1.758427 1.747869 2.093629 1.932493 2.196956 1.511395
#>  [385] 1.840087 1.829491 1.679140 1.675719 1.730059 1.738492 1.737105 1.857898
#>  [393] 1.855115 1.857936 1.714148 1.699972 1.693838 1.707326 1.731132 1.730515
#>  [401] 1.739893 1.838180 1.843123 1.693357 1.685774 1.743564 1.746916 1.751028
#>  [409] 1.750630 1.757098 1.757737 1.757973 1.753087 1.752987 1.753256 1.751222
#>  [417] 1.850543 1.856269 1.925701 1.845700 1.702832 1.719454 1.688049 1.697136
#>  [425] 1.723015 1.808490 1.810331 1.798308 1.803218 1.671599 1.665956 1.724812
#>  [433] 1.730961 1.824113 1.732995 1.807932 1.865044 1.816601 1.827376 1.816191
#>  [441] 1.819738 1.817925 1.824976 1.823161 1.821312 1.819501 1.826665 1.815765
#>  [449] 1.812157 1.822913 1.821067 1.819259 1.826145 1.822532 1.815287 1.811609
#>  [457] 1.811368 1.809542 1.809301 1.839765 1.832539 1.816198 1.814446 1.810713
#>  [465] 1.796536 1.812403 1.812143 1.810329 1.838759 1.835201 1.806650 1.827846
#>  [473] 1.810068 1.795838 1.788842 1.831118 1.829296 1.827510 1.809803 1.806134
#>  [481] 1.799115 1.834305 1.832519 1.830735 1.828952 1.827206 1.807732 1.832165
#>  [489] 1.826863 1.807463 1.805736 1.837312 1.833784 1.823159 1.809053 1.802050
#>  [497] 1.795105 1.830065 1.824742 1.805395 1.803673 1.851139 1.787873 1.838513
#>  [505] 1.836734 1.834994 1.833218 1.829706 1.827899 1.826164 1.808633 1.806841
#>  [513] 1.805121 1.803368 1.801616 1.843640 1.822470 1.798018 1.794527 1.791041
#>  [521] 1.839870 1.838131 1.834584 1.832848 1.831077 1.829308 1.827539 1.806560
#>  [529] 1.804808 1.857344 1.844766 1.841217 1.839481 1.837710 1.835939 1.821925
#>  [537] 1.820166 1.804492 1.802780 1.825249 1.814777 1.842553 1.840820 1.839052
#>  [545] 1.819813 1.818094 1.800720 1.793666 1.786774 1.840383 1.838617 1.821209
#>  [553] 1.819421 1.817706 1.815991 1.802106 1.827952 1.807145 1.838216 1.824275
#>  [561] 1.820774 1.817349 1.813857 1.812149 1.803518 1.801816 1.800081 1.823899
#>  [569] 1.822153 1.820406 1.813539 1.811764 1.810060 1.803186 1.801453 1.798060
#>  [577] 1.796296 1.844507 1.806432 1.792737 1.789325 1.785919 1.821741 1.819999
#>  [585] 1.818293 1.816588 1.814850 1.811410 1.809675 1.807976 1.801120 1.796009
#>  [593] 1.794284 1.826692 1.842059 1.821327 1.819624 1.817887 1.816185 1.814485
#>  [601] 1.812716 1.809321 1.807625 1.802287 1.798909 1.788700 1.785308 1.843305
#>  [609] 1.841558 1.839847 1.838136 1.819209 1.817511 1.810656 1.807236 1.805545
#>  [617] 1.841089 1.839347 1.808567 1.806844 1.805156 1.803470 1.815159 1.798159
#>  [625] 1.811575 1.803045 1.799649 1.818156 1.784294 1.802653 1.800974 1.810957
#>  [633] 1.807560 1.797429 1.800582 1.798908 1.798515 1.796846 1.799784 1.798118
#>  [641] 1.796418 1.843397 1.789576 1.797717 1.799177 1.795655 1.793964 1.822104
#>  [649] 1.808732 1.796872 1.795217 1.793563 1.791876 1.848431 1.845078 1.841691
#>  [657] 1.788180 1.786534 1.834424 1.781411 1.854215 1.844218 1.842514 1.840843
#>  [665] 1.786140 1.784465 1.850560 1.847218 1.837214 1.827233 1.794157 1.790834
#>  [673] 1.787583 1.853588 1.851918 1.843617 1.849627 1.848000 1.846302 1.844638
#>  [681] 1.843012 1.841315 1.839654 1.838029 1.836335 1.785309 1.783641 1.776952
#>  [689] 1.773697 1.845685 1.844025 1.842404 1.839091 1.835743 1.852005 1.825572
#>  [697] 1.819014 1.792794 1.786265 1.846717 1.845065 1.843409 1.838149 1.769626
#>  [705] 1.842790 1.839493 1.857252 1.840212 1.836928 1.833680 1.828464 1.826826
#>  [713] 1.842795 1.810320 1.784458 1.778032 1.824603 1.823006 1.769815 1.771223
#>  [721] 1.820790 1.767782 1.867397 1.827915 1.824703 1.815085 1.811848 1.763959
#>  [729] 1.833713 1.801672 1.795303 1.776196 1.769874 1.823762 1.822181 1.820567
#>  [737] 1.818987 1.817371 1.812601 1.810955 1.809377 1.807801 1.826652 1.759957
#>  [745] 1.815170 1.813595 1.811952 1.808804 1.807199 1.811368 1.803454 1.767152
#>  [753] 1.765567 1.764015 1.806305 1.799973 1.768463 1.796534 1.765121 1.761959
#>  [761] 1.812016 1.786840 1.774275 1.794387 1.792830 1.764638 1.763060 1.761516
#>  [769] 1.759973 1.758398 1.801962 1.795700 1.789412 1.792279 1.790693 1.764153
#>  [777] 1.762611 1.761071 1.759498 1.757925 1.756387 1.760556 1.759019 1.757483
#>  [785] 1.755916 1.763419 1.758539 1.757006 1.755474 1.778421 1.772203 1.766027
#>  [793] 1.783488 1.770608 1.764508 1.751787 1.766483 1.763415 1.829889 1.754674
#>  [801] 1.742715 1.764762 1.720135 1.796946 1.739055 1.757236 1.755480 1.749252
#>  [809] 1.747531 1.741370 1.739589 1.727504 1.821172 1.805481 1.746888 1.747294
#>  [817] 1.744901 1.877054 1.877444 1.721441 1.751451 1.670847 1.685755 1.631921
#>  [825] 1.628519 1.690576 1.686081 1.719814 1.756403 1.650603 1.693660 1.670468
#>  [833] 1.745718 1.830095 1.749538 1.763366 1.814193 1.746518 1.746319 1.684177
#>  [841] 1.685909 1.690503 1.685646 1.694348 1.692081 1.688871 1.688231 1.721305
#>  [849] 1.722450 1.683860 1.723157 1.722747 1.693261 1.755171 1.727075 1.705418
#>  [857] 1.733170 1.731066 1.749767 1.687851 1.722450 1.721196 1.724554 1.651912
#>  [865] 1.653066 1.722734 1.790189 1.723796 1.752362 1.723993 1.763434 1.771042
#>  [873] 1.809406 1.740508 1.675639 2.235486 2.252153 2.323304 2.208434 2.250865
#>  [881] 1.811516 1.829866 1.824471 1.823633 1.824717 1.791671 1.752748 1.719613
#>  [889] 1.726186 1.736567 1.744214 1.828556 1.766138 1.762950 1.763315 1.785115
#>  [897] 1.781510 1.867741 1.762605 1.744501 1.838924 1.937244 1.968640 2.268880
#>  [905] 1.709301 1.715899 1.761975 1.739703 1.723015 1.721661 1.715924 1.750778
#>  [913] 1.751832 1.704446 1.751121 1.720173 1.679083 1.751800 1.750431 1.669351
#>  [921] 1.741110 1.730076 1.752253 1.729714 1.752265 1.728986 1.719970 1.718485
#>  [929] 1.719219 1.718344 1.716735 1.717824 1.681877 1.728158 1.717500 1.749991
#>  [937] 1.681298 1.751707 1.733796 1.809107 1.808380 1.809123 1.809471 1.753321
#>  [945] 1.718975 1.805130 1.876552 1.697981 1.762773 1.776767 1.728745 1.731148
#>  [953] 1.854250 1.681688 1.654981 1.825852 1.717656 1.709569 1.706009 1.746544
#>  [961] 1.762493 1.755638 1.755746 1.807934 1.855091 1.855091 1.855091 1.679158
#>  [969] 1.677990 1.727967 1.751884 1.813602 1.805025 1.800386 1.799339 1.802635
#>  [977] 1.802140 1.803687 1.799695 1.799695 1.799292 1.718042 1.796970 1.681703
#>  [985] 1.802570 1.801280 1.749976 1.817248 1.700206 1.700785 1.697038 1.713375
#>  [993] 1.704974 1.700588 1.701861 1.700057 1.661880 1.713030 1.802046 1.713450
#> [1001] 2.172937 1.730922 1.802146 1.721669 1.717269 1.713578 1.723060 1.720014
#> [1009] 1.887954 1.754189 1.746091 1.744474 1.748209 1.705441 1.712669 1.743022
#> [1017] 1.858376 1.680898 1.785286 1.742755 1.686254 1.800728 1.710766 1.821736
#> [1025] 1.798971 1.780596 1.682444 1.673937 1.681701 1.685902 1.685393 1.707119
#> [1033] 1.620191 1.748656 1.671531 1.631355 1.685250 1.720110 1.724575 1.730203
#> [1041] 1.902354 1.719411 1.760340 1.691071 1.744444 1.745464 1.750825 1.631324
#> [1049] 1.829387 1.776781 1.775198 1.778481 1.773368 2.059785 1.693474 1.691083
#> [1057] 1.714182 1.751663 1.754152 1.774787 1.774420 1.778958 1.773687 1.777692
#> [1065] 1.687565 2.056655 1.705106 1.708631 1.731093 1.730139 1.725093 1.721120
#> [1073] 1.741719 1.739338 1.771529 1.902603 1.674362 1.922708 1.922708 2.080973
#> [1081] 1.753582 1.753150 1.746072 1.753100 1.753140 1.752454 1.746174 1.751550
#> [1089] 1.752468 1.751195 1.751972 1.753244 1.752264 1.752026 1.751156 1.753501
#> [1097] 1.751956 1.750474 1.746534 1.752253 1.751865 1.752572 1.753549 1.752881
#> [1105] 1.751091 1.752154 1.751856 1.752842 1.753864 1.750969 1.753386 1.752504
#> [1113] 1.752504 1.751736 1.751959 1.753232 1.844090 1.846352 1.845205 1.844985
#> [1121] 1.844985 1.844985 1.844985 1.846352 1.846352 1.751592