Apply Rolling Functions using Circular Statistics

A generic function for applying a function to rolling margins of an array.

Usage

roll_circstats(
  x,
  w = NULL,
  FUN,
  axial = TRUE,
  na.rm = TRUE,
  width = NULL,
  by.column = FALSE,
  partial = TRUE,
  fill = NA,
  ...
)

Arguments

x

numeric vector. Values in degrees.

w

(optional) Weights. A vector of positive numbers and of the same length as x.

FUN

the function to be applied

axial

logical. Whether the data are axial, i.e. pi-periodical (TRUE, the default) or directional, i.e. \(2 \pi\)-periodical (FALSE).

na.rm

logical value indicating whether NA values in x should be stripped before the computation proceeds.

width

integer specifying the window width (in numbers of observations) which is aligned to the original sample according to the align argument. If NULL, an optimal width is calculated.

by.column

logical. If TRUE, FUN is applied to each column separately.

partial

logical or numeric. If FALSE then FUN is only applied when all indexes of the rolling window are within the observed time range. If TRUE (default), then the subset of indexes that are in range are passed to FUN. A numeric argument to partial can be used to determine the minimal window size for partial computations. See below for more details.

fill

a three-component vector or list (recycled otherwise) providing filling values at the left/within/to the right of the data range. See the fill argument of zoo::na.fill() for details

...

optional arguments passed to zoo::rollapply()

Value

numeric vector with the results of the rolling function.

Note

If the rolling statistics are applied to values that are a function of distance it is recommended to sort the values first.

Examples

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")
data("san_andreas")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
distance <- distance_from_pb(
  x = san_andreas,
  PoR = PoR,
  pb = plate_boundary,
  tangential = TRUE
)
dat <- san_andreas[order(distance), ]
roll_circstats(dat$azi, w = 1 / dat$unc, circular_mean, width = 51)
#>    [1] 1.799880e+02 1.781512e+02 1.131229e+00 2.766417e+00 3.293337e+00
#>    [6] 1.786585e+02 2.748376e-01 1.758607e+02 1.791289e+02 1.973633e+00
#>   [11] 5.458127e-01 3.985705e+00 6.240060e+00 8.770752e+00 9.678773e+00
#>   [16] 1.146947e+01 1.416107e+01 1.573097e+01 1.562831e+01 1.451829e+01
#>   [21] 1.448511e+01 1.644570e+01 1.772161e+01 1.742273e+01 1.826384e+01
#>   [26] 1.887892e+01 1.769970e+01 1.936875e+01 1.967131e+01 1.941892e+01
#>   [31] 1.746316e+01 1.727473e+01 1.802694e+01 1.985330e+01 2.073245e+01
#>   [36] 1.975227e+01 1.801281e+01 1.930633e+01 1.970018e+01 2.019271e+01
#>   [41] 2.010805e+01 2.050523e+01 2.098142e+01 2.155941e+01 2.292787e+01
#>   [46] 2.246985e+01 2.135827e+01 2.279009e+01 2.297850e+01 2.784309e+01
#>   [51] 2.824444e+01 2.784523e+01 2.839384e+01 2.911217e+01 2.828130e+01
#>   [56] 2.845896e+01 2.783420e+01 2.814176e+01 2.742388e+01 2.703529e+01
#>   [61] 2.611435e+01 2.693504e+01 2.670314e+01 2.600720e+01 2.466867e+01
#>   [66] 2.540523e+01 2.445506e+01 2.343434e+01 2.326515e+01 2.315653e+01
#>   [71] 2.248538e+01 2.172197e+01 1.969198e+01 1.847852e+01 2.026018e+01
#>   [76] 1.782430e+01 1.688439e+01 2.496475e+01 2.429514e+01 2.317905e+01
#>   [81] 2.320207e+01 2.364674e+01 2.260598e+01 2.197656e+01 2.141891e+01
#>   [86] 2.074532e+01 2.128483e+01 1.850251e+01 1.618467e+01 1.527770e+01
#>   [91] 1.516811e+01 1.333039e+01 1.347116e+01 1.151906e+01 1.100356e+01
#>   [96] 1.081148e+01 1.122993e+01 1.104133e+01 1.105287e+01 1.099289e+01
#>  [101] 1.070947e+01 1.087397e+01 1.061126e+01 1.128583e+01 1.128447e+01
#>  [106] 1.085420e+01 1.079475e+01 1.088460e+01 1.067403e+01 1.049175e+01
#>  [111] 1.028740e+01 1.027058e+01 1.073530e+01 1.133633e+01 1.177272e+01
#>  [116] 1.204600e+01 1.250348e+01 1.297342e+01 1.318875e+01 1.245051e+01
#>  [121] 1.247102e+01 1.264049e+01 1.245431e+01 1.291211e+01 1.363148e+01
#>  [126] 1.338545e+01 1.362220e+01 1.298660e+01 1.266328e+01 1.274863e+01
#>  [131] 1.285213e+01 1.290217e+01 1.281707e+01 1.275876e+01 1.266398e+01
#>  [136] 1.240606e+01 1.273939e+01 1.220073e+01 1.262823e+01 1.279233e+01
#>  [141] 1.263225e+01 1.258973e+01 1.286795e+01 1.320074e+01 1.438252e+01
#>  [146] 1.456402e+01 1.448672e+01 1.458973e+01 1.461149e+01 1.431967e+01
#>  [151] 1.492198e+01 1.482893e+01 1.492050e+01 1.484305e+01 1.458222e+01
#>  [156] 1.463573e+01 1.500154e+01 1.501138e+01 1.496293e+01 1.530617e+01
#>  [161] 1.518945e+01 1.553217e+01 1.558169e+01 1.535852e+01 1.471882e+01
#>  [166] 1.411903e+01 1.399138e+01 1.383255e+01 1.211596e+01 1.142042e+01
#>  [171] 1.180731e+01 1.169308e+01 1.165315e+01 1.115569e+01 1.107278e+01
#>  [176] 1.041848e+01 1.041665e+01 1.064084e+01 1.116088e+01 1.100559e+01
#>  [181] 1.055595e+01 1.045103e+01 1.059499e+01 1.057232e+01 1.005224e+01
#>  [186] 1.049944e+01 1.012811e+01 1.102465e+01 1.148942e+01 1.162378e+01
#>  [191] 1.145402e+01 1.172641e+01 1.189999e+01 1.339250e+01 1.298544e+01
#>  [196] 1.243287e+01 1.244368e+01 1.264826e+01 1.201444e+01 1.201502e+01
#>  [201] 1.161483e+01 1.118454e+01 1.145677e+01 1.145677e+01 1.204067e+01
#>  [206] 1.201657e+01 1.239217e+01 1.243788e+01 1.253644e+01 1.251777e+01
#>  [211] 1.258188e+01 1.275712e+01 1.260710e+01 1.249972e+01 1.268060e+01
#>  [216] 1.292445e+01 1.338116e+01 1.374992e+01 1.375100e+01 1.456712e+01
#>  [221] 1.531486e+01 1.564990e+01 1.583155e+01 1.518691e+01 1.641446e+01
#>  [226] 1.646060e+01 1.739376e+01 1.762968e+01 1.787638e+01 1.779503e+01
#>  [231] 1.739276e+01 1.821232e+01 1.900040e+01 1.776594e+01 1.548916e+01
#>  [236] 1.685417e+01 1.666292e+01 1.878193e+01 1.782351e+01 1.783822e+01
#>  [241] 1.379040e+01 1.423716e+01 1.403078e+01 1.315955e+01 1.008111e+01
#>  [246] 1.041979e+01 1.053829e+01 1.040386e+01 1.000931e+01 9.823722e+00
#>  [251] 1.135727e+01 1.310124e+01 1.192704e+01 1.123435e+01 1.095682e+01
#>  [256] 7.479820e+00 7.263160e+00 6.217212e+00 6.624945e+00 6.638394e+00
#>  [261] 6.781584e+00 6.048499e+00 5.540599e+00 6.630181e+00 6.033846e+00
#>  [266] 5.988906e+00 6.156661e+00 5.841055e+00 5.232358e+00 4.780816e+00
#>  [271] 5.842031e+00 5.819441e+00 5.615381e+00 7.279058e+00 7.623736e+00
#>  [276] 7.455062e+00 7.369186e+00 6.002120e+00 5.828278e+00 5.679322e+00
#>  [281] 5.905394e+00 5.731843e+00 6.001546e+00 4.258318e+00 5.620625e+00
#>  [286] 8.231576e+00 7.861391e+00 8.173763e+00 4.363588e+00 4.785064e+00
#>  [291] 5.011090e+00 1.233088e+01 1.093207e+01 1.116102e+01 1.367717e+01
#>  [296] 1.619605e+01 1.520776e+01 1.535508e+01 1.555553e+01 1.570950e+01
#>  [301] 1.842607e+01 1.554088e+01 1.192475e+01 1.437725e+01 1.514423e+01
#>  [306] 1.511182e+01 2.417627e+01 2.602840e+01 2.610815e+01 2.548231e+01
#>  [311] 2.500739e+01 2.509188e+01 2.514154e+01 2.519705e+01 2.306891e+01
#>  [316] 2.551494e+01 2.742633e+01 2.726605e+01 2.682351e+01 2.623808e+01
#>  [321] 2.633512e+01 2.394441e+01 2.454975e+01 2.461304e+01 2.335081e+01
#>  [326] 2.536196e+01 2.575247e+01 2.798735e+01 2.844210e+01 2.841472e+01
#>  [331] 2.730662e+01 2.708765e+01 2.667788e+01 2.535754e+01 2.517798e+01
#>  [336] 2.667133e+01 2.626774e+01 2.630379e+01 2.659778e+01 2.569173e+01
#>  [341] 2.556263e+01 2.534197e+01 2.553321e+01 2.687840e+01 2.739066e+01
#>  [346] 2.750914e+01 2.767783e+01 2.789952e+01 2.778859e+01 2.772366e+01
#>  [351] 2.745970e+01 2.790313e+01 2.780835e+01 2.938502e+01 2.906019e+01
#>  [356] 2.884989e+01 2.871198e+01 2.773374e+01 2.725344e+01 2.699176e+01
#>  [361] 2.643128e+01 2.526742e+01 2.511984e+01 2.494229e+01 2.466107e+01
#>  [366] 2.455512e+01 2.177039e+01 1.897839e+01 1.867336e+01 1.877668e+01
#>  [371] 2.018608e+01 2.013279e+01 2.121176e+01 2.064230e+01 2.025687e+01
#>  [376] 1.995094e+01 1.937118e+01 1.800084e+01 1.638546e+01 1.600870e+01
#>  [381] 1.656635e+01 1.648883e+01 1.684728e+01 1.760401e+01 1.726500e+01
#>  [386] 1.712971e+01 1.474797e+01 1.473575e+01 1.459749e+01 1.420367e+01
#>  [391] 1.458009e+01 1.388021e+01 1.334536e+01 1.309750e+01 1.186605e+01
#>  [396] 1.167404e+01 1.182006e+01 1.182915e+01 1.162849e+01 1.091056e+01
#>  [401] 1.104973e+01 1.104489e+01 1.072397e+01 1.075883e+01 1.075919e+01
#>  [406] 1.084934e+01 1.092478e+01 1.092970e+01 9.988584e+00 1.082323e+01
#>  [411] 1.078290e+01 1.084189e+01 1.298993e+01 1.329179e+01 1.325614e+01
#>  [416] 1.278901e+01 1.352277e+01 1.441930e+01 1.552971e+01 1.524813e+01
#>  [421] 1.529688e+01 1.397210e+01 1.357412e+01 1.382488e+01 1.382488e+01
#>  [426] 1.379192e+01 1.218754e+01 1.208780e+01 1.245716e+01 1.235135e+01
#>  [431] 1.304305e+01 1.242447e+01 1.249762e+01 1.206602e+01 1.095972e+01
#>  [436] 1.151657e+01 1.205539e+01 1.300495e+01 1.317616e+01 1.295743e+01
#>  [441] 1.604230e+01 1.608395e+01 1.667768e+01 1.732321e+01 1.695588e+01
#>  [446] 1.659120e+01 1.676734e+01 1.676057e+01 1.774879e+01 1.810717e+01
#>  [451] 1.915984e+01 1.771417e+01 1.801641e+01 1.791481e+01 1.659633e+01
#>  [456] 1.672105e+01 1.686397e+01 1.731635e+01 1.732594e+01 1.944660e+01
#>  [461] 1.865322e+01 1.873896e+01 1.869832e+01 1.646490e+01 1.622575e+01
#>  [466] 1.660187e+01 1.657974e+01 1.572246e+01 1.798734e+01 1.794468e+01
#>  [471] 1.833958e+01 1.831400e+01 1.829971e+01 1.893606e+01 1.860607e+01
#>  [476] 1.844802e+01 1.870093e+01 1.943455e+01 1.958879e+01 1.740557e+01
#>  [481] 1.757130e+01 1.706734e+01 1.706734e+01 1.677760e+01 1.681961e+01
#>  [486] 1.723637e+01 1.703531e+01 1.656674e+01 1.632617e+01 1.364829e+01
#>  [491] 1.369161e+01 1.053895e+01 9.977180e+00 9.930888e+00 1.002672e+01
#>  [496] 1.093457e+01 1.098853e+01 1.071993e+01 1.120071e+01 9.989677e+00
#>  [501] 9.747484e+00 9.676119e+00 1.064782e+01 1.035964e+01 1.094947e+01
#>  [506] 1.200757e+01 1.153085e+01 1.126268e+01 1.105614e+01 1.103197e+01
#>  [511] 9.602378e+00 9.844189e+00 9.734760e+00 1.064604e+01 1.058250e+01
#>  [516] 1.183315e+01 1.156459e+01 1.054194e+01 1.056127e+01 9.140603e+00
#>  [521] 8.796832e+00 8.847774e+00 8.846254e+00 8.897945e+00 7.774815e+00
#>  [526] 7.958279e+00 7.664496e+00 7.714766e+00 8.349788e+00 7.820774e+00
#>  [531] 8.422774e+00 9.997581e+00 9.673542e+00 1.004982e+01 1.002785e+01
#>  [536] 1.001614e+01 1.023060e+01 1.035592e+01 1.014438e+01 1.092620e+01
#>  [541] 1.359612e+01 1.369688e+01 1.393879e+01 1.456103e+01 1.443067e+01
#>  [546] 1.540580e+01 1.473076e+01 1.514673e+01 1.494364e+01 1.400508e+01
#>  [551] 1.379176e+01 1.219377e+01 1.233701e+01 1.299223e+01 1.384387e+01
#>  [556] 1.308565e+01 1.306299e+01 1.313875e+01 1.462879e+01 1.463816e+01
#>  [561] 1.461762e+01 1.471700e+01 1.374020e+01 1.376564e+01 1.258281e+01
#>  [566] 1.257279e+01 1.149717e+01 1.165666e+01 1.302552e+01 1.334153e+01
#>  [571] 1.260308e+01 1.229989e+01 1.254407e+01 1.238763e+01 1.283407e+01
#>  [576] 1.484315e+01 1.485732e+01 1.519677e+01 1.477802e+01 1.483947e+01
#>  [581] 1.552192e+01 1.458054e+01 1.317315e+01 1.316781e+01 1.187839e+01
#>  [586] 1.172765e+01 1.225241e+01 1.081028e+01 1.014020e+01 9.743085e+00
#>  [591] 9.080135e+00 9.090376e+00 9.198445e+00 9.026651e+00 8.984524e+00
#>  [596] 8.604543e+00 8.010696e+00 7.862417e+00 7.353218e+00 7.289164e+00
#>  [601] 6.632808e+00 6.714605e+00 7.745942e+00 7.730123e+00 7.274176e+00
#>  [606] 6.273326e+00 6.644232e+00 6.720047e+00 6.553316e+00 3.281092e+00
#>  [611] 4.700235e+00 4.925625e+00 4.809669e+00 5.545999e+00 5.825763e+00
#>  [616] 5.967365e+00 7.313789e+00 7.479102e+00 7.166267e+00 7.465615e+00
#>  [621] 7.234727e+00 7.770171e+00 8.101184e+00 7.562180e+00 7.575929e+00
#>  [626] 6.763910e+00 6.777606e+00 6.411826e+00 5.182203e+00 4.808350e+00
#>  [631] 4.667555e+00 4.774163e+00 5.214451e+00 5.492558e+00 5.649903e+00
#>  [636] 5.100519e+00 4.515438e+00 4.515438e+00 4.879461e+00 5.352323e+00
#>  [641] 5.890314e+00 6.118710e+00 5.288996e+00 5.021460e+00 3.986181e+00
#>  [646] 4.299173e+00 4.804148e+00 4.394064e+00 4.556104e+00 4.589471e+00
#>  [651] 4.847157e+00 5.807504e+00 6.118049e+00 4.773599e+00 4.668971e+00
#>  [656] 4.309788e+00 4.379198e+00 3.723744e+00 4.208290e+00 4.030432e+00
#>  [661] 4.126913e+00 2.864514e+00 2.747053e+00 3.284830e+00 3.253745e+00
#>  [666] 3.106205e+00 3.073123e+00 1.728069e+00 1.761572e+00 1.131342e+00
#>  [671] 8.184120e-01 9.202961e-01 1.109964e+00 1.643495e-01 6.533627e-01
#>  [676] 7.467960e-01 4.239402e-01 3.566573e-01 2.357998e-01 5.480811e-01
#>  [681] 1.124773e+00 1.098368e+00 1.042847e+00 5.386757e-01 1.797037e+02
#>  [686] 1.788387e+02 1.796287e+02 1.026974e+00 8.795439e-02 1.798562e+02
#>  [691] 3.904979e-02 6.280985e-01 5.048608e-01 9.233595e-01 9.496267e-01
#>  [696] 1.505929e+00 1.374655e+00 1.667046e+00 1.711375e+00 1.751840e+00
#>  [701] 1.715371e+00 1.500778e+00 1.577838e+00 1.522879e+00 2.145265e+00
#>  [706] 8.308535e-01 2.108894e-01 2.486243e-01 4.830500e-01 1.148648e-01
#>  [711] 5.189493e-01 4.195617e-01 3.050743e-01 2.092210e-01 1.799013e+02
#>  [716] 1.798816e+02 1.796452e+02 1.793997e+02 1.788302e+02 1.787895e+02
#>  [721] 1.790237e+02 1.798099e+02 1.240426e-03 7.264005e-01 1.014093e+00
#>  [726] 6.561114e-01 5.498937e-01 6.773649e-01 7.502621e-01 7.405619e-01
#>  [731] 8.099665e-01 4.330812e-01 5.808066e-01 1.794003e+02 6.981542e-01
#>  [736] 7.461776e-01 1.509192e+00 1.725548e+00 1.237290e+00 1.474792e+00
#>  [741] 1.756545e+00 1.936324e+00 1.421891e+00 1.421891e+00 1.737877e+00
#>  [746] 1.539881e+00 2.292026e+00 2.375179e+00 2.319394e+00 2.248823e+00
#>  [751] 2.203371e+00 2.343951e+00 2.591935e+00 2.620891e+00 2.624906e+00
#>  [756] 2.792902e+00 4.445214e+00 4.841163e+00 4.256415e+00 3.726351e+00
#>  [761] 4.497594e+00 4.810848e+00 4.865264e+00 4.988340e+00 5.314843e+00
#>  [766] 5.867952e+00 5.617265e+00 5.345209e+00 5.358555e+00 6.939653e+00
#>  [771] 6.764275e+00 7.201039e+00 6.831910e+00 6.868331e+00 6.655097e+00
#>  [776] 6.645628e+00 6.776590e+00 6.812577e+00 6.955041e+00 6.648690e+00
#>  [781] 5.699973e+00 5.575963e+00 5.228286e+00 5.313893e+00 6.654591e+00
#>  [786] 5.810105e+00 6.448007e+00 6.181917e+00 6.285345e+00 6.273628e+00
#>  [791] 6.500180e+00 6.656357e+00 6.727815e+00 6.591495e+00 6.554501e+00
#>  [796] 6.013403e+00 5.907945e+00 5.341938e+00 5.230735e+00 5.285473e+00
#>  [801] 5.265491e+00 5.125526e+00 5.192677e+00 5.006897e+00 4.837833e+00
#>  [806] 4.806472e+00 4.674524e+00 3.956121e+00 3.827366e+00 4.062318e+00
#>  [811] 5.201649e+00 4.221949e+00 2.693400e+00 2.117927e+00 2.387436e+00
#>  [816] 2.243379e+00 1.680031e+00 2.263555e+00 2.547437e+00 2.011555e+00
#>  [821] 4.583436e-01 5.056362e-01 1.798115e+02 1.796558e+02 1.796779e+02
#>  [826] 1.793244e+02 1.793248e+02 1.793493e+02 1.794409e+02 1.796438e+02
#>  [831] 1.779002e+02 1.789037e+02 1.788854e+02 1.792837e+02 1.789293e+02
#>  [836] 1.788514e+02 1.786779e+02 1.785417e+02 1.784929e+02 1.783049e+02
#>  [841] 1.793631e+02 1.788174e+02 1.792981e+02 1.795400e+02 1.799014e+02
#>  [846] 3.125548e-01 5.211763e-01 1.158986e+00 1.108312e+00 1.115938e+00
#>  [851] 1.125063e+00 1.028954e+00 1.067692e+00 9.045859e-01 9.619556e-01
#>  [856] 1.064568e+00 1.150212e+00 1.750837e+00 1.970893e+00 2.026458e+00
#>  [861] 2.171555e+00 2.272500e+00 2.237594e+00 2.338908e+00 2.740262e+00
#>  [866] 2.435474e+00 2.496719e+00 2.596066e+00 2.522239e+00 3.073246e+00
#>  [871] 3.540555e+00 3.432889e+00 3.475369e+00 4.129814e+00 4.097047e+00
#>  [876] 4.633086e+00 4.699319e+00 4.835089e+00 4.524650e+00 4.578149e+00
#>  [881] 5.084190e+00 6.981916e+00 6.922827e+00 7.417738e+00 7.470067e+00
#>  [886] 7.611638e+00 7.630554e+00 8.087721e+00 7.923419e+00 7.967702e+00
#>  [891] 7.598443e+00 6.945283e+00 7.111511e+00 7.348870e+00 6.953052e+00
#>  [896] 6.778698e+00 6.514898e+00 6.673636e+00 6.556584e+00 7.671907e+00
#>  [901] 7.775740e+00 7.451732e+00 7.926613e+00 7.886241e+00 8.068007e+00
#>  [906] 8.167266e+00 8.596204e+00 8.267529e+00 8.331379e+00 9.046439e+00
#>  [911] 8.835083e+00 9.398025e+00 9.200402e+00 9.216610e+00 1.047596e+01
#>  [916] 1.136240e+01 1.182432e+01 1.259084e+01 1.254141e+01 1.293858e+01
#>  [921] 1.269595e+01 1.258540e+01 1.276020e+01 1.279014e+01 1.248257e+01
#>  [926] 1.384048e+01 1.291676e+01 1.299308e+01 1.387046e+01 1.478040e+01
#>  [931] 1.471961e+01 1.347434e+01 1.355366e+01 1.375809e+01 1.367673e+01
#>  [936] 1.363441e+01 1.406732e+01 1.408979e+01 1.398278e+01 1.391121e+01
#>  [941] 1.530749e+01 1.577085e+01 1.573294e+01 1.546499e+01 1.519603e+01
#>  [946] 1.560111e+01 1.613818e+01 1.611255e+01 1.632684e+01 1.647172e+01
#>  [951] 1.567221e+01 1.582570e+01 1.581771e+01 1.541661e+01 1.562614e+01
#>  [956] 1.550925e+01 1.559495e+01 1.576204e+01 1.608194e+01 1.547470e+01
#>  [961] 1.461644e+01 1.499830e+01 1.477640e+01 1.332603e+01 1.321403e+01
#>  [966] 1.285827e+01 1.212689e+01 1.182475e+01 1.129861e+01 1.106447e+01
#>  [971] 1.056999e+01 1.005395e+01 1.014917e+01 1.032291e+01 1.096254e+01
#>  [976] 1.229897e+01 1.170471e+01 1.159099e+01 1.170934e+01 1.012262e+01
#>  [981] 9.534551e+00 1.133671e+01 1.174107e+01 1.197712e+01 1.243706e+01
#>  [986] 1.240396e+01 1.248519e+01 1.209039e+01 1.223257e+01 1.256003e+01
#>  [991] 1.215132e+01 9.298501e+00 9.517991e+00 9.675070e+00 9.923424e+00
#>  [996] 1.035724e+01 1.007828e+01 9.570058e+00 9.619503e+00 9.721428e+00
#> [1001] 9.891135e+00 9.721547e+00 9.785430e+00 1.005501e+01 1.013047e+01
#> [1006] 1.043993e+01 1.163788e+01 1.181874e+01 1.182961e+01 1.186722e+01
#> [1011] 1.199590e+01 1.246620e+01 1.225702e+01 1.263839e+01 1.509644e+01
#> [1016] 1.556423e+01 1.617880e+01 1.597049e+01 1.568262e+01 1.545105e+01
#> [1021] 1.578898e+01 1.659009e+01 1.728775e+01 1.779125e+01 1.776288e+01
#> [1026] 1.698353e+01 1.649453e+01 1.664272e+01 1.791085e+01 1.781659e+01
#> [1031] 1.879613e+01 1.912975e+01 1.709963e+01 1.738739e+01 1.666520e+01
#> [1036] 1.465385e+01 1.448049e+01 1.553228e+01 1.522446e+01 1.476212e+01
#> [1041] 1.395364e+01 1.497291e+01 1.507018e+01 1.514037e+01 1.450909e+01
#> [1046] 1.461455e+01 1.412849e+01 1.479050e+01 1.573114e+01 1.673712e+01
#> [1051] 1.683507e+01 1.669723e+01 1.656704e+01 1.662889e+01 1.692914e+01
#> [1056] 1.773524e+01 1.782575e+01 1.640447e+01 1.604180e+01 1.438028e+01
#> [1061] 1.431477e+01 1.436222e+01 1.419186e+01 1.444722e+01 1.422901e+01
#> [1066] 1.343121e+01 1.292993e+01 1.118368e+01 1.172612e+01 1.202209e+01
#> [1071] 1.250369e+01 1.191427e+01 9.430511e+00 9.117685e+00 8.257524e+00
#> [1076] 7.584101e+00 7.819139e+00 8.380583e+00 6.512324e+00 6.473596e+00
#> [1081] 7.683783e+00 8.048409e+00