Apply Rolling Functions using Circular Statistics

A generic function for applying a function to rolling margins of an array.

Usage

roll_circstats(
  x,
  w = NULL,
  FUN,
  axial = TRUE,
  na.rm = TRUE,
  width = NULL,
  by.column = FALSE,
  partial = TRUE,
  fill = NA,
  ...
)

Arguments

x

numeric vector. Values in degrees.

w

(optional) Weights. A vector of positive numbers and of the same length as x.

FUN

the function to be applied

axial

logical. Whether the data are axial, i.e. pi-periodical (TRUE, the default) or directional, i.e. \(2 \pi\)-periodical (FALSE).

na.rm

logical value indicating whether NA values in x should be stripped before the computation proceeds.

width

integer specifying the window width (in numbers of observations) which is aligned to the original sample according to the align argument. If NULL, an optimal width is calculated.

by.column

logical. If TRUE, FUN is applied to each column separately.

partial

logical or numeric. If FALSE then FUN is only applied when all indexes of the rolling window are within the observed time range. If TRUE (default), then the subset of indexes that are in range are passed to FUN. A numeric argument to partial can be used to determine the minimal window size for partial computations. See below for more details.

fill

a three-component vector or list (recycled otherwise) providing filling values at the left/within/to the right of the data range. See the fill argument of zoo::na.fill() for details

...

optional arguments passed to zoo::rollapply()

Value

numeric vector with the results of the rolling function.

Note

If the rolling statistics are applied to values that are a function of distance it is recommended to sort the values first.

Examples

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")
data("san_andreas")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
distance <- distance_from_pb(
  x = san_andreas,
  PoR = PoR,
  pb = plate_boundary,
  tangential = TRUE
)
dat <- san_andreas[order(distance), ]
roll_circstats(dat$azi, w = 1 / dat$unc, circular_mean, width = 51)
#>   [1]  26.0719516  26.5591873  26.5898087  25.9146558  26.8033304  27.7032645
#>   [7]  27.0783010  28.7199053  29.5983841  30.8478051  30.6375966  29.7701367
#>  [13]  28.3894133  28.1754624  28.7135982  28.3721800  27.3510014  26.9263843
#>  [19]  26.4217232  25.6874853  26.5083729  25.8397437  25.4272631  24.4455125
#>  [25]  23.4548601  22.7597399  22.3429995  22.1688608  22.1688608  21.6284451
#>  [31]  21.0816265  21.0344202  21.7463600  21.6080906  21.3842729  21.8593229
#>  [37]  21.4200591  22.0740838  22.6171474  21.9034489  20.9299709  19.6644415
#>  [43]  18.3430813  18.5209684  17.8231899  16.9681458  18.4165589  18.6906939
#>  [49]  17.9405574  17.6431213  17.1011541  18.4455322  17.6191800  18.0282207
#>  [55]  18.4228480  17.5471994  17.5095353  18.0627379  18.0784242  17.9691226
#>  [61]  17.4713026  17.4220972  17.4220972  18.3199119  18.1415494  17.6343856
#>  [67]  17.6343856  18.3294991  18.7517265  19.5113059  20.0524365  19.4504084
#>  [73]  19.2134442  19.4735830  20.1605497  20.3572559  20.9626915  21.1853149
#>  [79]  21.9505739  22.1001677  22.6177589  23.0914325  23.9691738  23.6532479
#>  [85]  24.0168695  24.1511968  23.4093070  23.3828010  22.4037547  21.8951095
#>  [91]  21.9006088  21.9294810  22.1478279  22.8791658  22.0846583  22.3820153
#>  [97]  22.1297205  21.8122277  21.6668816  22.7274080  23.1386437  22.7454742
#> [103]  21.5502875  21.2574507  20.1591526  19.2467720  18.9927088  18.9136832
#> [109]  18.3956279  17.2542312  16.4544383  16.0461739  16.3149301  15.8200377
#> [115]  14.8937252  14.7362145  14.5005158  14.4228251  13.7289582  13.5781451
#> [121]  12.7968577  12.2371824  11.5776569  11.9789581  11.4406322  11.3033638
#> [127]  11.2828890  10.5142654  10.2664602   9.4674082   9.6037713   8.7518428
#> [133]   8.2666557   7.4331128   7.1279225   6.1172912   5.5005888  16.2356162
#> [139]  16.2356162  19.1895464  18.9219703  18.8074191  22.7336802  23.3359493
#> [145]  22.9805695  22.9506099  22.3640280  22.4051200  22.1419580  21.5607304
#> [151]  20.8367888  20.2647829  20.4783942  20.4251651  20.5621658  20.5065620
#> [157]  20.9724337  20.9724337  20.5954218  20.5207713  20.5120871  20.3990981
#> [163]  20.2336466  19.9690389  19.9973819  20.0081724  19.7917883  19.5325853
#> [169]  19.1844825  19.2271669  18.8356448  19.0501531  19.1659018  19.1768304
#> [175]  18.8860372  18.9264960  18.8468201  18.8232520  18.8495399  18.6220407
#> [181]  18.9903973  18.5561368  18.5660874  18.9358260  19.0129856  18.9793707
#> [187]  21.2643060  21.5066325  13.1306996  12.6243867  10.5276260  10.4776070
#> [193]  11.2737693   6.9271312   6.1960987   6.0268606   6.1526285   6.1526285
#> [199]   6.6511439   6.1724491   6.0673448   6.3229016   6.2260528   6.2572450
#> [205]   6.3322685   6.2472384   6.5323555   6.0018462   5.8818967   5.5649082
#> [211]   5.4736273   5.4736273   5.4575530   5.5928472   5.0885772   6.2455809
#> [217]   6.3945105   6.3805784   6.2971455   6.3986914   6.3042922   6.9947254
#> [223]   6.9487056   6.7394338   6.6979912   6.6839272   6.5122020   6.2890654
#> [229]   6.2890654   6.5101255   6.7676829   6.3807030   6.9368494   7.0812053
#> [235]   7.2378672   7.5119076   7.6579841   5.1987763   5.3745663   5.4197978
#> [241]   5.4796641   3.2475216   3.4716167   2.2246067   2.3194021   2.4457729
#> [247]   2.7087417   3.2600861   3.7874835   3.4253023   3.4969486   4.1479554
#> [253]   3.7593883   3.7418291   3.1648685   3.2384039   3.1642573   3.4078913
#> [259]   3.5509406   3.7825011   4.3588684   5.1671436   5.1671436   4.8112058
#> [265]   5.0516673   5.5156187   4.1270077   3.9060393   3.7853368   4.0156879
#> [271]   3.6538297   3.7345727   2.7282241   2.4233258   2.1440642   1.9879462
#> [277]   2.0310717   2.7014653   3.1166364   3.7559005   4.3265058   4.0113283
#> [283]   4.3868107   4.2325660   4.7323996   4.8949320   4.8734470   5.0331557
#> [289]   5.3796762   4.6479729   5.0566649   4.9902437   4.4424502   4.0977893
#> [295]   3.6354974   3.3613795   3.5019165   2.7655950   1.8165322   1.5168388
#> [301]   0.8967593   0.4045310 179.8275046 179.7823676   0.1609225   0.3204255
#> [307] 179.9459337 179.6397068 179.3048780 178.8472406 178.5368495 177.9214804
#> [313] 177.2868858 178.1197004 178.9860195 178.4308344 178.4509432 178.6487188
#> [319] 178.2238616 178.2015372 178.4872398 178.1666004 177.6927316 177.9040026
#> [325] 178.7174243 179.6018550   0.4254896   0.5756712 179.8075176 178.8079179
#> [331] 178.6069623 177.7171788 177.5695204 177.0379287 177.1760110 176.5515986
#> [337] 175.6541279 175.4006415 175.8006162 174.7473161 174.9365082 174.5542628
#> [343] 174.0602407 175.0545008 174.8704757 174.4913192 174.3974851 174.5082281
#> [349] 174.7657083 175.0358281 175.2500797 175.2097325 176.3646730 176.3646730
#> [355] 176.4367725 176.0526134 175.6235855 176.6804999 177.7364847 177.2956412
#> [361] 177.5572966 177.4386665 177.0437021 177.2268652 176.7607661 176.0979930
#> [367] 176.3468903 175.9176190 175.2162542 175.5054379 175.5517000 175.5517000
#> [373] 176.4977803 177.0717778 176.4086505 175.6695127 175.4067639 175.6230584
#> [379] 175.7598435 176.7663739 178.2755123 178.5904709 178.8322139 179.0573355
#> [385] 179.4896745 178.7313089 178.8303827 179.2623920 179.1169337 178.1514442
#> [391] 178.9125458 178.8232349 178.8600592 179.9208365 179.1924391 179.9141498
#> [397]   0.9737624   1.6988428   1.3425649   2.1150382   2.7451598   1.7668518
#> [403]   3.2543389   1.3289640   1.5752454   2.3243921   3.1942258