Apply Rolling Functions using Circular Statistics

A generic function for applying a function to rolling margins of an array.

Usage

roll_circstats(
  x,
  w = NULL,
  FUN,
  axial = TRUE,
  na.rm = TRUE,
  width = NULL,
  by.column = FALSE,
  partial = TRUE,
  fill = NA,
  ...
)

Arguments

x

numeric vector. Values in degrees.

w

(optional) Weights. A vector of positive numbers and of the same length as x.

FUN

the function to be applied

axial

logical. Whether the data are axial, i.e. pi-periodical (TRUE, the default) or directional, i.e. \(2 \pi\)-periodical (FALSE).

na.rm

logical value indicating whether NA values in x should be stripped before the computation proceeds.

width

integer specifying the window width (in numbers of observations) which is aligned to the original sample according to the align argument. If NULL, an optimal width is calculated.

by.column

logical. If TRUE, FUN is applied to each column separately.

partial

logical or numeric. If FALSE then FUN is only applied when all indexes of the rolling window are within the observed time range. If TRUE (default), then the subset of indexes that are in range are passed to FUN. A numeric argument to partial can be used to determine the minimal window size for partial computations. See below for more details.

fill

a three-component vector or list (recycled otherwise) providing filling values at the left/within/to the right of the data range. See the fill argument of zoo::na.fill() for details

...

optional arguments passed to zoo::rollapply()

Value

numeric vector with the results of the rolling function.

Note

If the rolling statistics are applied to values that are a function of distance it is recommended to sort the values first.

Examples

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")
data("san_andreas")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
distance <- distance_from_pb(
  x = san_andreas,
  PoR = PoR,
  pb = plate_boundary,
  tangential = TRUE
)
dat <- san_andreas[order(distance), ]
roll_circstats(dat$azi, w = 1 / dat$unc, circular_mean, width = 51)
#>    [1] 1.724601e+02 1.754018e+02 1.785472e+02 1.766702e+02 1.797096e+02
#>    [6] 1.298709e+00 1.977206e+00 1.772320e+02 1.789699e+02 1.742201e+02
#>   [11] 1.775516e+02 4.990087e-01 1.791327e+02 2.752321e+00 5.089828e+00
#>   [16] 7.674313e+00 8.644423e+00 1.051984e+01 1.339802e+01 1.506398e+01
#>   [21] 1.498504e+01 1.381015e+01 1.372673e+01 1.572840e+01 1.706427e+01
#>   [26] 1.678509e+01 1.785922e+01 1.854090e+01 1.834575e+01 1.936875e+01
#>   [31] 1.967131e+01 1.941892e+01 1.746316e+01 1.727473e+01 1.802694e+01
#>   [36] 1.985330e+01 2.073245e+01 1.975227e+01 1.801281e+01 1.930633e+01
#>   [41] 1.970018e+01 2.019271e+01 2.010805e+01 2.050523e+01 2.098142e+01
#>   [46] 2.155941e+01 2.292787e+01 2.246985e+01 2.135827e+01 2.279009e+01
#>   [51] 2.297850e+01 2.784309e+01 2.824444e+01 2.784523e+01 2.839384e+01
#>   [56] 2.911217e+01 2.828130e+01 2.845896e+01 2.783420e+01 2.814176e+01
#>   [61] 2.742388e+01 2.703529e+01 2.611435e+01 2.693504e+01 2.670314e+01
#>   [66] 2.600720e+01 2.466867e+01 2.540523e+01 2.445506e+01 2.343434e+01
#>   [71] 2.326515e+01 2.315653e+01 2.248538e+01 2.210288e+01 2.065676e+01
#>   [76] 1.959055e+01 2.002060e+01 1.840215e+01 1.740098e+01 1.977531e+01
#>   [81] 1.750693e+01 1.659217e+01 2.485734e+01 2.501685e+01 2.481565e+01
#>   [86] 2.381217e+01 2.292208e+01 2.202509e+01 2.212079e+01 2.260659e+01
#>   [91] 2.344424e+01 2.293214e+01 2.329718e+01 1.914217e+01 1.559815e+01
#>   [96] 1.510324e+01 1.435598e+01 1.396812e+01 1.275891e+01 1.305937e+01
#>  [101] 1.318872e+01 1.113146e+01 1.063375e+01 1.075567e+01 1.080398e+01
#>  [106] 1.036508e+01 1.021172e+01 1.002661e+01 1.016035e+01 1.045512e+01
#>  [111] 1.019787e+01 1.084363e+01 1.100784e+01 1.076890e+01 1.050166e+01
#>  [116] 1.044889e+01 1.038232e+01 1.039967e+01 1.005835e+01 9.997745e+00
#>  [121] 1.066371e+01 1.120207e+01 1.165474e+01 1.203518e+01 1.266397e+01
#>  [126] 1.316309e+01 1.338841e+01 1.338772e+01 1.291060e+01 1.305969e+01
#>  [131] 1.291121e+01 1.290385e+01 1.321918e+01 1.270699e+01 1.267078e+01
#>  [136] 1.266081e+01 1.210527e+01 1.288388e+01 1.295390e+01 1.301435e+01
#>  [141] 1.304217e+01 1.296256e+01 1.282237e+01 1.270072e+01 1.284806e+01
#>  [146] 1.331972e+01 1.272314e+01 1.285782e+01 1.296091e+01 1.315990e+01
#>  [151] 1.299226e+01 1.279162e+01 1.387474e+01 1.457389e+01 1.456580e+01
#>  [156] 1.456580e+01 1.488436e+01 1.460822e+01 1.439690e+01 1.498209e+01
#>  [161] 1.487171e+01 1.503191e+01 1.460960e+01 1.456138e+01 1.495378e+01
#>  [166] 1.511563e+01 1.509902e+01 1.506401e+01 1.535930e+01 1.530112e+01
#>  [171] 1.562198e+01 1.543097e+01 1.510050e+01 1.456828e+01 1.410154e+01
#>  [176] 1.363675e+01 1.353876e+01 1.197782e+01 1.142042e+01 1.180731e+01
#>  [181] 1.169308e+01 1.165315e+01 1.182068e+01 1.091439e+01 1.052757e+01
#>  [186] 1.040623e+01 1.035796e+01 1.086841e+01 1.041891e+01 1.030477e+01
#>  [191] 1.079797e+01 1.041472e+01 1.040687e+01 1.044855e+01 1.071410e+01
#>  [196] 1.022515e+01 1.022057e+01 1.074860e+01 1.070007e+01 1.001174e+01
#>  [201] 1.150377e+01 1.156248e+01 1.155488e+01 1.128203e+01 1.085667e+01
#>  [206] 1.088974e+01 1.238486e+01 1.232363e+01 1.245589e+01 1.261713e+01
#>  [211] 1.197207e+01 1.226367e+01 1.146431e+01 1.146431e+01 1.087091e+01
#>  [216] 1.098517e+01 1.118573e+01 1.135760e+01 1.189065e+01 1.164159e+01
#>  [221] 1.214926e+01 1.197377e+01 1.183890e+01 1.196358e+01 1.256029e+01
#>  [226] 1.308265e+01 1.310503e+01 1.302151e+01 1.469930e+01 1.538001e+01
#>  [231] 1.552051e+01 1.610862e+01 1.631643e+01 1.555228e+01 1.659874e+01
#>  [236] 1.644067e+01 1.706924e+01 1.641932e+01 1.677530e+01 1.698254e+01
#>  [241] 1.753948e+01 1.708029e+01 1.812772e+01 1.817342e+01 1.817663e+01
#>  [246] 1.808738e+01 2.020001e+01 1.918340e+01 1.640923e+01 1.638339e+01
#>  [251] 1.754539e+01 1.779635e+01 1.795559e+01 1.791413e+01 1.387545e+01
#>  [256] 1.404173e+01 1.393924e+01 1.091067e+01 1.008111e+01 1.041979e+01
#>  [261] 1.053829e+01 1.040386e+01 1.000931e+01 9.823722e+00 1.135727e+01
#>  [266] 1.310124e+01 1.192704e+01 1.123435e+01 1.095682e+01 7.479820e+00
#>  [271] 7.263160e+00 6.217212e+00 6.624945e+00 6.980246e+00 6.781584e+00
#>  [276] 6.048499e+00 5.540599e+00 7.090132e+00 6.033846e+00 5.988906e+00
#>  [281] 6.156661e+00 5.841055e+00 5.232358e+00 4.780816e+00 5.842031e+00
#>  [286] 5.819441e+00 6.204140e+00 7.279058e+00 7.623736e+00 7.455062e+00
#>  [291] 7.369186e+00 6.002120e+00 5.828278e+00 5.679322e+00 5.905394e+00
#>  [296] 6.153719e+00 6.001546e+00 4.258318e+00 5.620625e+00 8.231576e+00
#>  [301] 8.551064e+00 8.173763e+00 4.363588e+00 4.785064e+00 5.011090e+00
#>  [306] 1.233088e+01 1.093207e+01 1.116102e+01 1.367717e+01 1.619605e+01
#>  [311] 1.520776e+01 1.535508e+01 1.555553e+01 1.570950e+01 1.842607e+01
#>  [316] 1.554088e+01 1.192475e+01 1.437725e+01 1.514423e+01 1.511182e+01
#>  [321] 2.417627e+01 2.505782e+01 2.610815e+01 2.548231e+01 2.500739e+01
#>  [326] 2.509188e+01 2.514154e+01 2.519705e+01 2.306891e+01 2.551494e+01
#>  [331] 2.742633e+01 2.726605e+01 2.682351e+01 2.623808e+01 2.633512e+01
#>  [336] 2.394441e+01 2.454975e+01 2.461304e+01 2.335081e+01 2.536196e+01
#>  [341] 2.575247e+01 2.798735e+01 2.779929e+01 2.841472e+01 2.730662e+01
#>  [346] 2.708765e+01 2.667788e+01 2.535754e+01 2.517798e+01 2.667133e+01
#>  [351] 2.626774e+01 2.630379e+01 2.659778e+01 2.569173e+01 2.556263e+01
#>  [356] 2.534197e+01 2.553321e+01 2.687840e+01 2.739066e+01 2.750914e+01
#>  [361] 2.767783e+01 2.789952e+01 2.778859e+01 2.772366e+01 2.745970e+01
#>  [366] 2.790313e+01 2.780835e+01 2.938502e+01 2.906019e+01 2.884989e+01
#>  [371] 2.871198e+01 2.773374e+01 2.781379e+01 2.699176e+01 2.643128e+01
#>  [376] 2.526742e+01 2.511984e+01 2.494229e+01 2.466107e+01 2.455512e+01
#>  [381] 2.177039e+01 1.897839e+01 1.867336e+01 1.877668e+01 2.018608e+01
#>  [386] 2.013279e+01 2.121176e+01 2.064230e+01 2.025687e+01 1.995094e+01
#>  [391] 1.937118e+01 1.800084e+01 1.638546e+01 1.644370e+01 1.656635e+01
#>  [396] 1.648883e+01 1.684728e+01 1.760401e+01 1.726500e+01 1.712971e+01
#>  [401] 1.474797e+01 1.473575e+01 1.459749e+01 1.420367e+01 1.458009e+01
#>  [406] 1.388021e+01 1.334536e+01 1.309750e+01 1.186605e+01 1.167404e+01
#>  [411] 1.182006e+01 1.182915e+01 1.162849e+01 1.091056e+01 1.104973e+01
#>  [416] 1.104489e+01 1.072397e+01 1.075883e+01 1.075919e+01 1.084934e+01
#>  [421] 1.092478e+01 1.092970e+01 9.988584e+00 1.082323e+01 1.078290e+01
#>  [426] 1.084189e+01 1.298993e+01 1.301972e+01 1.325614e+01 1.278901e+01
#>  [431] 1.352277e+01 1.441930e+01 1.552971e+01 1.524813e+01 1.529688e+01
#>  [436] 1.397210e+01 1.357412e+01 1.382488e+01 1.382488e+01 1.379192e+01
#>  [441] 1.218754e+01 1.208780e+01 1.245716e+01 1.235135e+01 1.304305e+01
#>  [446] 1.242447e+01 1.249762e+01 1.206602e+01 1.138332e+01 1.151657e+01
#>  [451] 1.205539e+01 1.300495e+01 1.317616e+01 1.295743e+01 1.604230e+01
#>  [456] 1.608395e+01 1.667768e+01 1.732321e+01 1.695588e+01 1.659120e+01
#>  [461] 1.676734e+01 1.676057e+01 1.774879e+01 1.783541e+01 1.915984e+01
#>  [466] 1.771417e+01 1.808369e+01 1.827115e+01 1.826162e+01 1.659205e+01
#>  [471] 1.697528e+01 1.736323e+01 1.759619e+01 1.845253e+01 1.738780e+01
#>  [476] 1.894928e+01 1.912263e+01 1.688999e+01 1.698885e+01 1.675051e+01
#>  [481] 1.672105e+01 1.630327e+01 1.713171e+01 1.649391e+01 1.831117e+01
#>  [486] 1.875079e+01 1.885002e+01 1.977269e+01 1.923706e+01 1.912618e+01
#>  [491] 1.959719e+01 2.142427e+01 2.135134e+01 2.008426e+01 2.023784e+01
#>  [496] 1.738792e+01 1.790471e+01 1.765915e+01 1.782022e+01 1.724279e+01
#>  [501] 1.746524e+01 1.683450e+01 1.682576e+01 1.660980e+01 1.673893e+01
#>  [506] 1.348472e+01 1.090559e+01 1.082116e+01 1.068811e+01 1.056493e+01
#>  [511] 1.076728e+01 1.055589e+01 1.092666e+01 9.600779e+00 9.730432e+00
#>  [516] 1.019621e+01 1.121809e+01 1.095115e+01 1.062483e+01 1.059278e+01
#>  [521] 1.161679e+01 1.162588e+01 1.148754e+01 1.139075e+01 1.117127e+01
#>  [526] 1.115298e+01 9.419509e+00 9.461653e+00 9.881183e+00 9.366357e+00
#>  [531] 1.074744e+01 1.068608e+01 1.157058e+01 1.171594e+01 1.065748e+01
#>  [536] 9.274085e+00 8.805829e+00 8.874262e+00 8.825091e+00 8.997421e+00
#>  [541] 8.938833e+00 7.790278e+00 7.754862e+00 7.648779e+00 8.408601e+00
#>  [546] 8.349217e+00 8.457275e+00 8.370379e+00 9.782252e+00 9.673542e+00
#>  [551] 1.004982e+01 1.002785e+01 1.001614e+01 1.023060e+01 1.035592e+01
#>  [556] 1.014438e+01 1.092620e+01 1.359612e+01 1.369688e+01 1.393879e+01
#>  [561] 1.456103e+01 1.443067e+01 1.540580e+01 1.473076e+01 1.514673e+01
#>  [566] 1.494364e+01 1.400508e+01 1.379176e+01 1.219377e+01 1.233701e+01
#>  [571] 1.299223e+01 1.384387e+01 1.308565e+01 1.306299e+01 1.313875e+01
#>  [576] 1.462879e+01 1.463816e+01 1.461762e+01 1.471700e+01 1.374020e+01
#>  [581] 1.376564e+01 1.258281e+01 1.257279e+01 1.149717e+01 1.140215e+01
#>  [586] 1.302552e+01 1.334153e+01 1.354210e+01 1.265322e+01 1.236675e+01
#>  [591] 1.254674e+01 1.253132e+01 1.504019e+01 1.466594e+01 1.502228e+01
#>  [596] 1.513620e+01 1.489092e+01 1.568851e+01 1.561039e+01 1.316457e+01
#>  [601] 1.332480e+01 1.286199e+01 1.204185e+01 1.186972e+01 1.150414e+01
#>  [606] 1.080911e+01 1.030887e+01 9.178199e+00 9.215876e+00 9.080928e+00
#>  [611] 8.909113e+00 9.130941e+00 9.144745e+00 7.683591e+00 8.013691e+00
#>  [616] 7.475540e+00 7.469875e+00 7.411402e+00 6.846689e+00 7.864445e+00
#>  [621] 7.781765e+00 7.073437e+00 6.428211e+00 6.327497e+00 6.768449e+00
#>  [626] 6.377884e+00 3.414160e+00 3.382487e+00 4.865662e+00 5.042201e+00
#>  [631] 5.637077e+00 5.698183e+00 5.817778e+00 6.129906e+00 7.205861e+00
#>  [636] 7.626983e+00 7.314983e+00 7.276684e+00 7.234727e+00 7.770171e+00
#>  [641] 8.101184e+00 7.562180e+00 7.575929e+00 6.763910e+00 6.777606e+00
#>  [646] 6.411826e+00 5.182203e+00 4.808350e+00 4.748589e+00 4.764327e+00
#>  [651] 5.491879e+00 5.087452e+00 5.591747e+00 6.702282e+00 5.417773e+00
#>  [656] 3.424805e+00 4.999426e+00 5.652284e+00 5.963221e+00 5.953882e+00
#>  [661] 6.229486e+00 5.444007e+00 5.174882e+00 4.107902e+00 4.832468e+00
#>  [666] 4.292612e+00 4.536999e+00 4.647877e+00 4.768637e+00 5.810110e+00
#>  [671] 5.937340e+00 6.216347e+00 4.763681e+00 4.363889e+00 4.433653e+00
#>  [676] 3.751095e+00 3.706665e+00 4.035833e+00 4.107000e+00 2.783804e+00
#>  [681] 2.749966e+00 2.916800e+00 3.306647e+00 3.048532e+00 2.916904e+00
#>  [686] 1.431524e+00 1.225672e+00 1.871284e+00 9.241198e-01 7.737926e-01
#>  [691] 9.802925e-01 6.714390e-01 2.311542e-01 7.061669e-01 4.629574e-01
#>  [696] 5.027500e-01 8.819544e-02 5.826701e-01 6.170688e-01 1.124773e+00
#>  [701] 1.098368e+00 1.042847e+00 5.386757e-01 1.797037e+02 1.788387e+02
#>  [706] 1.796287e+02 1.026974e+00 8.795439e-02 1.798562e+02 3.904979e-02
#>  [711] 6.280985e-01 5.048608e-01 9.233595e-01 9.496267e-01 1.505929e+00
#>  [716] 1.374655e+00 1.667046e+00 1.711375e+00 1.751840e+00 1.715371e+00
#>  [721] 1.500778e+00 1.577838e+00 1.522879e+00 2.145265e+00 8.308535e-01
#>  [726] 2.108894e-01 2.486243e-01 4.830500e-01 1.148648e-01 5.189493e-01
#>  [731] 4.195617e-01 3.050743e-01 2.092210e-01 1.799013e+02 1.798816e+02
#>  [736] 1.796452e+02 1.793997e+02 1.788302e+02 1.787895e+02 1.790237e+02
#>  [741] 1.798099e+02 1.240426e-03 7.264005e-01 1.014093e+00 6.561114e-01
#>  [746] 5.498937e-01 6.773649e-01 7.502621e-01 7.405619e-01 8.099665e-01
#>  [751] 4.330812e-01 5.808066e-01 1.794003e+02 6.981542e-01 7.461776e-01
#>  [756] 1.509192e+00 1.725548e+00 1.237290e+00 1.474792e+00 1.756545e+00
#>  [761] 1.741290e+00 1.421891e+00 1.421891e+00 1.737877e+00 1.539881e+00
#>  [766] 2.292026e+00 2.375179e+00 2.319394e+00 2.248823e+00 2.203371e+00
#>  [771] 2.343951e+00 2.591935e+00 2.620891e+00 2.624906e+00 2.792902e+00
#>  [776] 4.445214e+00 4.841163e+00 4.256415e+00 3.726351e+00 4.497594e+00
#>  [781] 4.810848e+00 4.865264e+00 4.988340e+00 5.314843e+00 5.867952e+00
#>  [786] 5.617265e+00 5.345209e+00 5.358555e+00 6.939653e+00 6.943852e+00
#>  [791] 7.201039e+00 6.831910e+00 6.868331e+00 6.363125e+00 6.645628e+00
#>  [796] 6.776590e+00 6.812577e+00 6.955041e+00 6.648690e+00 5.699973e+00
#>  [801] 5.575963e+00 5.228286e+00 5.313893e+00 6.654591e+00 5.810105e+00
#>  [806] 6.448007e+00 6.181917e+00 6.285345e+00 6.273628e+00 6.500180e+00
#>  [811] 6.656357e+00 6.838133e+00 6.591495e+00 6.554501e+00 6.013403e+00
#>  [816] 5.907945e+00 5.341938e+00 5.230735e+00 5.285473e+00 5.265491e+00
#>  [821] 5.125526e+00 4.977549e+00 5.006897e+00 4.837833e+00 4.806472e+00
#>  [826] 4.674524e+00 3.956121e+00 3.827366e+00 4.062318e+00 5.201649e+00
#>  [831] 4.221949e+00 2.693400e+00 2.117927e+00 2.387436e+00 2.243379e+00
#>  [836] 1.680031e+00 2.263555e+00 2.547437e+00 2.011555e+00 4.583436e-01
#>  [841] 2.830896e-01 1.798115e+02 1.796558e+02 1.796779e+02 1.796776e+02
#>  [846] 1.793248e+02 1.793493e+02 1.794409e+02 1.796438e+02 1.779002e+02
#>  [851] 1.789037e+02 1.788854e+02 1.792837e+02 1.789293e+02 1.788514e+02
#>  [856] 1.786779e+02 1.785417e+02 1.784597e+02 1.783276e+02 1.783404e+02
#>  [861] 1.788950e+02 1.786688e+02 1.791074e+02 1.795869e+02 1.799744e+02
#>  [866] 5.932819e-01 8.142827e-01 1.182631e+00 1.142983e+00 7.020284e-01
#>  [871] 1.112058e+00 1.114752e+00 1.136306e+00 9.736434e-01 1.035730e+00
#>  [876] 1.036147e+00 1.222258e+00 1.928232e+00 1.993980e+00 2.375820e+00
#>  [881] 1.918504e+00 2.138304e+00 2.457452e+00 2.705794e+00 2.455862e+00
#>  [886] 2.470185e+00 2.245108e+00 2.467857e+00 2.706127e+00 3.486133e+00
#>  [891] 3.509718e+00 3.521121e+00 4.168441e+00 4.231933e+00 4.102403e+00
#>  [896] 4.679260e+00 4.743800e+00 4.725551e+00 4.561994e+00 4.283968e+00
#>  [901] 7.056977e+00 6.998673e+00 7.150757e+00 7.487719e+00 7.743352e+00
#>  [906] 7.646370e+00 7.946741e+00 7.849936e+00 7.923419e+00 7.967702e+00
#>  [911] 7.598443e+00 6.945283e+00 7.111511e+00 7.348870e+00 6.953052e+00
#>  [916] 6.778698e+00 6.514898e+00 6.673636e+00 6.556584e+00 7.671907e+00
#>  [921] 7.775740e+00 7.451732e+00 7.926613e+00 7.886241e+00 8.068007e+00
#>  [926] 8.167266e+00 8.596204e+00 8.472689e+00 7.850879e+00 7.983638e+00
#>  [931] 8.451233e+00 9.541136e+00 9.182259e+00 9.292638e+00 1.018876e+01
#>  [936] 1.025564e+01 1.062475e+01 1.168155e+01 1.190215e+01 1.233975e+01
#>  [941] 1.218232e+01 1.259071e+01 1.273104e+01 1.272826e+01 1.335192e+01
#>  [946] 1.402409e+01 1.259967e+01 1.369030e+01 1.334952e+01 1.369553e+01
#>  [951] 1.447513e+01 1.417808e+01 1.419717e+01 1.364550e+01 1.344800e+01
#>  [956] 1.345863e+01 1.374747e+01 1.391619e+01 1.368038e+01 1.357988e+01
#>  [961] 1.377287e+01 1.447251e+01 1.446824e+01 1.580375e+01 1.543603e+01
#>  [966] 1.570522e+01 1.610266e+01 1.609892e+01 1.635814e+01 1.653050e+01
#>  [971] 1.557268e+01 1.586664e+01 1.605245e+01 1.608381e+01 1.610675e+01
#>  [976] 1.613440e+01 1.613281e+01 1.576083e+01 1.535443e+01 1.585087e+01
#>  [981] 1.596046e+01 1.527488e+01 1.467493e+01 1.554740e+01 1.538206e+01
#>  [986] 1.537146e+01 1.510605e+01 1.488399e+01 1.397109e+01 1.182643e+01
#>  [991] 1.103888e+01 1.078587e+01 1.031130e+01 1.037558e+01 1.057237e+01
#>  [996] 1.040628e+01 1.051652e+01 1.108723e+01 1.019780e+01 1.093881e+01
#> [1001] 1.184996e+01 1.116305e+01 1.076380e+01 1.027064e+01 1.098828e+01
#> [1006] 1.085765e+01 1.103940e+01 1.023432e+01 1.022928e+01 1.215685e+01
#> [1011] 1.192809e+01 1.182669e+01 1.229047e+01 1.242644e+01 9.531662e+00
#> [1016] 9.771542e+00 9.928226e+00 1.076947e+01 1.083338e+01 1.024602e+01
#> [1021] 1.020822e+01 1.028727e+01 1.030058e+01 1.010566e+01 1.048839e+01
#> [1026] 1.041714e+01 9.559291e+00 9.577248e+00 9.887646e+00 1.034046e+01
#> [1031] 1.043899e+01 1.025333e+01 1.042792e+01 1.046337e+01 9.688415e+00
#> [1036] 9.491946e+00 9.294928e+00 1.012643e+01 1.124710e+01 1.168827e+01
#> [1041] 1.450945e+01 1.489297e+01 1.501958e+01 1.486392e+01 1.474585e+01
#> [1046] 1.479544e+01 1.536120e+01 1.574117e+01 1.643632e+01 1.713792e+01
#> [1051] 1.662759e+01 1.579071e+01 1.525674e+01 1.502628e+01 1.651904e+01
#> [1056] 1.652267e+01 1.730900e+01 1.729331e+01 1.866814e+01 1.902420e+01
#> [1061] 1.711500e+01 1.759087e+01 1.734444e+01 1.659481e+01 1.710217e+01
#> [1066] 1.718665e+01 1.702050e+01 1.686682e+01 1.611886e+01 1.653948e+01
#> [1071] 1.726611e+01 1.700947e+01 1.676143e+01 1.556701e+01 1.580898e+01
#> [1076] 1.666574e+01 1.705653e+01 1.827755e+01 1.928318e+01 1.913154e+01
#> [1081] 1.903307e+01 1.885907e+01 1.897217e+01 1.832425e+01 1.805049e+01
#> [1086] 1.848665e+01 1.851450e+01 1.887459e+01 1.710561e+01 1.508844e+01
#> [1091] 1.424360e+01 1.298450e+01 1.267487e+01 1.260627e+01 1.322830e+01
#> [1096] 1.302658e+01 1.306316e+01 1.220803e+01 1.187895e+01 1.230182e+01
#> [1101] 1.108212e+01 1.159684e+01 1.187704e+01 1.233265e+01 1.213593e+01
#> [1106] 1.156568e+01 1.159522e+01 9.126036e+00 8.813854e+00 7.979494e+00
#> [1111] 7.330222e+00 7.555373e+00 8.096769e+00 7.910124e+00 7.722796e+00
#> [1116] 5.862083e+00 5.807563e+00 7.000009e+00 7.350891e+00 7.466724e+00
#> [1121] 6.225417e+00 6.811142e+00 8.103863e+00 9.223694e+00 1.223148e+01
#> [1126] 1.258952e+01