Apply Rolling Functions using circular statistics

A generic function for applying a function to rolling margins of an array.

Usage

roll_circstats(
  azi,
  w = NULL,
  FUN,
  width,
  by.column = FALSE,
  partial = TRUE,
  align = "center",
  fill = NA,
  ...
)

Arguments

azi

numeric. Azimuth of \(\sigma_{Hmax}\)

...

Arguments passed on to zoo::rollapply

unc

numeric. Uncertainty of observed \(\sigma_{Hmax}\), either a numeric vector or a number

Value

numeric vector with the results of the rolling function.

Note

If the rolling statistics are applied to values that are a function of distance it is recommended to sort the values first.

Examples

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")
data("san_andreas")
ep <- subset(nuvel1, nuvel1$plate.rot == "na")
distance <- distance_from_pb(
  x = san_andreas,
  euler = ep,
  pb = plate_boundary,
  tangential = TRUE
)
dat <- san_andreas[order(distance), ]
roll_circstats(dat$azi, w = 1 / dat$unc, circular_mean, width = 51)
#>    [1]   5.4288059   6.0541673   4.0694800   7.4780988   6.2569433   7.8090857
#>    [7]   8.3776621   8.1387772   8.4996015   8.4930619   7.9875433   7.8157945
#>   [13]   6.3244567   6.1818651   6.3387962   6.7073876   6.5517873   6.4187081
#>   [19]   6.6644471   5.7420887   5.6229030   5.4803960   5.5418065   5.7950382
#>   [25]   6.5106022   6.4118259   6.8329355   6.7925986   6.3493853   8.1706689
#>   [31]   7.6123409   7.6185307   7.6185307   7.6185307   7.7469457   7.6298206
#>   [37]   7.6251751   8.2477597   6.9620332   5.9404250   6.0130334   6.7407262
#>   [43]   6.0626755   6.0466599   5.2666696   4.9767460   5.0002380   5.1461846
#>   [49]   5.7158755   5.7701181   5.1720028   5.0004596   4.9835686   4.6756093
#>   [55]   3.6014417   3.8473630   2.8939587   2.5407681   2.5681541   2.4535043
#>   [61]   1.7556750   1.9527298   2.1167563   2.7314800   2.7600040   5.6552259
#>   [67]   5.3653210   5.5065567   5.9882447   5.7514147   6.3406101   7.3088167
#>   [73]   8.0959306   8.0594955   8.0118460   6.3289101   6.5365455   5.2155494
#>   [79]   5.1595920   5.5228737   4.5018189   4.5145903   4.5538571   4.5942983
#>   [85]   4.5192458   4.6386052   5.2204386   4.8396270   4.8516981   5.9734817
#>   [91]   6.7834631   7.9981681   7.1524212   7.7175895   7.9280707   7.7492529
#>   [97]   7.9006335   7.9376237   8.2066210   7.9247735   8.8433146   9.4587618
#>  [103]   9.5241377   9.3391042  10.7408647  10.9282258   9.6891846   9.7786434
#>  [109]   9.5697750   9.7692477   9.8014810  11.3940899  11.3609013  11.4722858
#>  [115]  11.9648171  10.7466261   8.4442635   8.5144648   8.3816615   7.9408622
#>  [121]   9.9797381  10.0777212   8.4951221   7.4994598   7.4701409   7.5870809
#>  [127]   8.8993296   8.8864557  10.0502330  10.6267975   8.9741170   9.2125220
#>  [133]   9.1958433   9.2155789   9.1785764   8.3650786   7.9766177   6.8720878
#>  [139]   7.1406604   7.1192591   5.6787906   5.5099683   4.2978451   4.3276434
#>  [145]   3.2975013   3.7463259   3.7890900   5.2553676   5.2081948   6.1201382
#>  [151]   6.0778436   5.2892542   4.7954310   4.6452866   5.0058099   5.2574094
#>  [157]   5.3824955   6.0184014   5.9167485   5.9625820   5.8398746   5.7156413
#>  [163]   5.1039279   5.1362959   5.5806006   5.2840182   5.9536047   5.8176774
#>  [169]   5.7283505   5.7598858   5.7257328   4.1168736   4.2763869   4.6136876
#>  [175]   4.8729765   5.6059419   5.6316799   5.7798370   5.5210677   6.3229913
#>  [181]   6.2536363   7.3087762   6.6510666   6.6766990   6.7019240   6.6786464
#>  [187]   5.9000039   5.7547141   6.4757693   6.2955268   6.3411717   7.6830225
#>  [193]   7.5993172   5.9082308   5.9472221   6.2693503   7.1085330   7.0112071
#>  [199]   5.2236795   5.2425473   3.6818137   3.7312944   4.5356199   5.1698579
#>  [205]   5.5734481   4.8953166   4.7745971   3.6284751   3.8052532   3.8817283
#>  [211]   3.9820460   3.4942372   3.4331871   3.4602760   3.4324922   2.7552160
#>  [217]   2.6981051   3.1385693   3.1903792   2.6008006   2.7078617   2.7104231
#>  [223]   2.7351798   2.7103733   2.7256668   2.9006309   1.9380473   2.5486927
#>  [229]   3.6193064   3.7032401   3.0790054   3.1356714   3.1915970   4.0990708
#>  [235]   4.2080537   4.1377389   4.6486546   5.6995598   5.9088962   7.0239874
#>  [241]   7.1252919   7.1382199   7.2233608   7.3089585   9.2746234   9.1980016
#>  [247]   9.0559626   8.3790538   8.4472265   8.6822067   8.6822067   8.8233308
#>  [253]   8.9023604   8.6403763   8.3613591   8.4240907   7.6271969   7.5699053
#>  [259]   8.1552531   9.1675981   7.0642461   8.1107324   9.1003869   9.6023023
#>  [265]   9.4746847   9.2474262   8.8336364   9.0003973   8.9529534   9.0021906
#>  [271]  10.1270914  10.2839336  13.6013653  13.2064968  13.0270211  12.3500329
#>  [277]  11.9457684  13.0155190  12.5199432  11.3282964  11.5522395  12.1459755
#>  [283]  11.5975848  12.1742215  11.7929415  11.7126955  11.1105588  10.6875775
#>  [289]  10.8801404  10.6459235  10.1665413  10.1290354  10.4230411  10.3228850
#>  [295]  10.3105491   9.7992399  10.0906507   9.8156208   9.9757533   9.9303224
#>  [301]  10.0268009  10.1647056  10.2444498  10.1516693  10.0699121  10.8912494
#>  [307]  10.8389530  11.5296288  11.2014172  11.0310403  10.0711490  12.2728134
#>  [313]  10.8156238   9.9455009   8.6378302   9.4293363  10.1661998  10.4372622
#>  [319]  10.3979173  10.6019788  10.2763824  10.9055447  10.7779905   8.7707892
#>  [325]   8.8465951  10.4225768  11.0109993  11.0766358  10.8197600  10.8408472
#>  [331]  11.1711499  11.7821931  11.0086472  11.1791996  10.7410871  10.7655905
#>  [337]  10.2409341  10.6918385  10.5841702  10.6262109  10.6356352   9.9943178
#>  [343]  10.0051458   9.7397164   9.2200065   9.2560864   9.5005880   9.5583962
#>  [349]  10.6770939  10.7970078  10.7452316  10.3527386  10.0612430  11.1591740
#>  [355]  11.0713307  11.0769549  10.7781804  10.6852169  10.9165408  11.1483982
#>  [361]  11.3811567  11.6057747  11.8904022  13.0089443  13.2515044  15.6402523
#>  [367]  14.4605005  14.1750122  13.6210521  11.6713519  11.1265730  11.5684146
#>  [373]   9.8985779  11.0303523  11.2222496  10.1428340   7.8986222   7.9365820
#>  [379]   7.5195782   8.2670146   8.3011518   7.6541463   6.3335885   6.5972148
#>  [385]   6.8984516   7.6749465   7.2767639   8.1944766   8.6161434   8.7212119
#>  [391]   8.7097277   8.6905403   8.5709163   8.7699008   8.4412743   9.1678839
#>  [397]   9.1981170   8.7458913   8.4221051   7.5699306   6.6429146   6.6332172
#>  [403]   6.8730217   8.2231136   6.6472352   6.8697570   6.9037845   6.8318942
#>  [409]   6.7741959   6.4449307   6.4908264   6.3458555   6.2958239   6.0063167
#>  [415]   5.9677503   5.9769006   4.9161896   4.8049983   4.0223998   4.4766699
#>  [421]   5.8481510   6.3794663   6.6691493   6.3986824   5.0554591   4.9008804
#>  [427]   5.5279381   5.1651474   5.2558134   5.6197040   4.9799344   5.0023675
#>  [433]   4.6583343   4.1573372   4.2345204   3.9358959   3.3979828   2.6530760
#>  [439]   2.4338043   1.7772086   1.9555434   1.4832523   2.0369848   2.0607384
#>  [445]   1.8051343   1.6639705   1.5521987   1.5586675   2.1132809   2.1351850
#>  [451]   0.1244773   0.7815596   0.6361085   0.6856969 179.6040999 179.0116869
#>  [457] 178.9172096 178.9211898 178.8305944 179.3172062 179.2920634 179.1108487
#>  [463]   0.2359063   0.2392914   0.1742757   0.5628241   0.7423999   1.9146660
#>  [469]   2.3466614   2.7770594   2.6638556   2.5015114   2.1150774   2.2942789
#>  [475]   2.7108878   3.1444194   2.9079620   3.1846217   3.6084984   3.1221751
#>  [481]   2.9094177   3.4064666   4.3033446   4.5105965   5.3702090   5.3417673
#>  [487]   5.2974602   5.4066848   6.5348308   6.4227139   6.8289091   7.3107913
#>  [493]   7.5018891   6.8084895   6.8765988   6.4081823   6.7826600   6.6237435
#>  [499]   6.1151932   5.7418174   5.7656353   8.2999153   8.3695410   8.9327512
#>  [505]   9.3652695   9.4557049  10.1802016  10.6025430  10.4673990  11.0369702
#>  [511]  10.7164075  11.6703126  11.8429790  11.1602847  10.7634713  10.2570271
#>  [517]  10.4933553  10.4014456  10.6531163  10.3601935  10.1356708  10.0642177
#>  [523]   9.4773436   9.4704018  10.1450378  10.3095830  10.2930015  10.0688253
#>  [529]  10.0217801  10.1854776  10.4502157  11.5974176  11.3300453  10.6235472
#>  [535]  10.8374544  10.5120106  10.5270696  10.7078062  10.8012432  10.9122542
#>  [541]  10.5725179  10.4232652  10.0760178  10.3008314  10.7461089  11.1303755
#>  [547]  12.2270387  11.8670030  11.8043480  13.3853728  13.5827018  13.0943753
#>  [553]  13.4487551  13.5551430  13.1467625  12.8730745  13.8663911  13.4284864
#>  [559]  13.1447189  12.9756946  14.1316180  14.2455320  13.3758292  13.3854978
#>  [565]  13.3630933  14.0888451  14.9928914  14.5365698  14.8834465  14.5729469
#>  [571]  14.9470558  15.5121532  16.0289680  17.5130078  17.5646055  16.3179507
#>  [577]  17.7267863  17.8957880  18.1248680  18.1870421  18.4553276  17.5969898
#>  [583]  16.6908392  16.8908452  16.4761132  16.3679988  16.7502534  16.4585760
#>  [589]  16.3153859  16.4018609  16.2431216  17.3837189  17.4465337  17.2077723
#>  [595]  16.8715549  16.9041890  16.7753279  16.3267247  16.9400757  17.6385386
#>  [601]  17.8669852  17.8475655  19.1370743  19.7583872  20.2824863  20.9982331
#>  [607]  21.3586804  20.4141524  21.2562903  21.2574973  21.9267689  20.6503174
#>  [613]  20.7273516  21.3275476  21.2210909  21.5465961  21.9035564  22.0680800
#>  [619]  22.1960248  21.0554947  22.8635149  23.7255360  22.9714363  22.8982362
#>  [625]  22.2302646  21.9968542  23.2404822  21.7348377  21.6527882  21.3984355
#>  [631]  21.4080959  17.9777799  19.6227898  19.1560604  20.7513576  21.1714754
#>  [637]  19.5070018  17.8786376  17.6574968  18.8265360  20.2314979  20.3065984
#>  [643]  20.0265172  19.9369573  18.4768558  18.4606089  18.1573886  18.2271075
#>  [649]  18.4567449  18.2283241  18.5812869  18.1814769  18.0622605  17.9342155
#>  [655]  17.1442231  19.0186110  18.5711555  18.2379817  18.3463431  18.8574238
#>  [661]  18.5328849  17.9313440  17.6044904  17.4747258  17.2894964  17.2586381
#>  [667]  17.1340888  16.8878990  16.7189306  16.3814197  17.0323588  15.2035251
#>  [673]  14.6290615  14.3802287  14.0121791  13.9277973  14.7290718  14.4201034
#>  [679]  14.1995808  13.8089077  11.7904566  11.7111563  13.0089951  12.6396243
#>  [685]  10.7239069   9.6760969   9.4076183  10.1401287  10.4841835  10.9628752
#>  [691]   9.0583165   8.5156514   7.6552295   9.3999025   9.4993158   8.9298382
#>  [697]   9.0055504   9.1260603   8.9409480   8.8821938   9.4922174   8.7399815
#>  [703]   8.8678569   7.6486343   6.6729578   7.0372635   4.0011576   4.4150539
#>  [709]   5.0932139   5.2125279   4.3730063   4.5292063   5.3621689   2.3408195
#>  [715]   2.0576057   2.2265911   2.1749987   2.3149952   4.2986767   5.7066342
#>  [721]   6.0977607   6.1650410   6.2502656   6.3059176   5.2852002   5.8400407
#>  [727]   5.0306065   5.0993296   4.8928622   4.9432151   5.0533076   6.6933735
#>  [733]   6.8333517   6.4831407   7.0652121   8.4610227   8.8291117   9.1095657
#>  [739]   8.6924344   9.8665273   9.8338372  10.4072890  10.0802733   9.1619178
#>  [745]   7.5973116   7.7579964   8.8080329   9.0631145   8.8161939   8.3257193
#>  [751]   8.2410986   8.0342800   8.7136234   8.5984253   9.5192350  10.4963337
#>  [757]  10.1317375   9.5627801   9.6756310   8.9559711   7.9117629   8.8182145
#>  [763]  10.1998895  10.0261488  14.3403329  15.1931096  15.1882916  15.3551774
#>  [769]  15.5979441  14.0870695  12.1439341  12.0445020  13.0124589  14.0552126
#>  [775]  13.9904413  17.5207628  16.7605007  18.6522168  18.4184757  18.3562907
#>  [781]  18.6103502  18.6760621  18.5968908  19.0776640  19.3213424  19.2198463
#>  [787]  19.1497799  18.8962846  18.8111872  19.7862192  17.3719511  17.5534132
#>  [793]  17.3107307  15.8125186  16.9192152  17.0871862  17.0570487  17.3252904
#>  [799]  17.1047871  17.1836436  19.4088453  20.3337712  20.4852105  20.0157117
#>  [805]  20.3996358  20.2824563  20.0521399  20.8133054  19.8829692  20.1454786
#>  [811]  20.5699144  21.6162808  21.4668482  21.3421178  19.6843009  19.5414205
#>  [817]  19.9152131  19.9904230  19.7707529  19.4480994  19.1274715  18.2218182
#>  [823]  18.5985862  18.0064351  17.6158697  17.4033211  17.0700979  17.1368175
#>  [829]  15.9743732  15.7368893  15.7746571  15.4831902  15.5280829  15.3899918
#>  [835]  14.8204902  14.6327117  13.6899771  12.1414797  12.2562263  11.9155237
#>  [841]  11.3124157  11.6559067  11.3995866  11.4514054  12.2275404  11.8354123
#>  [847]  11.5675454  11.4219499  11.7803967  11.4398391  11.4887601   9.9051484
#>  [853]   9.8275504   9.6652698   9.8758338   9.7676372   9.8362443   9.8234843
#>  [859]   8.2827099   9.0224974   9.6994082   8.9219293   8.9217800   8.3929832
#>  [865]   8.4056230   9.1402597   9.6813966   9.9881823  10.0363359   9.9942980
#>  [871]   9.9348941  10.1672884  10.5815468   9.8632114  10.3584316   9.4170293
#>  [877]   9.9247670   9.8678902   9.6276696  10.2359331   9.8581813   9.9460079
#>  [883]  10.0713289  10.2646355  11.2595094  10.7595062  10.8939754  11.4147682
#>  [889]  13.4224339  13.4101633  13.2694274  13.7859847  14.2001106  13.4081654
#>  [895]  13.5935730  13.8974868  14.3912414  14.6186547  14.9751439  14.9974466
#>  [901]  15.6615809  15.7523126  15.4082075  15.4467971  15.4509575  15.2905797
#>  [907]  15.1611412  14.8690645  15.0765703  15.2585332  15.3069890  14.6601439
#>  [913]  14.9538964  15.1767321  15.1547254  14.9489620  14.8665665  14.3557670
#>  [919]  13.7555753  13.2209744  13.1326175  13.4543897  12.9355675  12.8016138
#>  [925]  12.8179094  12.3779993  12.5258411  12.1610262  12.4375635  12.5504598
#>  [931]  12.6710412  12.6927185  12.6572297  12.6000269  12.4091766  11.4444544
#>  [937]  12.0152740  12.0926398  12.0805430  11.4250968  11.0686968  12.4684223
#>  [943]  12.2082353  12.1375198  12.6510836  12.3268906  12.0667396  11.2009976
#>  [949]  11.1547192  10.6290637   9.8228528   8.9428388   8.9541951   9.0493687
#>  [955]   8.9386491   8.9729195   9.2404207   9.3667548   9.6161203   9.3061724
#>  [961]   8.4182111   8.7572328   8.5191905   8.4030206   8.4974378   8.7317989
#>  [967]   9.0300538   8.6608610   8.6814942   8.9633207  10.2825855  10.7701408
#>  [973]  10.5123169  11.4908970  12.0707211  12.0136080  12.3165339  14.6686347
#>  [979]  14.6972314  15.0963154  15.5043069  15.9275125  15.9368259  15.7137706
#>  [985]  16.5364900  17.5676970  18.5550302  19.4400729  19.3652060  19.3056952
#>  [991]  19.8985792  19.7144050  14.0822848  15.4468841  16.5795706  17.6643571
#>  [997]  18.7967819  17.8785960  18.2909466  16.6633847  17.4413009  18.2378419
#> [1003]  18.4959059  19.1104741  19.8101476  21.2752704  21.8873501  21.2505270
#> [1009]  21.4549553  22.6722997  23.4233064  23.6687357  22.7006001  23.6006720
#> [1015]  24.2290415  24.7745122  26.6500262  27.1392051  28.2029924  28.2664141
#> [1021]  27.2062281  27.4250175  28.5158512  30.4155450  29.9960880  28.6820385
#> [1027]  28.7148040  29.0895262  29.9384663  29.7747962  29.0705134  29.0549158
#> [1033]  27.8659947  25.1281822  26.1469738  25.0903147  24.4802179  24.5509123
#> [1039]  23.3586737  23.3146645  23.4511895  23.0204760  24.8309840  25.3511564
#> [1045]  26.7598789  22.5780225  23.0779754  22.4107655  21.8634645  22.0001362
#> [1051]  22.5905769  22.8669671  23.0044098  23.1570532  23.0456341  21.6342640
#> [1057]  20.0288718  19.1573847  19.5177597  18.3555432  16.4357398  16.3441669
#> [1063]  17.3228054  17.4791796  16.1020446  13.7755488  12.1243111  11.3293219
#> [1069]   8.8358675   6.6908509   3.5223874   5.1370700   2.4067835 179.1427953
#> [1075]   3.0081629   1.5944526   6.1569906   5.9443793   4.1478723   1.1312291
#> [1081] 178.1511646 179.9880158