Apply Rolling Functions using Circular Statistical Tests for Uniformity

A generic function for applying a function to rolling margins of an array.

Usage

roll_normchisq(
  obs,
  prd,
  unc = NULL,
  width = NULL,
  by.column = FALSE,
  partial = TRUE,
  fill = NA,
  ...
)

roll_rayleigh(
  obs,
  prd,
  unc = NULL,
  width = NULL,
  by.column = FALSE,
  partial = TRUE,
  fill = NA,
  ...
)

roll_dispersion(
  x,
  mean,
  w = NULL,
  width = NULL,
  by.column = FALSE,
  partial = TRUE,
  fill = NA,
  ...
)

Arguments

obs

Numeric vector containing the observed azimuth of \(\sigma_{Hmax}\), same length as prd

prd

Numeric vector containing the modeled azimuths of \(\sigma_{Hmax}\), i.e. the return object from model_shmax()

unc

Uncertainty of observed \(\sigma_{Hmax}\), either a numeric vector or a number

width

integer specifying the window width (in numbers of observations) which is aligned to the original sample according to the align argument. If NULL, an optimal width is estimated.

by.column

logical. If TRUE, FUN is applied to each column separately.

partial

logical or numeric. If FALSE then FUN is only applied when all indexes of the rolling window are within the observed time range. If TRUE (default), then the subset of indexes that are in range are passed to FUN. A numeric argument to partial can be used to determine the minimal window size for partial computations. See below for more details.

fill

a three-component vector or list (recycled otherwise) providing filling values at the left/within/to the right of the data range. See the fill argument of zoo::na.fill() for details

...

Arguments passed on to zoo::rollapply

x

numeric. Directions in dgerees

mean

numeric. The angle about which the angles x disperse (in degrees).

w

(optional) Weights. A vector of positive numbers and of the same length as x.

Value

numeric vector with the test statistic of the rolling test.

Note

If the rolling functions are applied to values that are a function of distance it is recommended to sort the values first.

Examples

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")
data("san_andreas")
ep <- subset(nuvel1, nuvel1$plate.rot == "na")
distance <- distance_from_pb(
  x = san_andreas,
  euler = ep,
  pb = plate_boundary,
  tangential = TRUE
)
dat <- san_andreas[order(distance), ]
dat.PoR <- PoR_shmax(san_andreas, ep, "right")
roll_normchisq(dat.PoR$azi.PoR, 135, dat$unc)
#>    [1] 0.027352804 0.025944123 0.026121190 0.033374447 0.033329800 0.033259021
#>    [7] 0.033564252 0.030438497 0.030401819 0.030368882 0.030365725 0.030363120
#>   [13] 0.031241592 0.026031701 0.024639812 0.031773432 0.031807575 0.031860297
#>   [19] 0.031825576 0.035258105 0.035075606 0.035113984 0.051189346 0.051183169
#>   [25] 0.051211256 0.049543781 0.050723369 0.050499257 0.049576545 0.044203050
#>   [31] 0.044159857 0.044186279 0.045314208 0.040939737 0.038836523 0.035152401
#>   [37] 0.035121703 0.034964867 0.035551178 0.029768008 0.029748140 0.029741928
#>   [43] 0.029804586 0.029393656 0.032236490 0.034130651 0.027047463 0.026763775
#>   [49] 0.025124057 0.025407058 0.030381906 0.027905504 0.025620186 0.012839946
#>   [55] 0.012844795 0.012837598 0.013069911 0.013070494 0.013064419 0.013185516
#>   [61] 0.014108887 0.014158319 0.014149126 0.013152165 0.014108456 0.012999148
#>   [67] 0.014361340 0.014380822 0.014389168 0.014658484 0.018076011 0.018260938
#>   [73] 0.018403534 0.018309747 0.018305785 0.022584582 0.022584481 0.022567344
#>   [79] 0.023477440 0.024291940 0.023657734 0.019152686 0.035814558 0.038418900
#>   [85] 0.038443884 0.036034629 0.036068156 0.035278609 0.035320036 0.035326147
#>   [91] 0.035329913 0.035836355 0.035822010 0.035938330 0.038131233 0.043492546
#>   [97] 0.044592096 0.039428217 0.039416463 0.039410286 0.039402615 0.040326060
#>  [103] 0.040267352 0.040604510 0.040595350 0.040747880 0.055193762 0.055195747
#>  [109] 0.055016697 0.055332061 0.049935900 0.051148144 0.054183644 0.034929968
#>  [115] 0.035597323 0.035421708 0.051610904 0.051640751 0.057704213 0.057859702
#>  [121] 0.057907186 0.055400100 0.055388724 0.055355322 0.054884137 0.069963713
#>  [127] 0.058075134 0.056597074 0.071539871 0.071557091 0.067546712 0.067675920
#>  [133] 0.065731961 0.065568246 0.065594763 0.065599006 0.065317495 0.065116285
#>  [139] 0.062091623 0.062277441 0.035406517 0.036480424 0.036463097 0.036425639
#>  [145] 0.034892346 0.033747364 0.033813168 0.025484457 0.025470691 0.025530624
#>  [151] 0.030142660 0.030147170 0.030697548 0.030774348 0.029371964 0.020672789
#>  [157] 0.012249631 0.011395727 0.011394723 0.010988431 0.010985915 0.007610644
#>  [163] 0.007062514 0.006777092 0.006660216 0.006607633 0.006397569 0.006393600
#>  [169] 0.006399290 0.006440738 0.006387987 0.007920136 0.007801852 0.007813241
#>  [175] 0.008022796 0.008130646 0.008218892 0.008224528 0.008426805 0.008441209
#>  [181] 0.008410705 0.001743073 0.001742610 0.001743733 0.001692914 0.002286134
#>  [187] 0.003840548 0.003894527 0.004642839 0.004654811 0.005219364 0.005255484
#>  [193] 0.006714192 0.009072358 0.009463120 0.010486901 0.006123430 0.006941445
#>  [199] 0.006949340 0.007082498 0.007071795 0.007109051 0.007520098 0.010381898
#>  [205] 0.010392976 0.012040539 0.017343300 0.017389413 0.018975365 0.019037830
#>  [211] 0.019027131 0.019034555 0.024618536 0.025877504 0.026125347 0.026098207
#>  [217] 0.025617871 0.049753929 0.049743790 0.065566703 0.066093337 0.064670793
#>  [223] 0.064708498 0.065271662 0.065278780 0.065237784 0.065234118 0.086510964
#>  [229] 0.086388838 0.084029139 0.083402882 0.083352391 0.083309883 0.082118201
#>  [235] 0.087503673 0.087294586 0.080552029 0.082093979 0.080649630 0.085621772
#>  [241] 0.085571030 0.085526061 0.078308936 0.068955365 0.068513403 0.069938964
#>  [247] 0.067721789 0.068836783 0.051166986 0.049144969 0.018595733 0.018538643
#>  [253] 0.018421953 0.018455473 0.018021417 0.018002571 0.016285641 0.016290389
#>  [259] 0.017237926 0.017233236 0.010639632 0.010799773 0.010801583 0.011050890
#>  [265] 0.010933680 0.010919627 0.014225798 0.014331761 0.013565909 0.010294608
#>  [271] 0.019235838 0.018515318 0.018453966 0.020046129 0.026492219 0.029800667
#>  [277] 0.029005042 0.029967271 0.029973610 0.029973611 0.030805377 0.030867607
#>  [283] 0.059802869 0.063392004 0.061890470 0.060739252 0.060766877 0.064724625
#>  [289] 0.064667095 0.064757209 0.064729448 0.065360749 0.065337975 0.065342165
#>  [295] 0.065318920 0.066029007 0.066348864 0.066496405 0.066468854 0.064428803
#>  [301] 0.064874737 0.077437318 0.074572750 0.074533948 0.074639009 0.076284947
#>  [307] 0.074082171 0.069719011 0.067560800 0.067558796 0.067560465 0.069051825
#>  [313] 0.069033357 0.039076085 0.030628817 0.033289020 0.031203650 0.032037801
#>  [319] 0.032342829 0.032785918 0.068080600 0.068112834 0.057765470 0.056832983
#>  [325] 0.065966214 0.064426971 0.062400830 0.062414977 0.062406578 0.062580715
#>  [331] 0.060794083 0.061656722 0.061654744 0.064360956 0.064446902 0.061805281
#>  [337] 0.061816651 0.061811534 0.064684181 0.065967347 0.068303051 0.067438413
#>  [343] 0.067013977 0.067017636 0.066812618 0.065354358 0.066620834 0.066628177
#>  [349] 0.067555586 0.067649527 0.068332613 0.052605354 0.052617605 0.059995647
#>  [355] 0.059345235 0.036289139 0.040410673 0.046222068 0.046577643 0.044395256
#>  [361] 0.043858273 0.049616114 0.044630971 0.102746342 0.102616450 0.102583621
#>  [367] 0.118010188 0.118806138 0.118942387 0.118954095 0.118881169 0.120765800
#>  [373] 0.126600151 0.130503409 0.134158213 0.135560746 0.147497993 0.144693021
#>  [379] 0.132407937 0.114917849 0.114489943 0.114808291 0.108278715 0.119478224
#>  [385] 0.115084612 0.118693288 0.116055844 0.114727952 0.109707960 0.133030619
#>  [391] 0.146284799 0.146218146 0.142617243 0.145387550 0.099771506 0.100228022
#>  [397] 0.121994657 0.121956973 0.117469048 0.117476950 0.111051549 0.111125814
#>  [403] 0.111059256 0.111430423 0.105538278 0.099278683 0.099379616 0.098613020
#>  [409] 0.098851619 0.098895026 0.110308667 0.110381030 0.110158689 0.111103402
#>  [415] 0.174735758 0.180696974 0.186518332 0.191731610 0.192899133 0.195684887
#>  [421] 0.170381722 0.172968997 0.172944482 0.185080185 0.194813168 0.197551989
#>  [427] 0.173653816 0.139068647 0.152119114 0.157659441 0.156918419 0.170533352
#>  [433] 0.170712442 0.184066427 0.178566039 0.186338195 0.186832728 0.188525662
#>  [439] 0.188094945 0.187964192 0.199077556 0.193899159 0.195492706 0.180025529
#>  [445] 0.183244930 0.124496048 0.124502382 0.143265737 0.143335991 0.144665032
#>  [451] 0.148595537 0.148674095 0.154309378 0.154413879 0.154438480 0.154548930
#>  [457] 0.154810088 0.170600765 0.180624610 0.173996735 0.122426488 0.138731089
#>  [463] 0.135684107 0.136507212 0.126933927 0.124862781 0.125559048 0.125058039
#>  [469] 0.122695477 0.128908491 0.129056989 0.128858168 0.147783983 0.145573612
#>  [475] 0.154818419 0.156797840 0.156178199 0.140924109 0.122581453 0.122612071
#>  [481] 0.121019001 0.120899222 0.116671963 0.110437862 0.110411178 0.110491605
#>  [487] 0.110505946 0.111081624 0.106911034 0.110110418 0.110173486 0.166231016
#>  [493] 0.130074791 0.131263039 0.129166249 0.129288290 0.130957105 0.130625603
#>  [499] 0.131381457 0.128775716 0.118626520 0.127566930 0.127452547 0.089705468
#>  [505] 0.089538954 0.078560691 0.073138000 0.070229342 0.282064764 0.283856446
#>  [511] 0.283737377 0.272137528 0.265665887 0.276731014 0.274918177 0.274841911
#>  [517] 0.273085390 0.268568614 0.267871760 0.279477603 0.278428244 0.278774269
#>  [523] 0.276410362 0.290524861 0.265993230 0.249621329 0.250442530 0.253763561
#>  [529] 0.248467492 0.248445712 0.247853991 0.245852378 0.241554462 0.241674150
#>  [535] 0.230983670 0.229390463 0.246384048 0.256050917 0.262585964 0.111619074
#>  [541] 0.114136169 0.115883165 0.120870052 0.124263966 0.124412059 0.128038224
#>  [547] 0.128296648 0.129656760 0.132787720 0.133029853 0.126820773 0.126615357
#>  [553] 0.120468441 0.121867209 0.121636061 0.135041365 0.142468272 0.128593022
#>  [559] 0.105396443 0.109747091 0.109849202 0.086736065 0.083122180 0.073246896
#>  [565] 0.075636744 0.083081622 0.082979510 0.077312901 0.077227762 0.070767629
#>  [571] 0.079285518 0.080198565 0.073682320 0.073744890 0.073897090 0.072207619
#>  [577] 0.072315256 0.072026417 0.071841543 0.069989925 0.068447038 0.070922857
#>  [583] 0.070858413 0.071233519 0.069376268 0.067503428 0.053043869 0.053003260
#>  [589] 0.055107059 0.059487459 0.058398020 0.055999467 0.060963275 0.060599530
#>  [595] 0.062315434 0.059049833 0.059021327 0.049195131 0.052563154 0.053117160
#>  [601] 0.051373899 0.035982158 0.034838682 0.021906720 0.021801969 0.021459038
#>  [607] 0.022029246 0.022434823 0.023177764 0.023182306 0.024591708 0.025320765
#>  [613] 0.027035824 0.020602918 0.019087680 0.019380799 0.020247043 0.020988140
#>  [619] 0.021027157 0.020905956 0.019357695 0.017870460 0.019552580 0.021868183
#>  [625] 0.022101615 0.022481285 0.022323581 0.022227820 0.016733391 0.014650465
#>  [631] 0.012014874 0.012855765 0.016784589 0.016194392 0.016249813 0.016272031
#>  [637] 0.016580120 0.016751462 0.016629444 0.015793482 0.015757755 0.030196963
#>  [643] 0.030301653 0.029106714 0.042409274 0.042377932 0.040061865 0.035902059
#>  [649] 0.032914715 0.034190533 0.032849284 0.058907216 0.063229944 0.063213299
#>  [655] 0.053204737 0.053376364 0.053257205 0.053285675 0.053438578 0.053362609
#>  [661] 0.054776783 0.057170879 0.055311551 0.056608252 0.056593143 0.035165162
#>  [667] 0.035116992 0.036553818 0.036534991 0.036414937 0.036896091 0.036809918
#>  [673] 0.033066682 0.033021035 0.033024707 0.031501369 0.031540032 0.031915586
#>  [679] 0.032506253 0.033911032 0.033168447 0.036011374 0.014001458 0.014018977
#>  [685] 0.014046554 0.016272249 0.018216826 0.025661629 0.025618547 0.026298870
#>  [691] 0.028309583 0.028262891 0.027688032 0.027356769 0.047185749 0.046377599
#>  [697] 0.082365788 0.082571877 0.078754056 0.078487586 0.075789933 0.073229067
#>  [703] 0.073419658 0.067438301 0.067283603 0.064471621 0.059807927 0.059748128
#>  [709] 0.059801780 0.062120510 0.061817587 0.060220896 0.057463037 0.058115558
#>  [715] 0.061272900 0.061262269 0.060523306 0.066412486 0.059802387 0.059791012
#>  [721] 0.095711216 0.107003892 0.108291580 0.090256089 0.089858900 0.077024903
#>  [727] 0.078505793 0.078176052 0.071484390 0.064683119 0.064686059 0.065967559
#>  [733] 0.065639538 0.065613262 0.048343714 0.048213253 0.048157805 0.048144530
#>  [739] 0.048115948 0.048103592 0.048073415 0.048133606 0.048680932 0.049064012
#>  [745] 0.048570097 0.047010151 0.046955321 0.047241258 0.046593381 0.045962073
#>  [751] 0.045954609 0.027822005 0.027758906 0.025641314 0.007232209 0.005276581
#>  [757] 0.005099206 0.006197790 0.006129152 0.005631235 0.024252635 0.024246611
#>  [763] 0.025037559 0.025607868 0.025610714 0.054841078 0.055702014 0.060541544
#>  [769] 0.067844218 0.067824011 0.058534946 0.055136446 0.042821961 0.042905983
#>  [775] 0.042157617 0.040309263 0.040634503 0.040715706 0.040678472 0.042130282
#>  [781] 0.044345986 0.044320028 0.043196107 0.041941717 0.042360546 0.042265086
#>  [787] 0.042676405 0.043652016 0.041284935 0.041324720 0.040841091 0.006556578
#>  [793] 0.006592618 0.003815015 0.003728428 0.004862394 0.007175962 0.006812795
#>  [799] 0.005792331 0.007156930 0.006143678 0.006142161 0.007818399 0.009210094
#>  [805] 0.011203998 0.011379292 0.011846369 0.011293087 0.011357749 0.011350234
#>  [811] 0.016422666 0.016413361 0.016420228 0.016606784 0.016754615 0.016775826
#>  [817] 0.016718136 0.013137665 0.013098388 0.013275835 0.013430889 0.013423495
#>  [823] 0.017187605 0.017189261 0.017358832 0.017361461 0.017128931 0.016907998
#>  [829] 0.017187966 0.017754474 0.017453649 0.017333941 0.017711529 0.017171378
#>  [835] 0.017736524 0.016913278 0.016558606 0.016456667 0.016729638 0.016648307
#>  [841] 0.016651710 0.014057750 0.014059805 0.014036510 0.014061446 0.014102574
#>  [847] 0.012981575 0.017806121 0.021064560 0.021258083 0.021003693 0.021573397
#>  [853] 0.021594256 0.017694330 0.017325400 0.017832601 0.017825814 0.017795229
#>  [859] 0.017615873 0.017238660 0.017478286 0.017417379 0.019145155 0.019430541
#>  [865] 0.019422529 0.019391342 0.019611498 0.019762924 0.021393426 0.020442429
#>  [871] 0.020209479 0.019872439 0.018640163 0.018645079 0.017993205 0.017965725
#>  [877] 0.017680517 0.019345094 0.009006035 0.009933115 0.009682840 0.009986814
#>  [883] 0.008455049 0.008448580 0.008203404 0.007250965 0.007115075 0.023737458
#>  [889] 0.026097645 0.026493168 0.026976680 0.026334886 0.027269600 0.025762201
#>  [895] 0.069012743 0.074364672 0.074436734 0.076388797 0.077940998 0.076795853
#>  [901] 0.080031917 0.082120981 0.083540078 0.092319950 0.088069457 0.090429820
#>  [907] 0.089059356 0.084019711 0.081501248 0.081654052 0.080932185 0.084099762
#>  [913] 0.084124091 0.084022674 0.083943658 0.085194658 0.094435280 0.092437417
#>  [919] 0.088015827 0.085594402 0.085786563 0.086752483 0.089266060 0.090401295
#>  [925] 0.095616698 0.049977565 0.038779908 0.038710011 0.034313189 0.028493926
#>  [931] 0.029008397 0.029011202 0.025685326 0.025752623 0.026650270 0.031490294
#>  [937] 0.032263204 0.028243503 0.028630690 0.030291867 0.030924698 0.030444516
#>  [943] 0.030446930 0.030418265 0.030413594 0.030346420 0.030550920 0.030486794
#>  [949] 0.031931184 0.031650363 0.032872832 0.033121399 0.028794726 0.028764833
#>  [955] 0.028756417 0.028723691 0.028070708 0.031137807 0.031681379 0.032650618
#>  [961] 0.038495319 0.038719965 0.038732261 0.045588460 0.045498805 0.046307979
#>  [967] 0.060806380 0.060497335 0.062402872 0.062510868 0.060852578 0.061613571
#>  [973] 0.062130830 0.062339155 0.062225788 0.056203588 0.056520413 0.056610211
#>  [979] 0.058746744 0.059258367 0.056093183 0.055726302 0.060764724 0.192257888
#>  [985] 0.200053764 0.208592188 0.210384014 0.194122793 0.168310544 0.165256934
#>  [991] 0.160673611 0.134613490 0.135978300 0.140226191 0.133212143 0.144643995
#>  [997] 0.152299201 0.153943445 0.163257316 0.161773389 0.168953159 0.168147473
#> [1003] 0.168379987 0.166335792 0.168645553 0.171923532 0.352677742 0.360635640
#> [1009] 0.321435438 0.307916024 0.319211325 0.338370729 0.327426775 0.326695465
#> [1015] 0.311506019 0.298335887 0.284488934 0.284020810 0.277268455 0.267392450
#> [1021] 0.258274121 0.252679444 0.246781204 0.231587865 0.198242386 0.187433985
#> [1027] 0.174258414 0.171132535 0.164812417 0.146102797 0.149876973 0.138017354
#> [1033] 0.128214177 0.105730855 0.109455486 0.101219406 0.108175160 0.097178834
#> [1039] 0.094878897 0.101073209 0.101232461 0.093151819 0.074371159 0.055923076
#> [1045] 0.055240613 0.054021220 0.053860523 0.056274107 0.060099088 0.060254662
#> [1051] 0.060458565 0.062599847 0.061743965 0.060960973 0.075074549 0.075193137
#> [1057] 0.073749576 0.073159841 0.068546244 0.068737432 0.068139712 0.067056148
#> [1063] 0.070638730 0.070109764 0.077100911 0.074326297 0.074143396 0.066452590
#> [1069] 0.066698146 0.067830217 0.068792508 0.069247459 0.068932029 0.067426756
#> [1075] 0.099501310 0.102622481 0.103383401 0.101679905 0.099104381 0.091809634
#> [1081] 0.095187034 0.096621148
roll_rayleigh(dat.PoR$azi.PoR, prd = 135, unc = dat$unc)
#>    [1]  0.872444614  0.878948222  0.874690720  0.851892423  0.853665153
#>    [6]  0.855259330  0.847821371  0.860044752  0.860883340  0.861604688
#>   [11]  0.861449764  0.861410461  0.856638415  0.868073440  0.873810937
#>   [16]  0.848966086  0.850488132  0.848826039  0.849922237  0.838384850
#>   [21]  0.841683761  0.840644277  0.777152045  0.777347536  0.777344546
#>   [26]  0.785084532  0.784695262  0.788729319  0.793551363  0.809316733
#>   [31]  0.810067048  0.809255063  0.805082465  0.820688560  0.828880732
#>   [36]  0.840740348  0.841469087  0.846010505  0.844459434  0.858268471
#>   [41]  0.858959639  0.859172436  0.857143968  0.860229333  0.855544351
#>   [46]  0.850509653  0.874011174  0.876950819  0.882829463  0.880728157
#>   [51]  0.857268784  0.865074989  0.876476421  0.930170386  0.929935177
#>   [56]  0.929988590  0.928094719  0.928074966  0.928277314  0.926798724
#>   [61]  0.924011224  0.922415126  0.922714044  0.926751383  0.922618947
#>   [66]  0.928311493  0.922348220  0.921852956  0.921588754  0.919684509
#>   [71]  0.912372233  0.907650200  0.907341920  0.909687623  0.909793641
#>   [76]  0.900752832  0.901196457  0.901331260  0.893656713  0.886588717
#>   [81]  0.892736980  0.924941055  0.844502061  0.831206231  0.830486392
#>   [86]  0.841774051  0.841251114  0.842082918  0.840260687  0.839582885
#>   [91]  0.839209880  0.834178699  0.834885150  0.829611916  0.814819329
#>   [96]  0.777599033  0.763270843  0.785664660  0.786141042  0.786455524
#>  [101]  0.786850098  0.778428434  0.781226497  0.778183578  0.778944472
#>  [106]  0.772338443  0.740959981  0.740892840  0.743922630  0.744745531
#>  [111]  0.767426494  0.758772228  0.741716024  0.790925153  0.779903778
#>  [116]  0.782161122  0.737129921  0.736079207  0.722927714  0.722473037
#>  [121]  0.722920840  0.737316609  0.737675072  0.738721058  0.748295097
#>  [126]  0.711314507  0.763006167  0.778933852  0.747029343  0.746466349
#>  [131]  0.744972741  0.740559703  0.751339008  0.753359740  0.752402962
#>  [136]  0.752154891  0.760726504  0.767245812  0.778031788  0.774681288
#>  [141]  0.815170043  0.809473842  0.810388188  0.812548655  0.829207671
#>  [146]  0.844242349  0.843776809  0.880417122  0.881126816  0.878972412
#>  [151]  0.859849183  0.859606303  0.852522072  0.848675022  0.856869846
#>  [156]  0.878172457  0.917339369  0.922435089  0.922501965  0.928117881
#>  [161]  0.928281307  0.944013232  0.950562465  0.953503117  0.954970519
#>  [166]  0.956286271  0.958139152  0.958391747  0.958032076  0.957225816
#>  [171]  0.959464956  0.954761591  0.955076016  0.954452414  0.950476987
#>  [176]  0.945694959  0.944584912  0.944312494  0.942142433  0.941352964
#>  [181]  0.943008295  0.968583480  0.970091204  0.970037088  0.972501812
#>  [186]  0.965829125  0.958014255  0.956356816  0.951037746  0.951663035
#>  [191]  0.947630506  0.947142587  0.939104181  0.927058132  0.923156552
#>  [196]  0.916583255  0.931334063  0.910722891  0.910320587  0.904236833
#>  [201]  0.904781644  0.902920944  0.900323925  0.897394907  0.896837697
#>  [206]  0.881293796  0.873444459  0.871078589  0.868752541  0.865543706
#>  [211]  0.866111048  0.865706430  0.827950228  0.816338546  0.820062033
#>  [216]  0.821484797  0.826291159  0.746043972  0.746590849  0.711175890
#>  [221]  0.701760782  0.709816723  0.709201522  0.700835589  0.701143188
#>  [226]  0.702053765  0.702296410  0.670446009  0.676396256  0.678144340
#>  [231]  0.686912034  0.687150753  0.688531452  0.695288374  0.681045025
#>  [236]  0.683980744  0.709549191  0.711589066  0.715106496  0.703043822
#>  [241]  0.705556308  0.705497276  0.725160095  0.768239998  0.772465847
#>  [246]  0.769308777  0.777121091  0.772529823  0.839977653  0.847607758
#>  [251]  0.907133125  0.910172279  0.911234791  0.909912319  0.916202532
#>  [256]  0.916956281  0.922234762  0.922032814  0.918149473  0.918313814
#>  [261]  0.934824640  0.932064138  0.931450604  0.924072035  0.924528223
#>  [266]  0.925089716  0.910428505  0.905406005  0.908986209  0.924074337
#>  [271]  0.911687136  0.915499434  0.917458977  0.910073532  0.883873263
#>  [276]  0.849971062  0.856205798  0.850734587  0.850404137  0.850404113
#>  [281]  0.846703206  0.844708588  0.780809240  0.754050184  0.761648379
#>  [286]  0.763744022  0.762081311  0.748622397  0.748517929  0.746370565
#>  [291]  0.746847929  0.737397695  0.737418955  0.737524825  0.738160170
#>  [296]  0.733040509  0.727988406  0.729349473  0.730889040  0.742263674
#>  [301]  0.733371054  0.716446666  0.719885150  0.720321165  0.719869755
#>  [306]  0.720884723  0.744484804  0.751669531  0.759586698  0.759694830
#>  [311]  0.759604003  0.747195567  0.748945941  0.805798631  0.841070573
#>  [316]  0.823627691  0.828155086  0.828767964  0.822645755  0.821877125
#>  [321]  0.756144571  0.755918751  0.780322546  0.787929279  0.757075463
#>  [326]  0.769129030  0.776653521  0.775702394  0.776294560  0.771139608
#>  [331]  0.776134546  0.771435326  0.771578452  0.758711657  0.755711630
#>  [336]  0.766041470  0.765233161  0.765829856  0.754395888  0.746809287
#>  [341]  0.730278174  0.737349993  0.743870122  0.743606997  0.747083239
#>  [346]  0.756199109  0.745116178  0.744624475  0.744852204  0.741389198
#>  [351]  0.737381412  0.780420439  0.779660222  0.770104286  0.770590979
#>  [356]  0.814380368  0.786512520  0.767795222  0.761679773  0.768579635
#>  [361]  0.773578113  0.761098894  0.781657070  0.658634214  0.661914406
#>  [366]  0.663848632  0.634592167  0.619954685  0.615701183  0.615722962
#>  [371]  0.616722912  0.627210759  0.612194673  0.593052963  0.562287462
#>  [376]  0.553525733  0.521455510  0.550586767  0.583993941  0.612136315
#>  [381]  0.623569097  0.615202667  0.632552333  0.575819477  0.594805051
#>  [386]  0.578206928  0.585475047  0.598998151  0.612539381  0.553074623
#>  [391]  0.536403110  0.538669295  0.543432084  0.531294352  0.614573717
#>  [396]  0.607351900  0.541729591  0.542112731  0.569687548  0.569305391
#>  [401]  0.588812571  0.586222468  0.588332146  0.583004536  0.606078289
#>  [406]  0.630871032  0.627441042  0.629659280  0.626881304  0.626691961
#>  [411]  0.609437197  0.607148015  0.611670994  0.607976485  0.535633424
#>  [416]  0.516407708  0.505505293  0.495320280  0.491938226  0.480802365
#>  [421]  0.550857364  0.540594898  0.541447945  0.517941905  0.508911281
#>  [426]  0.524994478  0.563976924  0.666104267  0.609576369  0.594313773
#>  [431]  0.602274760  0.547625249  0.542521745  0.485426202  0.475689750
#>  [436]  0.448225155  0.438360091  0.420133145  0.424265977  0.426116395
#>  [441]  0.386282531  0.396774195  0.379620228  0.408211497  0.393199312
#>  [446]  0.486412445  0.486196567  0.437927067  0.435220162  0.422813170
#>  [451]  0.405892121  0.404129452  0.373635872  0.369774810  0.369263511
#>  [456]  0.365224389  0.359289955  0.317561500  0.275267010  0.297633840
#>  [461]  0.375713305  0.327383216  0.343703303  0.339232186  0.376478279
#>  [466]  0.383017374  0.372691322  0.384110803  0.394575709  0.350686266
#>  [471]  0.344695698  0.350947553  0.282839318  0.293460246  0.258026211
#>  [476]  0.253935777  0.267461529  0.328513383  0.389575144  0.388339598
#>  [481]  0.400283310  0.404759315  0.429003028  0.469069068  0.470169985
#>  [486]  0.466182817  0.465902261  0.459052213  0.483580985  0.469320263
#>  [491]  0.466374683  0.360844616  0.452051813  0.449372060  0.467114303
#>  [496]  0.463533737  0.462869874  0.471373004  0.464659738  0.473773334
#>  [501]  0.510057813  0.475868793  0.479447794  0.555521803  0.560057051
#>  [506]  0.589732434  0.609217929  0.625531314  0.363012540  0.364964331
#>  [511]  0.370111542  0.394197512  0.414916966  0.394418113  0.407416631
#>  [516]  0.410786102  0.423660577  0.442990759  0.458289175  0.437488554
#>  [521]  0.462807147  0.437787014  0.474937228  0.440564517  0.461415199
#>  [526]  0.466595158  0.445791798  0.412555918  0.429611886  0.430670017
#>  [531]  0.424026363  0.425564427  0.439611197  0.434834036  0.452248976
#>  [536]  0.440665518  0.413866306  0.385224491  0.346164743  0.507673158
#>  [541]  0.487161924  0.477754287  0.460221716  0.445811678  0.443316133
#>  [546]  0.428354624  0.422005340  0.404934157  0.385811092  0.380790085
#>  [551]  0.404836511  0.411110522  0.438228659  0.421939804  0.428505589
#>  [556]  0.391024420  0.378641227  0.437067670  0.519215716  0.501075469
#>  [561]  0.497830491  0.562900361  0.579765907  0.618550254  0.613464660
#>  [566]  0.593819529  0.594107390  0.614561258  0.616696912  0.634977123
#>  [571]  0.621553348  0.613592823  0.636044788  0.633786839  0.632800412
#>  [576]  0.643791335  0.642561400  0.646191933  0.651203540  0.662960096
#>  [581]  0.674644305  0.665414982  0.666092747  0.661895339  0.678883379
#>  [586]  0.691134794  0.737736641  0.739039174  0.726642796  0.704056831
#>  [591]  0.710739054  0.725615983  0.717115432  0.721354381  0.714692954
#>  [596]  0.729375053  0.730332694  0.752988402  0.731774062  0.720560414
#>  [601]  0.739287239  0.784078695  0.794078349  0.815630731  0.819438945
#>  [606]  0.824013125  0.818163856  0.808862790  0.798023565  0.797921155
#>  [611]  0.785730813  0.777534983  0.754625643  0.785958733  0.796943189
#>  [616]  0.790858081  0.780394562  0.759490939  0.758508838  0.761564901
#>  [621]  0.783184998  0.793292856  0.777762273  0.756636002  0.751891094
#>  [626]  0.743139072  0.746714461  0.749119115  0.778522882  0.810522068
#>  [631]  0.846541935  0.848040346  0.847833801  0.860904502  0.859374625
#>  [636]  0.858751379  0.849252472  0.844755201  0.847396003  0.863149098
#>  [641]  0.861786609  0.835264718  0.831977317  0.852932542  0.809384191
#>  [646]  0.810253823  0.826169807  0.837540062  0.846144972  0.841044243
#>  [651]  0.846314829  0.778518099  0.765925394  0.766191908  0.787464141
#>  [656]  0.780419152  0.784542967  0.783265176  0.779228845  0.781998354
#>  [661]  0.772827178  0.760817794  0.767799003  0.763745392  0.764335883
#>  [666]  0.803018858  0.804358718  0.784605583  0.785535851  0.790004177
#>  [671]  0.778778740  0.783364281  0.795042109  0.797299559  0.797113299
#>  [676]  0.811176886  0.809214264  0.805450675  0.802671119  0.789733998
#>  [681]  0.792305185  0.765090250  0.838125622  0.839254645  0.837786584
#>  [686]  0.815610946  0.829287474  0.793458460  0.796684632  0.793487690
#>  [691]  0.792582714  0.793820043  0.802287580  0.801929220  0.738348803
#>  [696]  0.746397494  0.690880212  0.688789256  0.724953471  0.729177020
#>  [701]  0.737141960  0.761511808  0.759970506  0.775671692  0.777378012
#>  [706]  0.787358348  0.799545494  0.801625729  0.800526928  0.792603132
#>  [711]  0.796986631  0.803413753  0.823453489  0.814686485  0.792285889
#>  [716]  0.791591457  0.795575619  0.791075357  0.819861037  0.818893425
#>  [721]  0.737713336  0.727799682  0.710455538  0.698660579  0.694352366
#>  [726]  0.740856010  0.732163341  0.734716558  0.748565089  0.764370863
#>  [731]  0.764283118  0.758267727  0.759599942  0.760216615  0.783860429
#>  [736]  0.785140599  0.785682906  0.785780574  0.786492201  0.787232694
#>  [741]  0.788116173  0.787132575  0.782195205  0.777694329  0.785192462
#>  [746]  0.803042459  0.805070354  0.801163264  0.805258475  0.807684251
#>  [751]  0.808496069  0.881802092  0.885136649  0.904185101  0.942394978
#>  [756]  0.957222400  0.962105634  0.954470813  0.955608112  0.951297453
#>  [761]  0.886862554  0.886916450  0.870428681  0.862757817  0.862585945
#>  [766]  0.826442317  0.822736457  0.800183704  0.789772817  0.789080457
#>  [771]  0.806725776  0.814075272  0.834151001  0.831526678  0.836136773
#>  [776]  0.841860221  0.836684250  0.835975367  0.836557188  0.832700946
#>  [781]  0.828120345  0.828289816  0.834775090  0.840132459  0.840920774
#>  [786]  0.844094051  0.841640727  0.837066164  0.848591787  0.847970195
#>  [791]  0.854091302  0.923379125  0.922024648  0.951959422  0.955694481
#>  [796]  0.941255306  0.936493676  0.939611633  0.952685517  0.942912605
#>  [801]  0.947960142  0.947966460  0.939067940  0.934940864  0.927635212
#>  [806]  0.925757126  0.922864807  0.929548419  0.927509301  0.927754923
#>  [811]  0.920281541  0.920720697  0.920452500  0.918647100  0.917807537
#>  [816]  0.916870888  0.918092221  0.926575666  0.927262498  0.925508968
#>  [821]  0.924680020  0.925009473  0.909237652  0.909153230  0.907154616
#>  [826]  0.906959209  0.913490510  0.913778865  0.911046598  0.906511887
#>  [831]  0.910233988  0.910871379  0.908372864  0.913153225  0.909166060
#>  [836]  0.918530356  0.921080030  0.922291454  0.920440585  0.922735388
#>  [841]  0.922379889  0.930884579  0.930770648  0.931558291  0.929712862
#>  [846]  0.927928472  0.933334935  0.920855928  0.917106242  0.916251054
#>  [851]  0.918189949  0.913797572  0.912860022  0.930681170  0.933396118
#>  [856]  0.932055055  0.932164395  0.933140786  0.935683677  0.937790201
#>  [861]  0.935275601  0.936804212  0.930204946  0.929099106  0.929635717
#>  [866]  0.932198178  0.930590975  0.931114703  0.917794802  0.924064579
#>  [871]  0.924264136  0.926752173  0.929305547  0.929211311  0.931386304
#>  [876]  0.933110611  0.936079521  0.932591313  0.952233914  0.948286346
#>  [881]  0.949395858  0.946681166  0.952034632  0.950303668  0.951975672
#>  [886]  0.954095402  0.956021774  0.912824826  0.900957479  0.898783650
#>  [891]  0.897382100  0.901814352  0.894861727  0.900895273  0.803538054
#>  [896]  0.776389125  0.773739422  0.757419571  0.746079186  0.755667345
#>  [901]  0.746673382  0.736191118  0.729764768  0.717289499  0.727500963
#>  [906]  0.722211671  0.715240095  0.718952567  0.724254248  0.723235395
#>  [911]  0.721612337  0.712198067  0.711261092  0.713475797  0.715381783
#>  [916]  0.709240895  0.689362207  0.693364554  0.709638910  0.720147833
#>  [921]  0.715411522  0.706265649  0.695916636  0.694934920  0.678591701
#>  [926]  0.769788252  0.812153475  0.814641735  0.841026942  0.860790391
#>  [931]  0.857319741  0.857578127  0.871707445  0.869199433  0.863511710
#>  [936]  0.852234560  0.844485482  0.860061308  0.858303701  0.838950872
#>  [941]  0.833097724  0.835922352  0.835825516  0.837195550  0.837680112
#>  [946]  0.839778724  0.833974986  0.836043994  0.827123749  0.819724912
#>  [951]  0.799078403  0.805009591  0.830161246  0.831321433  0.831721472
#>  [956]  0.833355692  0.821168247  0.784092319  0.769321939  0.748620513
#>  [961]  0.692769149  0.685180810  0.684338246  0.622689122  0.627072174
#>  [966]  0.609809070  0.516988622  0.496422093  0.477658555  0.475238582
#>  [971]  0.484410896  0.470044468  0.459000691  0.454249146  0.456895953
#>  [976]  0.497031303  0.491604701  0.483473791  0.449234499  0.437957594
#>  [981]  0.458811705  0.461337600  0.397588014  0.178218712  0.136153468
#>  [986]  0.090460940  0.079128157  0.111158006  0.147673373  0.152669218
#>  [991]  0.182039781  0.220872430  0.207887333  0.180217647  0.190450321
#>  [996]  0.137737629  0.093567839  0.077449378  0.052557867  0.041655476
#> [1001]  0.014308076  0.033659201  0.039057880  0.055658508  0.038848234
#> [1006]  0.013804902 -0.191708885 -0.217202356 -0.145972670 -0.100915950
#> [1011] -0.132581072 -0.162498866 -0.129861109 -0.116423423 -0.075758776
#> [1016] -0.040272301  0.007033634  0.010944639  0.037040852  0.077975779
#> [1021]  0.110457877  0.117353856  0.145985945  0.181525468  0.249452420
#> [1026]  0.290948818  0.329309753  0.337950477  0.371180119  0.416275520
#> [1031]  0.409056015  0.451463447  0.478176215  0.532410670  0.511655332
#> [1036]  0.544189914  0.521731432  0.564688990  0.579089274  0.568209606
#> [1041]  0.571086172  0.601423261  0.641561165  0.679377169  0.685476961
#> [1046]  0.694390792  0.695469489  0.679647147  0.657508782  0.656148226
#> [1051]  0.655383922  0.641296333  0.647754964  0.654565117  0.601473639
#> [1056]  0.601072128  0.607834177  0.615180921  0.645725360  0.644139259
#> [1061]  0.642782503  0.649864132  0.624266719  0.628402998  0.605338445
#> [1066]  0.625547128  0.626456365  0.663497348  0.659901406  0.649711875
#> [1071]  0.640717779  0.633819204  0.634910606  0.644917401  0.600733350
#> [1076]  0.585463545  0.580159736  0.590933336  0.604520596  0.635946377
#> [1081]  0.619247266  0.607490389
roll_dispersion(dat.PoR$azi.PoR, mean = 135, w = 1/ dat$unc)
#>    [1] 0.031888846 0.030262945 0.031327320 0.037026894 0.036583712 0.036185167
#>    [7] 0.038044657 0.034988812 0.034779165 0.034598828 0.034637559 0.034647385
#>   [13] 0.035840396 0.032981640 0.031547266 0.037758478 0.037377967 0.037793490
#>   [19] 0.037519441 0.040403787 0.039579060 0.039838931 0.055711989 0.055663116
#>   [25] 0.055663864 0.053728867 0.053826184 0.052817670 0.051612159 0.047670817
#>   [31] 0.047483238 0.047686234 0.048729384 0.044827860 0.042779817 0.039814913
#>   [37] 0.039632728 0.038497374 0.038885142 0.035432882 0.035260090 0.035206891
#>   [43] 0.035714008 0.034942667 0.036113912 0.037372587 0.031497206 0.030762295
#>   [49] 0.029292634 0.029817961 0.035682804 0.033731253 0.030880895 0.017457404
#>   [55] 0.017516206 0.017502853 0.017976320 0.017981258 0.017930672 0.018300319
#>   [61] 0.018997194 0.019396218 0.019321489 0.018312154 0.019345263 0.017922127
#>   [67] 0.019412945 0.019536761 0.019602812 0.020078873 0.021906942 0.023087450
#>   [73] 0.023164520 0.022578094 0.022551590 0.024811792 0.024700886 0.024667185
#>   [79] 0.026585822 0.028352821 0.026815755 0.018764736 0.038874485 0.042198442
#>   [85] 0.042378402 0.039556487 0.039687222 0.039479271 0.039934828 0.040104279
#>   [91] 0.040197530 0.041455325 0.041278712 0.042597021 0.046295168 0.055600242
#>   [97] 0.059182289 0.053583835 0.053464739 0.053386119 0.053287475 0.055392891
#>  [103] 0.054693376 0.055454105 0.055263882 0.056915389 0.064760005 0.064776790
#>  [109] 0.064019343 0.063813617 0.058143376 0.060306943 0.064570994 0.052268712
#>  [115] 0.055024056 0.054459719 0.065717520 0.065980198 0.069268072 0.069381741
#>  [121] 0.069269790 0.065670848 0.065581232 0.065319735 0.062926226 0.072171373
#>  [127] 0.059248458 0.055266537 0.063242664 0.063383413 0.063756815 0.064860074
#>  [133] 0.062165248 0.061660065 0.061899259 0.061961277 0.059818374 0.058188547
#>  [139] 0.055492053 0.056329678 0.046207489 0.047631539 0.047402953 0.046862836
#>  [145] 0.042698082 0.038939413 0.039055798 0.029895720 0.029718296 0.030256897
#>  [151] 0.035037704 0.035098424 0.036869482 0.037831244 0.035782538 0.030456886
#>  [157] 0.020665158 0.019391228 0.019374509 0.017970530 0.017929673 0.013996692
#>  [163] 0.012359384 0.011624221 0.011257370 0.010928432 0.010465212 0.010402063
#>  [169] 0.010491981 0.010693546 0.010133761 0.011309602 0.011230996 0.011386897
#>  [175] 0.012380753 0.013576260 0.013853772 0.013921877 0.014464392 0.014661759
#>  [181] 0.014247926 0.007854130 0.007477199 0.007490728 0.006874547 0.008542719
#>  [187] 0.010496436 0.010910796 0.012240564 0.012084241 0.013092374 0.013214353
#>  [193] 0.015223955 0.018235467 0.019210862 0.020854186 0.017166484 0.022319277
#>  [199] 0.022419853 0.023940792 0.023804589 0.024269764 0.024919019 0.025651273
#>  [205] 0.025790576 0.029676551 0.031638885 0.032230353 0.032811865 0.033614073
#>  [211] 0.033472238 0.033573393 0.043012443 0.045915363 0.044984492 0.044628801
#>  [217] 0.043427210 0.063489007 0.063352288 0.072206027 0.074559805 0.072545819
#>  [223] 0.072699619 0.074791103 0.074714203 0.074486559 0.074425898 0.082388498
#>  [229] 0.080900936 0.080463915 0.078271991 0.078212312 0.077867137 0.076177907
#>  [235] 0.079738744 0.079004814 0.072612702 0.072102734 0.071223376 0.074239044
#>  [241] 0.073610923 0.073625681 0.068709976 0.057940000 0.056883538 0.057672806
#>  [247] 0.055719727 0.056867544 0.040005587 0.038098061 0.023216719 0.022456930
#>  [253] 0.022191302 0.022521920 0.020949367 0.020760930 0.019441310 0.019491797
#>  [259] 0.020462632 0.020421546 0.016293840 0.016983965 0.017137349 0.018981991
#>  [265] 0.018867944 0.018727571 0.022392874 0.023648499 0.022753448 0.018981416
#>  [271] 0.022078216 0.021125142 0.020635256 0.022481617 0.029031684 0.037507234
#>  [277] 0.035948551 0.037316353 0.037398966 0.037398972 0.038324198 0.038822853
#>  [283] 0.054797690 0.061487454 0.059587905 0.059063994 0.059479672 0.062844401
#>  [289] 0.062870518 0.063407359 0.063288018 0.065650576 0.065645261 0.065618794
#>  [295] 0.065459958 0.066739873 0.068002898 0.067662632 0.067277740 0.064434081
#>  [301] 0.066657237 0.070888333 0.070028712 0.069919709 0.070032561 0.069778819
#>  [307] 0.063878799 0.062082617 0.060103325 0.060076292 0.060098999 0.063201108
#>  [313] 0.062763515 0.048550342 0.039732357 0.044093077 0.042961229 0.042808009
#>  [319] 0.044338561 0.044530719 0.060963857 0.061020312 0.054919363 0.053017680
#>  [325] 0.060731134 0.057717742 0.055836620 0.056074402 0.055926360 0.057215098
#>  [331] 0.055966364 0.057141168 0.057105387 0.060322086 0.061072093 0.058489633
#>  [337] 0.058691710 0.058542536 0.061401028 0.063297678 0.067430457 0.065662502
#>  [343] 0.064032469 0.064098251 0.063229190 0.060950223 0.063720955 0.063843881
#>  [349] 0.063786949 0.064652701 0.065654647 0.054894890 0.055084944 0.057473928
#>  [355] 0.057352255 0.046404908 0.053371870 0.058051195 0.059580057 0.057855091
#>  [361] 0.056605472 0.059725276 0.054585733 0.085341446 0.084521399 0.084037842
#>  [367] 0.091351958 0.095011329 0.096074704 0.096069259 0.095819272 0.093197310
#>  [373] 0.096951332 0.101736759 0.109428134 0.111618567 0.119636123 0.112353308
#>  [379] 0.104001515 0.096965921 0.094107726 0.096199333 0.091861917 0.106045131
#>  [385] 0.101298737 0.105448268 0.103631238 0.100250462 0.096865155 0.111731344
#>  [391] 0.115899223 0.115332676 0.114141979 0.117176412 0.096356571 0.098162025
#>  [397] 0.114567602 0.114471817 0.107578113 0.107673652 0.102796857 0.103444383
#>  [403] 0.102916963 0.104248866 0.098480428 0.092282242 0.093139739 0.092585180
#>  [409] 0.093279674 0.093327010 0.097640701 0.098212996 0.097082252 0.098005879
#>  [415] 0.116091644 0.120898073 0.123623677 0.126169930 0.127015443 0.129799409
#>  [421] 0.112285659 0.114851276 0.114638014 0.120514524 0.122772180 0.118751380
#>  [427] 0.109005769 0.083473933 0.097605908 0.101421557 0.099431310 0.113093688
#>  [433] 0.114369564 0.128643450 0.131077562 0.137943711 0.140409977 0.144966714
#>  [439] 0.143933506 0.143470901 0.153429367 0.150806451 0.155094943 0.147947126
#>  [445] 0.151700172 0.128396889 0.128450858 0.140518233 0.141194960 0.144296708
#>  [451] 0.148526970 0.148967637 0.156591032 0.157556297 0.157684122 0.158693903
#>  [457] 0.160177511 0.170609625 0.181183247 0.175591540 0.156071674 0.168154196
#>  [463] 0.164074174 0.165191954 0.155880430 0.154245656 0.156827170 0.153972299
#>  [469] 0.151356073 0.162328434 0.163826076 0.162263112 0.179290171 0.176634939
#>  [475] 0.185493447 0.186516056 0.183134618 0.167871654 0.152606214 0.152915100
#>  [481] 0.149929172 0.148810171 0.142749243 0.132732733 0.132457504 0.133454296
#>  [487] 0.133524435 0.135236947 0.129104754 0.132669934 0.133406329 0.159788846
#>  [493] 0.136987047 0.137656985 0.133221424 0.134116566 0.134282532 0.132156749
#>  [499] 0.133835065 0.131556667 0.122485547 0.131032802 0.130138052 0.111119549
#>  [505] 0.109985737 0.102566892 0.097695518 0.093617171 0.159246865 0.158758917
#>  [511] 0.157472115 0.151450622 0.146270758 0.151395472 0.148145842 0.147303474
#>  [517] 0.144084856 0.139252310 0.135427706 0.140627861 0.134298213 0.140553247
#>  [523] 0.131265693 0.139858871 0.134646200 0.133351211 0.138552051 0.146861021
#>  [529] 0.142597029 0.142332496 0.143993409 0.143608893 0.140097201 0.141291491
#>  [535] 0.136937756 0.139833621 0.146533423 0.153693877 0.163458814 0.123081710
#>  [541] 0.128209519 0.130561428 0.134944571 0.138547081 0.139170967 0.142911344
#>  [547] 0.144498665 0.148766461 0.153547227 0.154802479 0.148790872 0.147222370
#>  [553] 0.140442835 0.144515049 0.142873603 0.152243895 0.155339693 0.140733082
#>  [559] 0.120196071 0.124731133 0.125542377 0.109274910 0.105058523 0.095362436
#>  [565] 0.096633835 0.101545118 0.101473152 0.096359686 0.095825772 0.091255719
#>  [571] 0.094611663 0.096601794 0.090988803 0.091553290 0.091799897 0.089052166
#>  [577] 0.089359650 0.088452017 0.087199115 0.084259976 0.081338924 0.083646255
#>  [583] 0.083476813 0.084526165 0.080279155 0.077216301 0.065565840 0.065240207
#>  [589] 0.068339301 0.073985792 0.072315236 0.068596004 0.070721142 0.069661405
#>  [595] 0.071326761 0.067656237 0.067416826 0.061752900 0.067056485 0.069859896
#>  [601] 0.065178190 0.053980326 0.051480413 0.046092317 0.045140264 0.043996719
#>  [607] 0.045459036 0.047784302 0.050494109 0.050519711 0.053567297 0.055616254
#>  [613] 0.061343589 0.053510317 0.050764203 0.052285480 0.054901360 0.060127265
#>  [619] 0.060372791 0.059608775 0.054203750 0.051676786 0.055559432 0.060841000
#>  [625] 0.062027226 0.064215232 0.063321385 0.062720221 0.055369280 0.047369483
#>  [631] 0.038364516 0.037989913 0.038041550 0.034773874 0.035156344 0.035312155
#>  [637] 0.037686882 0.038811200 0.038150999 0.034212726 0.034553348 0.041183820
#>  [643] 0.042005671 0.036766865 0.047653952 0.047436544 0.043457548 0.040614984
#>  [649] 0.038463757 0.039738939 0.038421293 0.055370475 0.058518651 0.058452023
#>  [655] 0.053133965 0.054895212 0.053864258 0.054183706 0.055192789 0.054500412
#>  [661] 0.056793206 0.059795552 0.058050249 0.059063652 0.058916029 0.049245286
#>  [667] 0.048910320 0.053848604 0.053616037 0.052498956 0.055305315 0.054158930
#>  [673] 0.051239473 0.050675110 0.050721675 0.047205778 0.047696434 0.048637331
#>  [679] 0.049332220 0.052566500 0.051923704 0.058727437 0.040468594 0.040186339
#>  [685] 0.040553354 0.046097263 0.042678132 0.051635385 0.050828842 0.051628077
#>  [691] 0.051854321 0.051544989 0.049428105 0.049517695 0.065412799 0.063400627
#>  [697] 0.077279947 0.077802686 0.068761632 0.067705745 0.065714510 0.059622048
#>  [703] 0.060007373 0.056082077 0.055655497 0.053160413 0.050113627 0.049593568
#>  [709] 0.049868268 0.051849217 0.050753342 0.049146562 0.044136628 0.046328379
#>  [715] 0.051928528 0.052102136 0.051106095 0.052231161 0.045034741 0.045276644
#>  [721] 0.065571666 0.068050079 0.072386115 0.075334855 0.076411909 0.064785997
#>  [727] 0.066959165 0.066320860 0.062858728 0.058907284 0.058929220 0.060433068
#>  [733] 0.060100014 0.059945846 0.054034893 0.053714850 0.053579273 0.053554857
#>  [739] 0.053376950 0.053191826 0.052970957 0.053216856 0.054451199 0.055576418
#>  [745] 0.053701885 0.049239385 0.048732412 0.049709184 0.048685381 0.048078937
#>  [751] 0.047875983 0.029549477 0.028715838 0.023953725 0.014401255 0.010694400
#>  [757] 0.009473591 0.011382297 0.011097972 0.012175637 0.028284362 0.028270887
#>  [763] 0.032392830 0.034310546 0.034353514 0.043389421 0.044315886 0.049954074
#>  [769] 0.052556796 0.052729886 0.048318556 0.046481182 0.041462250 0.042118331
#>  [775] 0.040965807 0.039534945 0.040828938 0.041006158 0.040860703 0.041824763
#>  [781] 0.042969914 0.042927546 0.041306228 0.039966885 0.039769806 0.038976487
#>  [787] 0.039589818 0.040733459 0.037852053 0.038007451 0.036477175 0.019155219
#>  [793] 0.019493838 0.012010144 0.011076380 0.014686173 0.015876581 0.015097092
#>  [799] 0.011828621 0.014271849 0.013009964 0.013008385 0.015233015 0.016264784
#>  [805] 0.018091197 0.018560718 0.019283798 0.017612895 0.018122675 0.018061269
#>  [811] 0.019929615 0.019819826 0.019886875 0.020338225 0.020548116 0.020782278
#>  [817] 0.020476945 0.018356083 0.018184375 0.018622758 0.018829995 0.018747632
#>  [823] 0.022690587 0.022711693 0.023211346 0.023260198 0.021627373 0.021555284
#>  [829] 0.022238351 0.023372028 0.022441503 0.022282155 0.022906784 0.021711694
#>  [835] 0.022708485 0.020367411 0.019729993 0.019427136 0.019889854 0.019316153
#>  [841] 0.019405028 0.017278855 0.017307338 0.017110427 0.017571784 0.018017882
#>  [847] 0.016666266 0.019786018 0.020723439 0.020937236 0.020452513 0.021550607
#>  [853] 0.021784994 0.017329707 0.016650970 0.016986236 0.016958901 0.016714804
#>  [859] 0.016079081 0.015552450 0.016181100 0.015798947 0.017448763 0.017725223
#>  [865] 0.017591071 0.016950456 0.017352256 0.017221324 0.020551300 0.018983855
#>  [871] 0.018933966 0.018311957 0.017673613 0.017697172 0.017153424 0.016722347
#>  [877] 0.015980120 0.016852172 0.011941522 0.012928414 0.012651035 0.013329709
#>  [883] 0.011991342 0.012424083 0.012006082 0.011476149 0.010994557 0.021793794
#>  [889] 0.024760630 0.025304087 0.025654475 0.024546412 0.026284568 0.024776182
#>  [895] 0.049115486 0.055902719 0.056565145 0.060645107 0.063480203 0.061083164
#>  [901] 0.063331655 0.065952220 0.067558808 0.070677625 0.068124759 0.069447082
#>  [907] 0.071189976 0.070261858 0.068936438 0.069191151 0.069596916 0.071950483
#>  [913] 0.072184727 0.071631051 0.071154554 0.072689776 0.077659448 0.076658862
#>  [919] 0.072590272 0.069963042 0.071147120 0.073433588 0.076020841 0.076266270
#>  [925] 0.080352075 0.057552937 0.046961631 0.046339566 0.039743265 0.034802402
#>  [931] 0.035670065 0.035605468 0.032073139 0.032700142 0.034122073 0.036941360
#>  [937] 0.038878630 0.034984673 0.035424075 0.040262282 0.041725569 0.041019412
#>  [943] 0.041043621 0.040701112 0.040579972 0.040055319 0.041506254 0.040989002
#>  [949] 0.043219063 0.045068772 0.050230399 0.048747602 0.042459689 0.042169642
#>  [955] 0.042069632 0.041661077 0.044707938 0.053976920 0.057669515 0.062844872
#>  [961] 0.076807713 0.078704797 0.078915439 0.094327719 0.093231956 0.097547732
#>  [967] 0.120752845 0.125894477 0.130585361 0.131190354 0.128897276 0.132488883
#>  [973] 0.135249827 0.136437714 0.135776012 0.125742174 0.127098825 0.129131552
#>  [979] 0.137691375 0.140510601 0.135297074 0.134665600 0.150602997 0.205445322
#>  [985] 0.215961633 0.227384765 0.230217961 0.222210498 0.213081657 0.211832695
#>  [991] 0.204490055 0.194781892 0.198028167 0.204945588 0.202387420 0.215565593
#>  [997] 0.226608040 0.230637656 0.236860533 0.239586131 0.246422981 0.241585200
#> [1003] 0.240235530 0.236085373 0.240287942 0.246548775 0.297927221 0.304300589
#> [1009] 0.286493167 0.275228988 0.283145268 0.290624716 0.282465277 0.279105856
#> [1015] 0.268939694 0.260068075 0.248241592 0.247263840 0.240739787 0.230506055
#> [1021] 0.222385531 0.220661536 0.213503514 0.204618633 0.187636895 0.177262796
#> [1027] 0.167672562 0.165512381 0.157204970 0.145931120 0.147735996 0.137134138
#> [1033] 0.130455946 0.116897333 0.122086167 0.113952521 0.119567142 0.108827753
#> [1039] 0.105227681 0.107947598 0.107228457 0.099644185 0.089609709 0.080155708
#> [1045] 0.078630760 0.076402302 0.076132628 0.080088213 0.085622804 0.085962943
#> [1051] 0.086154020 0.089675917 0.088061259 0.086358721 0.099631590 0.099731968
#> [1057] 0.098041456 0.096204770 0.088568660 0.088965185 0.089304374 0.087533967
#> [1063] 0.093933320 0.092899251 0.098665389 0.093613218 0.093385909 0.084125663
#> [1069] 0.085024649 0.087572031 0.089820555 0.091545199 0.091272348 0.088770650
#> [1075] 0.099816663 0.103634114 0.104960066 0.102266666 0.098869851 0.091013406
#> [1081] 0.095188183 0.098127403