Apply Rolling Functions using Circular Statistics

A generic function for applying a function to rolling margins of an array along an additional value.

Usage

distroll_circstats(
  x,
  distance,
  FUN,
  width = NULL,
  min_n = 2,
  align = c("right", "center", "left"),
  w = NULL,
  sort = TRUE,
  ...
)

distroll_confidence(
  x,
  distance,
  w = NULL,
  width = NULL,
  min_n = 2,
  align = c("right", "center", "left"),
  sort = TRUE,
  ...
)

distroll_dispersion(
  x,
  y,
  w = NULL,
  w.y = NULL,
  distance,
  width = NULL,
  min_n = 2,
  align = c("right", "center", "left"),
  sort = TRUE,
  ...
)

distroll_dispersion_sde(
  x,
  y,
  w = NULL,
  w.y = NULL,
  distance,
  width = NULL,
  min_n = 2,
  align = c("right", "center", "left"),
  sort = TRUE,
  ...
)

Arguments

x, y

vectors of numeric values in degrees. length(y) is either 1 or length(x)

distance

numeric. the independent variable along the values in x are sorted, e.g. the plate boundary distances

FUN

the function to be applied

width

numeric. the range across distance on which FUN should be applied on x. If NULL, then width is a number that separates the distances in 10 equal groups.

min_n

integer. The minimum values that should be considered in FUN (2 by default), otherwise NA.

align

specifies whether the index of the result should be left- or right-aligned or centered (default) compared to the rolling window of observations. This argument is only used if width represents widths.

w

numeric. the weighting for x

sort

logical. Should the values be sorted after distance prior to applying the function (TRUE by default).

...

optional arguments to FUN

w.y

numeric. the weighting for y

Value

two-column vectors of (sorted) x and the rolled statistics along distance.

Examples

data("plates")
plate_boundary <- subset(plates, plates$pair == "na-pa")
data("san_andreas")
PoR <- subset(nuvel1, nuvel1$plate.rot == "na")
san_andreas$distance <- distance_from_pb(
  x = san_andreas,
  PoR = PoR,
  pb = plate_boundary,
  tangential = TRUE
)
dat <- san_andreas |> cbind(PoR_shmax(san_andreas, PoR, "right"))

distroll_circstats(dat$azi.PoR, distance = dat$distance, w = 1 / dat$unc, FUN = circular_mean)
#>         distance         x   n
#>  [1,] -9.2265048  65.71825   7
#>  [2,] -7.6044815 127.98953  19
#>  [3,] -5.9824582 155.84338  14
#>  [4,] -4.3604350 154.02577  39
#>  [5,] -2.7384117 138.51753 118
#>  [6,] -1.1163884 138.28500 746
#>  [7,]  0.5056349 142.58238 131
#>  [8,]  2.1276582 147.30864   5
#>  [9,]  3.7496815 178.57158   2
#> [10,]  5.3717047        NA   1
distroll_confidence(dat$azi.PoR, distance = dat$distance, w = 1 / dat$unc)
#> Warning: NaNs produced
#> Warning: NaNs produced
#>         distance         x   n
#>  [1,] -9.2265048       NaN   7
#>  [2,] -7.6044815 22.366208  19
#>  [3,] -5.9824582 50.995038  14
#>  [4,] -4.3604350 21.515646  39
#>  [5,] -2.7384117  4.995946 118
#>  [6,] -1.1163884  2.212704 746
#>  [7,]  0.5056349  6.138640 131
#>  [8,]  2.1276582 10.082906   5
#>  [9,]  3.7496815       NaN   2
#> [10,]  5.3717047        NA   1
distroll_dispersion(dat$azi.PoR, y = 135, distance = dat$distance, w = 1 / dat$unc)
#>         distance          x   n
#>  [1,] -9.2265048 0.55904215   7
#>  [2,] -7.6044815 0.31555788  19
#>  [3,] -5.9824582 0.39396588  14
#>  [4,] -4.3604350 0.32627346  39
#>  [5,] -2.7384117 0.06533460 118
#>  [6,] -1.1163884 0.09485111 746
#>  [7,]  0.5056349 0.11039516 131
#>  [8,]  2.1276582 0.06141144   5
#>  [9,]  3.7496815 0.47800478   2
#> [10,]  5.3717047         NA   1
distroll_dispersion_sde(dat$azi.PoR, y = 135, distance = dat$distance, w = 1 / dat$unc, R = 100)
#>         distance          x   n
#>  [1,] -9.2265048 0.20049859   7
#>  [2,] -7.6044815 0.13461160  19
#>  [3,] -5.9824582 0.22762507  14
#>  [4,] -4.3604350 0.11928758  39
#>  [5,] -2.7384117 0.02085958 118
#>  [6,] -1.1163884 0.01200510 746
#>  [7,]  0.5056349 0.02967653 131
#>  [8,]  2.1276582 0.03488652   5
#>  [9,]  3.7496815 0.05802001   2
#> [10,]  5.3717047         NA   1