Stress2Grid

Stress pattern and wavelength analysis

Usage

stress2grid(
  x,
  stat = c("mean", "median"),
  lon_range = NULL,
  lat_range = NULL,
  gridsize = 2.5,
  min_data = 3,
  threshold = 25,
  arte_thres = 200,
  method_weighting = FALSE,
  quality_weighting = TRUE,
  dist_weight = c("inverse", "linear", "none"),
  dist_threshold = 0.1,
  R_range = seq(50, 1000, 50),
  ...
)

Source

https://github.com/MorZieg/Stress2Grid

Arguments

x

sf object containing

azi

SHmax in degree

unc

Uncertainties of SHmax in degree

type

Methods used for the determination of the direction of SHmax

stat

Should the direction of interpolated SHmax be based on the circular mean and standard deviation ("mean", the default) or on the circular median and interquartile range ("median")?

lon_range, lat_range

(optional) numeric vector specifying the minimum and maximum longitudes and latitudes.

gridsize

Numeric. Target spacing of the regular grid in decimal degree. Default is 2.5

min_data

Integer. Minimum number of data per bin. Default is 3

threshold

Numeric. Threshold for deviation of direction. Default is 25

arte_thres

Numeric. Maximum distance (in km) of the gridpoint to the next datapoint. Default is 200

method_weighting

Logical. If a method weighting should be applied: Default is FALSE.

quality_weighting

Logical. If a quality weighting should be applied: Default is TRUE.

dist_weight

Distance weighting method which should be used. One of "none", "linear", or "inverse" (the default).

dist_threshold

Numeric. Distance weight to prevent overweight of data nearby (0 to 1). Default is 0.1

R_range

Numeric value or vector specifying the search radius (in km). Default is seq(50, 1000, 50)

...

optional arguments to dist_greatcircle()

Value

sf object containing

lon,lat

longitude and latitude in degree

azi

Mean SHmax in degree

sd

Standard deviation of SHmax in degree

R

Search radius in km

mdr

Mean distance of datapoint per search radius

N

Number of data points in search radius

Details

Calculates the weighted mean and standard deviation of the stress data

References

Ziegler, M. O. and Heidbach, O. (2019). Matlab Script Stress2Grid v1.1. GFZ Data Services. doi:10.5880/wsm.2019.002

See also

Examples

data("san_andreas")
stress2grid(san_andreas)
#> Simple feature collection with 671 features and 7 fields
#> Geometry type: POINT
#> Dimension:     XY
#> Bounding box:  xmin: -124.57 ymin: 24.33 xmax: -107.07 ymax: 39.33
#> Geodetic CRS:  WGS 84
#> # A tibble: 671 × 8
#> # Groups:   R [20]
#>      lon   lat   azi    sd     R    mdr     N        geometry
#>    <dbl> <dbl> <dbl> <dbl> <dbl>  <dbl> <int>     <POINT [°]>
#>  1 -115.  24.3   NA   0       50 NA         0 (-114.57 24.33)
#>  2 -115.  24.3   NA   0      100 NA         0 (-114.57 24.33)
#>  3 -115.  24.3   NA   0      150 NA         0 (-114.57 24.33)
#>  4 -115.  24.3   NA   0      200 NA         1 (-114.57 24.33)
#>  5 -115.  24.3  179.  5.32   250  0.820     3 (-114.57 24.33)
#>  6 -115.  24.3  170. 25.0    350  0.796     9 (-114.57 24.33)
#>  7 -115.  24.3  168. 19.3    400  0.816    18 (-114.57 24.33)
#>  8 -115.  24.3  171. 14.8    450  0.884    56 (-114.57 24.33)
#>  9 -115.  24.3  171. 16.3    500  0.876   107 (-114.57 24.33)
#> 10 -115.  24.3  171. 15.5    550  0.832   139 (-114.57 24.33)
#> # … with 661 more rows