Theoretical Plate Tectonic Stress Paths

Construct \(\sigma_{Hmax}\) lines that are following small circles, great circles, or loxodromes of an Euler pole for the relative plate motion.

Usage

eulerpole_paths(x, type = c("sc", "gc", "ld"), n = 10, angle, cw)

eulerpole_smallcircles(x, n = 10)

eulerpole_greatcircles(x, n = 10)

eulerpole_loxodromes(x, n = 10, angle = 45, cw)

Arguments

x

Either an object of class "euler.pole" or data.frame containing coordinates of Euler pole in lat, lon, and rotation angle (optional).

type

Character string specifying the type of curves to export. Either "sm" for small circles (default), "gc" for great circles, or "ld" for loxodromes.

n

Number of equally spaced curves; n = 10 by default (angular distance between curves: 180 / n)

angle

Direction of loxodromes; angle = 45 by default.

cw

logical. Sense of loxodromes: TRUE for clockwise loxodromes (right-lateral displaced plate boundaries). FALSE for counterclockwise loxodromes (left-lateral displaced plate boundaries).

Value

sf object

Details

Maximum horizontal stress can be aligned to three types of curves related to relative plate motion:

Small circles

Lines that have a constant distance to the Euler pole. If x contains angle, output additionally gives absolute velocity on small circle (degree/Myr -> km/Myr).

Great circles

Paths of the the shortest distance between the Euler pole and its antipodal position.

Loxodromes

Lines of constant bearing, i.e. curves cutting small circles at a constant angle.

Note

If package "smoothr" is installed, the sf objects will be "densified" via smoothr::densify().

Author

Tobias Stephan

Examples

data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to
# Pacific plate

eulerpole_smallcircles(euler)
#> Simple feature collection with 11 features and 2 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -84.709 ymin: -179.3255 xmax: 84.709 ymax: 179.7315
#> Geodetic CRS:  WGS 84
#> Warning: bounding box has potentially an invalid value range for longlat data
#> # A tibble: 11 × 3
#>                                                           geometry     d abs_vel
#>  *                                                <LINESTRING [°]> <dbl>   <dbl>
#>  1 (-48.709 101.833, -48.709 101.833, -48.709 101.833, -48.709 10…     0     0  
#>  2 (-30.709 101.833, -30.71571 102.4799, -30.73582 103.1265, -30.…   -18    26.9
#>  3 (-12.709 101.833, -12.72024 102.9175, -12.75396 104.0016, -12.…   -36    51.2
#>  4 (5.291 101.833, 5.275842 103.2953, 5.230387 104.757, 5.154684 …   -54    70.4
#>  5 (23.291 101.833, 23.27168 103.6965, 23.21377 105.5586, 23.1173…   -72    82.8
#>  6 (41.291 101.833, 41.26618 104.228, 41.19178 106.6194, 41.06806…    90    87.1
#>  7 (59.291 101.833, 59.25627 105.1832, 59.15234 108.521, 58.97993…    72    82.8
#>  8 (77.291 101.833, 77.22258 108.4309, 77.0195 114.9037, 76.68801…    54    70.4
#>  9 (84.709 -78.167, 84.59139 -89.46284, 84.25303 -99.79473, 83.72…    36    51.2
#> 10 (66.709 -78.167, 66.69442 -79.57283, 66.65076 -80.97315, 66.57…    18    26.9
#> 11 (48.709 -78.167, 48.709 -78.167, 48.709 -78.167, 48.709 -78.16…     0     0  
eulerpole_greatcircles(euler)
#> Simple feature collection with 11 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -89.991 ymin: -179.9989 xmax: 89.991 ymax: 179.9785
#> Geodetic CRS:  WGS 84
#> Warning: bounding box has potentially an invalid value range for longlat data
#> # A tibble: 11 × 2
#>        d                                                                geometry
#>  * <dbl>                                                        <LINESTRING [°]>
#>  1     0 (-48.709 101.833, -48.609 101.833, -48.509 101.833, -48.409 101.833, -…
#>  2    36 (-48.709 101.833, -48.62806 101.7441, -48.54706 101.6554, -48.46599 10…
#>  3    72 (-48.709 101.833, -48.67801 101.689, -48.64684 101.5451, -48.61549 101…
#>  4   108 (-48.709 101.833, -48.73981 101.6888, -48.77044 101.5444, -48.80089 10…
#>  5   144 (-48.709 101.833, -48.78987 101.7438, -48.87067 101.6543, -48.95139 10…
#>  6     0 (-48.709 101.833, -48.809 101.833, -48.909 101.833, -49.009 101.833, -…
#>  7    36 (-48.709 101.833, -48.78987 101.9222, -48.87067 102.0117, -48.95139 10…
#>  8    72 (-48.709 101.833, -48.73981 101.9772, -48.77044 102.1216, -48.80089 10…
#>  9   108 (-48.709 101.833, -48.67801 101.977, -48.64684 102.1209, -48.61549 102…
#> 10   144 (-48.709 101.833, -48.62806 101.9219, -48.54706 102.0106, -48.46599 10…
#> 11     0 (-48.709 101.833, -48.609 101.833, -48.509 101.833, -48.409 101.833, -…
eulerpole_loxodromes(x = euler, angle = 45, n = 10, cw = FALSE)
#> Simple feature collection with 15 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -82.82221 ymin: -179.9787 xmax: 82.82221 ymax: 179.996
#> Geodetic CRS:  WGS 84
#> Warning: bounding box has potentially an invalid value range for longlat data
#> # A tibble: 15 × 2
#>        d                                                                geometry
#>  * <dbl>                                                        <LINESTRING [°]>
#>  1   108 (66.58404 -78.07054, 66.51323 -78.01693, 66.44238 -77.96404, 66.37148 …
#>  2   144 (77.35336 101.6024, 77.42378 101.3391, 77.49398 101.0733, 77.56395 100…
#>  3     0 (41.291 101.833, 41.36167 101.7388, 41.43227 101.6444, 41.50279 101.54…
#>  4    36 (5.933699 101.3044, 6.003953 101.2458, 6.074159 101.1871, 6.144316 101…
#>  5    72 (-30.1275 101.6185, -30.05698 101.5917, -29.98648 101.5648, -29.916 10…
#>  6   108 (-48.5214 101.7399, -48.4543 101.7063, -48.38724 101.6725, -48.32023 1…
#>  7   144 (-48.59325 101.5909, -48.55193 101.5042, -48.5107 101.4174, -48.46954 …
#>  8     0 (-48.70929 101.5338, -48.70954 101.4266, -48.70986 101.3195, -48.71026…
#>  9    36 (-48.82535 101.591, -48.86716 101.5043, -48.90905 101.4177, -48.95102 …
#> 10    72 (-48.89695 101.7412, -48.96434 101.7086, -49.03178 101.6762, -49.09927…
#> 11   108 (-48.89653 101.9268, -48.96356 101.961, -49.03054 101.9956, -49.09744 …
#> 12   144 (-48.82424 102.0762, -48.86512 102.1639, -48.9058 102.2519, -48.94626 …
#> 13     0 (-48.70793 102.1322, -48.70703 102.2393, -48.70586 102.3465, -48.70441…
#> 14    36 (-48.59215 102.0739, -48.54991 102.1596, -48.50747 102.245, -48.46482 …
#> 15    72 (-48.52098 101.9241, -48.45352 101.9561, -48.38601 101.9878, -48.31843…
eulerpole_loxodromes(x = euler, angle = 30, cw = TRUE)
#> Simple feature collection with 13 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -85.21953 ymin: -179.9698 xmax: 85.21953 ymax: 179.9581
#> Geodetic CRS:  WGS 84
#> Warning: bounding box has potentially an invalid value range for longlat data
#> # A tibble: 13 × 2
#>        d                                                                geometry
#>  * <dbl>                                                        <LINESTRING [°]>
#>  1   144 (-47.93824 101.5141, -47.85449 101.4813, -47.77068 101.4487, -47.68683…
#>  2     0 (-48.20753 100.8943, -48.15215 100.7944, -48.09657 100.6949, -48.04078…
#>  3    36 (-48.6678 100.6237, -48.66181 100.4929, -48.65551 100.3621, -48.64892 …
#>  4    72 (-49.14575 100.8139, -49.19187 100.7018, -49.23775 100.5894, -49.2834 …
#>  5   108 (-49.45697 101.4023, -49.53763 101.3538, -49.61821 101.3048, -49.69872…
#>  6   144 (-49.47885 102.1618, -49.5624 102.1969, -49.64599 102.2319, -49.7296 1…
#>  7     0 (-49.20269 102.7905, -49.25631 102.8946, -49.30995 102.9989, -49.36362…
#>  8    36 (-48.73753 103.044, -48.74063 103.1752, -48.74373 103.3065, -48.74684 …
#>  9    72 (-48.26342 102.8344, -48.21527 102.9424, -48.16714 103.0504, -48.11904…
#> 10   108 (-47.95948 102.251, -47.87845 102.2967, -47.79746 102.3424, -47.71651 …
#> 11   144 (-21.06283 101.833, -20.97623 101.8579, -20.88965 101.883, -20.80309 1…
#> 12     0 (41.291 101.833, 41.37758 101.8996, 41.46413 101.9664, 41.55063 102.03…
#> 13    36 (76.35517 -78.167, 76.26854 -78.26446, 76.18185 -78.36015, 76.09512 -7…
eulerpole_smallcircles(data.frame(lat = 30, lon = 10))
#> Simple feature collection with 11 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -84 ymin: -179.8454 xmax: 84 ymax: 179.7965
#> Geodetic CRS:  WGS 84
#> Warning: bounding box has potentially an invalid value range for longlat data
#> # A tibble: 11 × 2
#>                                                                   geometry     d
#>  *                                                        <LINESTRING [°]> <dbl>
#>  1 (-30 -170, -30 -170, -30 -170, -30 -170, -30 -170, -30 -170, -30 -170,…     0
#>  2 (-12 -170, -12.00774 -169.4314, -12.03093 -168.8632, -12.06958 -168.29…   -18
#>  3 (6 -170, 5.985529 -168.9363, 5.942134 -167.8735, 5.869864 -166.8123, 5…   -36
#>  4 (24 -170, 23.97832 -168.4063, 23.91332 -166.8145, 23.80513 -165.2266, …   -54
#>  5 (42 -170, 41.96867 -167.6973, 41.87482 -165.3999, 41.7188 -163.1132, 4…   -72
#>  6 (60 -170, 59.95107 -166.4035, 59.80475 -162.8282, 59.56246 -159.2943, …    90
#>  7 (78 -170, 77.88851 -161.8144, 77.56006 -153.9057, 77.03116 -146.4959, …    72
#>  8 (84 10, 83.81339 -3.638991, 83.28465 -15.74781, 82.4858 -25.60532, 81.…    54
#>  9 (66 10, 65.96464 7.401912, 65.85889 4.822514, 65.68372 2.279844, 65.44…    36
#> 10 (48 10, 47.98869 9.169016, 47.9548 8.339768, 47.8984 7.51398, 47.81962…    18
#> 11 (30 10, 30 10, 30 10, 30 10, 30 10, 30 10, 30 10, 30 10, 30 10, 30 10,…     0