Theoretical Plate Tectonic Stress Paths

Construct \(\sigma_\text{Hmax}\) lines that are following small circles, great circles, or loxodromes of an Euler pole for the relative plate motion.

Usage

eulerpole_paths(x, type = c("sc", "gc", "ld"), n = 10, angle, cw)

eulerpole_smallcircles(x, n = 10)

eulerpole_greatcircles(x, n = 10)

eulerpole_loxodromes(x, n = 10, angle = 45, cw)

Arguments

x

Either an object of class "euler.pole" or data.frame containing coordinates of Euler pole in lat, lon, and rotation angle (optional).

type

Character string specifying the type of curves to export. Either "sm" for small circles (default), "gc" for great circles, or "ld" for loxodromes.

n

Number of equally spaced curves; n = 10 by default (angular distance between curves: 180 / n)

angle

Direction of loxodromes; angle = 45 by default.

cw

logical. Sense of loxodromes: TRUE for clockwise loxodromes (right-lateral displaced plate boundaries). FALSE for counterclockwise loxodromes (left-lateral displaced plate boundaries).

Value

sf object

Details

Maximum horizontal stress can be aligned to three types of curves related to relative plate motion:

Small circles

Lines that have a constant distance to the Euler pole. If x contains angle, output additionally gives absolute velocity on small circle (degree/Myr -> km/Myr).

Great circles

Paths of the the shortest distance between the Euler pole and its antipodal position.

Loxodromes

Lines of constant bearing, i.e. curves cutting small circles at a constant angle.

Note

If package "smoothr" is installed, the sf objects will be "densified" via smoothr::densify().

Author

Tobias Stephan

Examples

data("nuvel1")
euler <- subset(nuvel1, nuvel1$plate.rot == "na") # North America relative to
# Pacific plate

eulerpole_smallcircles(euler)
#> Simple feature collection with 11 features and 2 fields
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -90 xmax: 180 ymax: 90
#> CRS:           +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> # A tibble: 11 × 3
#>                                                           geometry     d abs_vel
#>  *                                                <LINESTRING [°]> <dbl>   <dbl>
#>  1 (-180 -90, -178.2 -90, -176.4 -90, -174.6 -90, -172.8 -90, -17…     0     0  
#>  2 (-180 -72, -178.2 -72, -176.4 -72, -174.6 -72, -172.8 -72, -17…   -18    26.9
#>  3 (-180 -54, -178.2 -54, -176.4 -54, -174.6 -54, -172.8 -54, -17…   -36    51.2
#>  4 (-180 -36, -178.2 -36, -176.4 -36, -174.6 -36, -172.8 -36, -17…   -54    70.4
#>  5 (-180 -18, -178.2 -18, -176.4 -18, -174.6 -18, -172.8 -18, -17…   -72    82.8
#>  6 (-180 0, -178.2 0, -176.4 0, -174.6 0, -172.8 0, -171 0, -169.…    90    87.1
#>  7 (-180 18, -178.2 18, -176.4 18, -174.6 18, -172.8 18, -171 18,…    72    82.8
#>  8 (-180 36, -178.2 36, -176.4 36, -174.6 36, -172.8 36, -171 36,…    54    70.4
#>  9 (-180 54, -178.2 54, -176.4 54, -174.6 54, -172.8 54, -171 54,…    36    51.2
#> 10 (-180 72, -178.2 72, -176.4 72, -174.6 72, -172.8 72, -171 72,…    18    26.9
#> 11 (-180 90, -178.2 90, -176.4 90, -174.6 90, -172.8 90, -171 90,…     0     0  
eulerpole_greatcircles(euler)
#> Simple feature collection with 11 features and 1 field
#> Geometry type: GEOMETRY
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -90 xmax: 180 ymax: 90
#> CRS:           +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> # A tibble: 11 × 2
#>        d                                                                geometry
#>  * <dbl>                                                          <GEOMETRY [°]>
#>  1     0 MULTILINESTRING ((180 -90, 180 -89.9, 180 -89.8, 180 -89.7, 180 -89.6,…
#>  2    36 LINESTRING (144 -90, 144 -89.9, 144 -89.8, 144 -89.7, 144 -89.6, 144 -…
#>  3    72 LINESTRING (108 -90, 108 -89.9, 108 -89.8, 108 -89.7, 108 -89.6, 108 -…
#>  4   108 LINESTRING (72 -90, 72 -89.9, 72 -89.8, 72 -89.7, 72 -89.6, 72 -89.5, …
#>  5   144 LINESTRING (36 -90, 36 -89.9, 36 -89.8, 36 -89.7, 36 -89.6, 36 -89.5, …
#>  6     0 LINESTRING (0 -90, 0 -89.9, 0 -89.8, 0 -89.7, 0 -89.6, 0 -89.5, 0 -89.…
#>  7    36 LINESTRING (-36 -90, -36 -89.9, -36 -89.8, -36 -89.7, -36 -89.6, -36 -…
#>  8    72 LINESTRING (-72 -90, -72 -89.9, -72 -89.8, -72 -89.7, -72 -89.6, -72 -…
#>  9   108 LINESTRING (-108 -90, -108 -89.9, -108 -89.8, -108 -89.7, -108 -89.6, …
#> 10   144 LINESTRING (-144 -90, -144 -89.9, -144 -89.8, -144 -89.7, -144 -89.6, …
#> 11     0 MULTILINESTRING ((-180 -90, -180 -89.9, -180 -89.8, -180 -89.7, -180 -…
eulerpole_loxodromes(x = euler, angle = 45, n = 10, cw = FALSE)
#> Simple feature collection with 15 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -89.80256 xmax: 180 ymax: 89.80256
#> CRS:           +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> # A tibble: 15 × 2
#>        d                                                                geometry
#>  * <dbl>                                                        <LINESTRING [°]>
#>  1   108 (162.1974 -89.80256, 162.2681 -89.73185, 162.3389 -89.66114, 162.4096 …
#>  2   144 (126.1974 -89.80256, 126.2681 -89.73185, 126.3389 -89.66114, 126.4096 …
#>  3     0 (90.19744 -89.80256, 90.26815 -89.73185, 90.33886 -89.66114, 90.40957 …
#>  4    36 (54.19744 -89.80256, 54.26815 -89.73185, 54.33886 -89.66114, 54.40957 …
#>  5    72 (18.19744 -89.80256, 18.26815 -89.73185, 18.33886 -89.66114, 18.40957 …
#>  6   108 (-17.80256 -89.80256, -17.73185 -89.73185, -17.66114 -89.66114, -17.59…
#>  7   144 (-53.80256 -89.80256, -53.73185 -89.73185, -53.66114 -89.66114, -53.59…
#>  8     0 (-89.80256 -89.80256, -89.73185 -89.73185, -89.66114 -89.66114, -89.59…
#>  9    36 (-125.8026 -89.80256, -125.7319 -89.73185, -125.6611 -89.66114, -125.5…
#> 10    72 (-161.8026 -89.80256, -161.7319 -89.73185, -161.6611 -89.66114, -161.5…
#> 11   108 (-179.4178 -71.41778, -179.3471 -71.34707, -179.2764 -71.27636, -179.2…
#> 12   144 (-179.3553 -35.35534, -179.2846 -35.28463, -179.2139 -35.21392, -179.1…
#> 13     0 (-180 0, -179.9293 0.07071068, -179.8586 0.1414214, -179.7879 0.212132…
#> 14    36 (-179.9376 36.06245, -179.8668 36.13316, -179.7961 36.20387, -179.7254…
#> 15    72 (-179.8751 72.12489, -179.8044 72.1956, -179.7337 72.26631, -179.663 7…
eulerpole_loxodromes(x = euler, angle = 30, cw = TRUE)
#> Simple feature collection with 13 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -89.20062 xmax: 180 ymax: 89.20062
#> CRS:           +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> # A tibble: 13 × 2
#>        d                                                                geometry
#>  * <dbl>                                                        <LINESTRING [°]>
#>  1   144 (180 62.35383, 179.95 62.44043, 179.9 62.52703, 179.85 62.61364, 179.8…
#>  2     0 (180 0, 179.95 0.08660254, 179.9 0.1732051, 179.85 0.2598076, 179.8 0.…
#>  3    36 (180 -62.35383, 179.95 -62.26723, 179.9 -62.18062, 179.85 -62.09402, 1…
#>  4    72 (159.5 -89.20062, 159.45 -89.11401, 159.4 -89.02741, 159.35 -88.94081,…
#>  5   108 (123.5 -89.20062, 123.45 -89.11401, 123.4 -89.02741, 123.35 -88.94081,…
#>  6   144 (87.5 -89.20062, 87.45 -89.11401, 87.4 -89.02741, 87.35 -88.94081, 87.…
#>  7     0 (51.5 -89.20062, 51.45 -89.11401, 51.4 -89.02741, 51.35 -88.94081, 51.…
#>  8    36 (15.5 -89.20062, 15.45 -89.11401, 15.4 -89.02741, 15.35 -88.94081, 15.…
#>  9    72 (-20.5 -89.20062, -20.55 -89.11401, -20.6 -89.02741, -20.65 -88.94081,…
#> 10   108 (-56.5 -89.20062, -56.55 -89.11401, -56.6 -89.02741, -56.65 -88.94081,…
#> 11   144 (-92.5 -89.20062, -92.55 -89.11401, -92.6 -89.02741, -92.65 -88.94081,…
#> 12     0 (-128.5 -89.20062, -128.55 -89.11401, -128.6 -89.02741, -128.65 -88.94…
#> 13    36 (-164.5 -89.20062, -164.55 -89.11401, -164.6 -89.02741, -164.65 -88.94…
eulerpole_smallcircles(data.frame(lat = 30, lon = 10))
#> Simple feature collection with 11 features and 1 field
#> Geometry type: LINESTRING
#> Dimension:     XY
#> Bounding box:  xmin: -180 ymin: -90 xmax: 180 ymax: 90
#> CRS:           +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0
#> # A tibble: 11 × 2
#>                                                                   geometry     d
#>  *                                                        <LINESTRING [°]> <dbl>
#>  1 (-180 -90, -178.2 -90, -176.4 -90, -174.6 -90, -172.8 -90, -171 -90, -…     0
#>  2 (-180 -72, -178.2 -72, -176.4 -72, -174.6 -72, -172.8 -72, -171 -72, -…   -18
#>  3 (-180 -54, -178.2 -54, -176.4 -54, -174.6 -54, -172.8 -54, -171 -54, -…   -36
#>  4 (-180 -36, -178.2 -36, -176.4 -36, -174.6 -36, -172.8 -36, -171 -36, -…   -54
#>  5 (-180 -18, -178.2 -18, -176.4 -18, -174.6 -18, -172.8 -18, -171 -18, -…   -72
#>  6 (-180 0, -178.2 0, -176.4 0, -174.6 0, -172.8 0, -171 0, -169.2 0, -16…    90
#>  7 (-180 18, -178.2 18, -176.4 18, -174.6 18, -172.8 18, -171 18, -169.2 …    72
#>  8 (-180 36, -178.2 36, -176.4 36, -174.6 36, -172.8 36, -171 36, -169.2 …    54
#>  9 (-180 54, -178.2 54, -176.4 54, -174.6 54, -172.8 54, -171 54, -169.2 …    36
#> 10 (-180 72, -178.2 72, -176.4 72, -174.6 72, -172.8 72, -171 72, -169.2 …    18
#> 11 (-180 90, -178.2 90, -176.4 90, -174.6 90, -172.8 90, -171 90, -169.2 …     0