SHmax direction resulting from multiple plate boundaries

Calculates a \(\sigma_{Hmax}\) direction at given coordinates, sourced by multiple plate boundaries. This first-order approximation is the circular mean of the superimposed theoretical directions, weighted by the rotation rates of the underlying PoRs.

Usage

superimposed_shmax(df, PoRs, types, absolute = TRUE, PoR_weighting = NULL)

Arguments

df

data.frame containing the coordinates of the point(s) (lat, lon), and the direction of \(\sigma_{Hmax}\) azi (in degrees)

PoRs

multirow data.frame or "euler.pole" object that must contain lat, lon and angle

types

character vector with length equal to number of rows in PoRs. Type of plate boundary. Must be "out", "in", "right", or "left" for outward, inward, right-lateral, or left-lateral moving plate boundaries, respectively.

absolute

logical. Whether the resultant azimuth should be weighted using the absolute rotation at the points or the angular rotation of the PoRs.

PoR_weighting

(optional) numeric vector with length equal to number of rows in PoRs. Extra weightings for the used PoRs.

Value

numeric. Resultant azimuth in degrees and geographical CRS

See also

Examples

data(san_andreas)
data(nuvel1)
pors <- subset(nuvel1, plate.rot %in% c("eu", "na"))
superimposed_shmax(san_andreas, pors, types = c("in", "right"), PoR_weighting = c(2, 1))
#>    [1] 109.21942 105.14570 105.29746 105.64951 106.30982 105.99733 106.59449
#>    [8] 106.60588 106.67391 106.49172 107.60195 107.62350 106.89011 108.22912
#>   [15] 108.85049 107.18473 108.41171 108.43000 107.36505 108.37538 108.48727
#>   [22] 105.95590 105.32995 105.05988 105.31072 105.26741 105.28205 105.52004
#>   [29] 105.46640 105.56822 106.28700 105.52233 106.44585 105.81804 105.97444
#>   [36] 105.79719 106.81349 106.58299 106.53782 105.97598 106.32228 105.68125
#>   [43] 106.46959 106.76564 106.44727 106.22236 106.56014 106.51451 106.46954
#>   [50] 106.29014 105.96883 106.81003 106.76445 106.53667 106.31245 106.87850
#>   [57] 106.83192 106.60468 106.62665 106.69360 106.60358 107.11657 106.69245
#>   [64] 106.75908 106.71411 106.73569 106.99389 108.35382 106.88792 108.50592
#>   [71] 108.68834 108.10523 108.87066 108.89074 108.74245 108.59518 108.90347
#>   [78] 107.07293 107.09243 108.78846 108.59415 108.75831 108.70947 108.61343
#>   [85] 108.53647 107.45594 108.40106 107.56589 107.52069 107.43000 107.38509
#>   [92] 107.60331 108.04889 106.98642 107.24309 107.41083 108.40049 109.14794
#>   [99] 109.00890 108.87981 105.20211 105.00968 105.24228 105.35503 105.54516
#>  [106] 105.32141 105.12042 105.61257 106.31765 106.22719 106.23982 106.19412
#>  [113] 105.70390 105.47507 105.72721 105.75095 106.23055 105.87284 105.69696
#>  [120] 105.77411 106.74407 106.69844 106.51490 106.46888 105.88560 105.74370
#>  [127] 106.67498 107.06921 106.97554 106.83614 106.42381 107.09162 106.44720
#>  [134] 105.85664 106.55998 106.82356 106.27818 106.83434 106.78769 106.24478
#>  [141] 106.15618 106.41312 106.85665 106.71898 106.67268 106.17857 106.13483
#>  [148] 107.48489 106.92427 106.78637 106.69504 106.55966 106.45743 106.10177
#>  [155] 107.50659 106.94601 106.90028 106.71787 106.58166 106.73969 106.64854
#>  [162] 106.50143 106.67072 106.58006 106.99969 106.90774 106.63537 106.78062
#>  [169] 106.80246 106.65587 106.91483 108.45149 108.39477 107.16928 107.07787
#>  [176] 106.98744 108.84131 108.79132 108.64371 108.54642 108.49712 107.04097
#>  [183] 106.99517 108.81141 108.76225 109.09391 108.97980 108.92934 108.94917
#>  [190] 108.85001 109.43376 108.66302 108.63447 108.59007 109.24509 108.80661
#>  [197] 108.47346 107.54056 107.78890 107.41090 108.70227 107.58480 107.53963
#>  [204] 107.44869 107.51314 108.24405 107.75579 107.06468 107.14143 107.39496
#>  [211] 104.61843 104.68492 109.18257 109.04971 109.08734 104.87359 105.08763
#>  [218] 105.28651 105.45777 105.22949 105.25354 106.11066 105.56394 105.54461
#>  [225] 106.08044 105.63607 106.39939 106.42220 106.50309 106.60580 105.84199
#>  [232] 105.56879 106.85969 105.95353 107.36442 106.88193 106.02014 105.99875
#>  [239] 107.06739 106.11064 106.58316 106.49228 106.17798 106.05609 106.58264
#>  [246] 107.02856 106.74093 106.36825 106.01334 107.55373 107.45956 107.57532
#>  [253] 106.73745 106.89350 106.84838 106.93608 107.71037 107.04795 106.95761
#>  [260] 106.91167 108.55541 106.90901 108.78621 108.56572 108.11795 107.36868
#>  [267] 108.88032 108.76629 109.00261 108.80577 108.64444 107.44888 108.61462
#>  [274] 107.15066 108.88210 108.27394 108.08505 107.34564 108.98356 108.88514
#>  [281] 107.46315 107.41702 108.66486 107.48199 108.16165 108.38280 107.70829
#>  [288] 107.55818 108.55487 108.18647 110.35544 104.98625 107.04100 105.40784
#>  [295] 105.51418 105.32084 106.35432 106.26319 106.21805 105.79491 106.59418
#>  [302] 107.03614 105.41566 106.42351 106.36672 105.65604 106.62959 106.49237
#>  [309] 106.51473 107.13675 106.13409 106.11151 106.95664 106.89841 106.21220
#>  [316] 108.14708 106.84188 106.74036 108.45752 106.86335 108.62441 108.69336
#>  [323] 108.86036 106.79972 108.69266 109.17031 108.97946 107.10041 108.83659
#>  [330] 107.37190 107.39115 107.54724 107.50108 108.51089 108.32153 108.44918
#>  [337] 107.47467 107.49456 108.23609 110.39610 108.32098 107.95433 106.33127
#>  [344] 108.35975 105.92930 106.14945 106.30804 105.92971 106.91797 106.46994
#>  [351] 106.74245 107.42813 105.88069 107.33688 107.00273 108.91991 108.66359
#>  [358] 108.30952 107.26642 108.93479 108.64281 109.11574 107.13990 107.61120
#>  [365] 107.86428 108.05837 107.32051 105.13067 105.19905 108.54996 108.27966
#>  [372] 109.30240 108.87957 106.92915 106.84838 109.48603 106.99763 108.39046
#>  [379] 114.04815 108.75527 109.42961 108.66237 103.41618 106.72085 108.19641
#>  [386] 112.98872 109.55519 108.56460 109.42615 109.57618 112.77471 107.52786
#>  [393] 106.65253 107.11356 108.40005 108.38139 108.63675 108.38457 108.27834
#>  [400] 108.92340 100.53711 110.63940 109.25548 108.75668 109.36141 106.61289
#>  [407] 106.83973 107.46796 108.23817 110.74513 107.51012 104.99090  98.65584
#>  [414]  98.65584  95.47403 100.53711 100.15395 100.58437  99.25156  99.25156
#>  [421]  99.13247  95.20891  98.36966  98.36966  98.22115  97.90407  96.37962
#>  [428] 101.03158  99.32329 110.90244 111.30380 111.82460 106.12249 110.99682
#>  [435] 108.83374 108.43560 111.49893 108.87738 108.50194 108.65109 111.66267
#>  [442] 108.85380 108.45522 107.19482 108.64302 108.42253 108.64310 108.48012
#>  [449] 107.76818 103.94010 108.27232 108.26593 108.60944 107.17345 107.57536
#>  [456] 110.00990 108.64307  97.51008 109.27422 100.72200 108.27525 107.56388
#>  [463] 108.37159 112.41896 109.68729 109.61121 108.77789 105.54526 104.86003
#>  [470] 108.49380 112.60137 111.39259 109.74258 110.10215 103.11680 106.87316
#>  [477] 112.03784 112.16450 110.14728 111.68088 109.68640 111.22647 109.08350
#>  [484] 108.49582 107.63149 109.33730 109.76856 109.01868 108.43070 112.77471
#>  [491] 108.87687 110.17733 109.41961 109.00796 108.77936 108.36381 107.36649
#>  [498] 108.87624 108.87687 109.47911 105.52237 110.08896 109.00754 107.06542
#>  [505] 112.10853 110.83974 107.07724 107.75002 108.81425 108.37209 111.53473
#>  [512] 111.43754 108.71410 108.95808 107.76022 108.46846 108.55495 106.40912
#>  [519] 112.64356 111.04523 103.23423 109.46508 109.72225 105.45540 105.14917
#>  [526] 107.04504 107.76201 107.07659 108.88987 108.71552 106.91715 108.65988
#>  [533] 108.39953 108.45680 108.22953 108.94566 108.40918 108.37873 109.37574
#>  [540] 107.78784 107.84628 110.94859 110.90216 108.87687 111.96995 108.30342
#>  [547] 108.31131 108.87687 109.78231 111.61861 110.05790 108.81340 110.12911
#>  [554] 108.26827 107.57287 108.46465 109.56712 108.98037 107.62198 108.88472
#>  [561] 108.81425 111.37729 107.67912 110.96098 111.74566 111.48246 109.73814
#>  [568] 109.66651 109.19380 108.89241 109.66995 111.94294 113.10401 112.98437
#>  [575] 112.95308 113.05325 108.94990 109.12695 108.79932 108.28912 113.78696
#>  [582] 109.96957 106.89391  95.67736 110.05826 108.62394 108.34411 108.96350
#>  [589] 111.38057 111.54110 111.77359 111.67127 109.77644 106.71511 106.02020
#>  [596] 105.93157 105.83742 105.83132 105.81135 105.97196 105.96741 106.00211
#>  [603] 106.03545 106.10131 106.09622 105.99631 106.12933 106.19270 106.22429
#>  [610] 105.42368 109.68632 109.95517 111.07195 111.23900 113.58187 113.96857
#>  [617] 109.80392 108.54446 110.97801 111.22649 109.54340 105.57092 105.76603
#>  [624] 105.41796 109.21334 112.60176 110.37595 108.92686 110.34598 110.05867
#>  [631] 107.90958 101.30121  94.96085  95.95741  97.41411  97.38821 101.61004
#>  [638]  99.77794  98.58509  98.07635 103.12757 103.12629  96.11467  96.90419
#>  [645]  96.23166  96.13436 104.02722 110.86367 110.33932 112.36478 109.70857
#>  [652] 109.84159 102.52349 102.92278 103.37176 103.06715 109.76602 110.92278
#>  [659] 105.99706 111.90043 111.87071 110.52881 111.94184 111.23583 110.53665
#>  [666] 110.60614 109.81374 108.75525 105.07548 103.47766 105.73299 107.51105
#>  [673] 105.76860 105.63058 103.49083 110.73444 105.99631 105.94608 113.42005
#>  [680] 111.20933 110.19762 111.27170 106.67343 112.53756 109.43458 108.20258
#>  [687] 108.71549 108.77080  96.31235  96.31540  97.28362  97.78331  96.25180
#>  [694] 110.43852 110.98069 110.84234  95.30180  94.89025  96.87392 101.40950
#>  [701] 113.64683 107.03162 110.25718 108.65251  95.65838  96.78635  95.37717
#>  [708]  95.23164  98.04821 104.66942 104.21330 104.38896 106.91160  94.35614
#>  [715]  94.32966  96.50542  97.96777 109.67997 109.65270 109.57606 111.20879
#>  [722] 111.20602 109.89319 100.59627 102.82606  95.03687  97.86983  97.84677
#>  [729]  95.18970  95.21859 109.57721 101.21128 111.30157 105.36058 112.97935
#>  [736] 108.33777  96.17647  96.12516  96.14003  96.13082  96.24251  96.20561
#>  [743]  96.20324  96.20277  96.66319  97.60048  94.77321  96.98147  96.67663
#>  [750]  95.23116  98.90086  98.37247  95.81926  98.43946  98.42291 100.03256
#>  [757]  95.26571  97.60048  97.52700  97.65762 100.05963 100.05225  97.59499
#>  [764] 102.76029  97.68616 102.30901  97.75207 102.51584 102.46055 104.55343
#>  [771]  96.78675  96.93490  95.89902  99.78964  99.77466  96.55788  98.87697
#>  [778]  96.74187  94.91313  99.74665  98.81590  94.87286  98.72422  98.36394
#>  [785]  99.24985  98.12455  99.13278  98.39091  99.29687  99.38589  99.31675
#>  [792]  99.42583  99.37502  99.37335  94.80459  98.36762  99.36296  99.36433
#>  [799]  95.23269  99.14527  98.59610 104.59642 104.57905 104.60933 104.60511
#>  [806] 101.83814 101.60197 104.51400 113.18006  96.05110 102.31603 106.86909
#>  [813]  98.02341  97.99207 107.18408  94.98427  95.23874 105.52749  96.48086
#>  [820]  96.42013  96.35197 101.32315 100.71553 101.58006 101.36711 104.50586
#>  [827] 109.24726 109.24726 109.24726  94.75066  95.08363  97.54974 100.74964
#>  [834] 104.51534 104.43673 104.97518 104.97465 104.03370 104.38073 104.03352
#>  [841] 104.98338 104.98338 104.92334  96.63005 105.05454  95.31745 105.91913
#>  [848] 104.27365 101.83538 104.58946  95.52745  95.61623  95.59718  95.60310
#>  [855]  95.74210  95.69291  95.72298  95.70449  94.38376  96.61903 104.31648
#>  [862]  97.30142 109.23611  98.49441 104.15800  96.31146  96.47455  96.46523
#>  [869]  96.50557  96.34589 111.62089  98.80732  98.85440 100.60139  98.96849
#>  [876]  96.12097  96.26377 110.37964 113.39853  98.06019 102.31131 100.72190
#>  [883]  95.08378 104.37206  96.06533 105.18088 104.95310 102.35143  96.46635
#>  [890]  96.32555  96.45405  94.78460  94.82576  96.22866  98.14225 110.06194
#>  [897]  96.16748 105.15864  95.35317  96.97174  97.22206  97.06813 111.02716
#>  [904] 107.99851 102.10219  95.23677  97.24382 103.80182 112.98444 105.12487
#>  [911] 108.95128 102.33953 102.48071 102.23586 106.89182 101.68467 103.03176
#>  [918]  95.20738  97.11866  99.03174 112.20244 102.56547 102.55747 102.58597
#>  [925] 102.43624 102.51145  94.84895 105.39240  95.59415 110.12155 103.98775
#>  [932] 109.84575  94.83319 113.03543 113.01147 112.94070 112.95330 113.03179
#>  [939] 112.95078 112.94609 113.02109 112.95428 112.96463 112.98552 113.00024
#>  [946] 112.94941 112.98761 112.97792 113.03824 113.01300 112.98868 112.95841
#>  [953] 112.96377 112.95630 112.89771 112.95568 112.98189 112.99018 112.98217
#>  [960] 112.94563 112.95230 113.00770 112.99439 112.95536 112.89807 112.89807
#>  [967] 112.95580 112.99392 113.01461 108.40087 106.64284 107.77233 107.06685
#>  [974] 108.53671 108.43511 108.77010 108.40843 104.54075 108.17402 108.76707
#>  [981] 108.38820 110.03600 110.81505 108.44268 108.40105 107.60931 110.82801
#>  [988] 108.70545 106.85469 108.72803 108.81425 107.59703 104.86003 105.52237
#>  [995] 112.60137 112.03784 112.16450 109.78231 110.07568 110.14728 109.30240
#> [1002] 109.47911 109.33730 111.67474 111.43754 108.46244 110.10215 110.82801
#> [1009] 110.03600 109.61121 109.37574 108.65109 108.28912 109.42961 108.29748
#> [1016] 107.51455 107.60931 107.63149 108.40015 109.02705 108.44200 113.54666
#> [1023] 106.85469 100.25371 104.51157 109.97921 112.66751 107.70627 113.02575
#> [1030] 112.33782 100.16180 109.84260 110.54186 113.17305 110.37205 107.74950
#> [1037] 103.79683 103.40024 104.28513 104.45914 104.53122 108.86054 108.99890
#> [1044] 110.72181 111.48381 111.43135 104.04570 110.82416 112.61565 111.15462
#> [1051] 104.56922 107.48448 110.53608 110.66041 106.58151 105.79740 106.11858
#> [1058] 106.47905 106.45453 110.34231 109.22228 109.64929 109.33655 109.76239
#> [1065] 109.91682 107.06406 110.90912 111.43868 112.15952 113.15235 113.30567
#> [1072] 110.59506 111.74232 110.86939 106.77517 109.64459 107.80560 104.28775
#> [1079] 109.73126 109.62040 107.46655 105.61911