The von Mises Distribution

Density, distribution function, and random generation for the circular normal distribution with mean and kappa.

Usage

rvm(n, mean, kappa)

dvm(theta, mean, kappa)

pvm(theta, mean, kappa, tol = 1e-20)

Arguments

n

number of observations in degrees

mean

mean in degrees

kappa

concentration parameter

theta

angular value in degrees

tol

the precision in evaluating the distribution function or the quantile.

Value

dvm gives the density, pvm gives the distribution function, and rvm generates random deviates.

Examples

x <- rvm(100, mean = 90, k = 100)
dvm(x, mean = 90, k = 100)
#>   [1] 3.72308353 3.93584007 3.29243329 3.95105156 3.46507767 1.99402182
#>   [7] 3.91043121 3.84516413 0.60506341 1.89295338 3.86904206 1.59091410
#>  [13] 3.74203021 3.83986663 3.42781545 3.81122797 2.48082501 3.22404244
#>  [19] 0.03317261 2.95694977 2.62146468 1.19817789 3.76972298 2.10518380
#>  [25] 3.97825934 3.90584235 0.10356087 3.10620749 3.79190017 0.94067252
#>  [31] 2.17059868 2.30632534 3.33523838 3.91410304 3.97606011 1.81685742
#>  [37] 3.15654378 3.86653742 2.90461787 3.02820912 3.54961981 3.36668926
#>  [43] 3.43776796 3.84609894 1.33025453 3.98259781 1.00719276 3.68554943
#>  [49] 3.94339861 3.76542461 3.97948036 3.98004955 3.91216612 3.84603498
#>  [55] 0.71142445 2.36160010 3.26411275 3.64459095 1.81813291 2.88691306
#>  [61] 3.76215188 3.80606288 3.26153655 2.98375882 3.96501439 1.63821851
#>  [67] 3.60957904 2.29111874 2.02449032 2.43320479 2.62110321 0.15170627
#>  [73] 3.98346831 3.30666317 3.31127830 3.25051111 0.73321453 3.53109621
#>  [79] 1.40869161 3.68708646 2.06877735 0.29247455 3.95552963 3.58109663
#>  [85] 3.03493288 3.91357086 1.20998285 3.98409978 3.94683142 2.29879745
#>  [91] 3.55569223 3.96226764 1.70677095 3.04799250 3.88775155 1.95518332
#>  [97] 3.89539167 3.97509474 3.48027852 3.80310288