The von Mises Distribution

Density, probability distribution function, quantiles, and random generation for the circular normal distribution with mean and kappa.

Usage

rvm(n, mean, kappa)

dvm(theta, mean, kappa)

pvm(theta, mean, kappa, from = NULL, tol = 1e-20)

qvm(p, mean = 0, kappa, from = NULL, tol = .Machine$double.eps^(0.6))

Arguments

n

number of observations in degrees

mean

mean in degrees

kappa

concentration parameter

theta

angular value in degrees

from

if NULL is set to \(\mu-\pi\). This is the value from which the pvm and qvm are evaluated. in degrees.

tol

the precision in evaluating the distribution function or the quantile.

p

numeric vector of probabilities with values in \([0,1]\).

Value

dvm gives the density, pvm gives the probability of the von Mises distribution function, rvm generates random deviates (in degrees), and qvm provides quantiles (in degrees).

Examples

x <- rvm(100, mean = 90, kappa = 2)
dvm(x, mean = 90, kappa = 2)
#>   [1] 0.438771107 0.511306894 0.505749886 0.509162374 0.209090693 0.359561777
#>   [7] 0.065641182 0.498427127 0.497203872 0.106384350 0.471629877 0.465844783
#>  [13] 0.223984135 0.414811188 0.401404062 0.458613214 0.458570469 0.258732038
#>  [19] 0.495647725 0.490169743 0.300912669 0.353556191 0.108328089 0.483629698
#>  [25] 0.477009960 0.490098042 0.246420996 0.429545860 0.452998706 0.424601711
#>  [31] 0.437763315 0.489054379 0.010693059 0.512096879 0.103981245 0.112023857
#>  [37] 0.049055100 0.399509389 0.505500143 0.402886811 0.310632344 0.009455589
#>  [43] 0.173086839 0.463913313 0.468031489 0.235324016 0.338386715 0.087357236
#>  [49] 0.498407769 0.402916911 0.403703201 0.221052119 0.504348579 0.509344517
#>  [55] 0.512387133 0.492052643 0.515177199 0.415903009 0.055433653 0.505621685
#>  [61] 0.409941348 0.053605906 0.289656831 0.265147904 0.040706513 0.480000652
#>  [67] 0.449787418 0.245128541 0.467945007 0.034213082 0.221136900 0.070289711
#>  [73] 0.454966494 0.418333337 0.089946945 0.500401241 0.447228323 0.196266434
#>  [79] 0.469732346 0.170118198 0.030739514 0.405224261 0.503251325 0.502935813
#>  [85] 0.242174209 0.390907239 0.224614050 0.433555362 0.510466195 0.470496669
#>  [91] 0.422254741 0.442553405 0.217275273 0.487054437 0.499335173 0.511564875
#>  [97] 0.462952497 0.411267072 0.515885083 0.475854900
pvm(x, mean = 90, kappa = 2)
#>   [1] 0.6981962278 0.4514176721 0.4277490606 0.4411395943 0.8875652932
#>   [6] 0.7803033338 0.9645650425 0.5947682332 0.4019776522 0.9440404400
#>  [11] 0.3494455358 0.6600163196 0.1214679306 0.7264521376 0.2592688601
#>  [16] 0.6710863832 0.3288503900 0.1436287422 0.6020103349 0.3850606762
#>  [21] 0.8270405846 0.7854894948 0.9430452656 0.6286592018 0.6411690778
#>  [26] 0.6150987773 0.1355915404 0.7095550596 0.3208070015 0.7153823043
#>  [31] 0.6994703535 0.6173948006 0.9965156256 0.4558043868 0.9452680454
#>  [36] 0.9411473327 0.0270923684 0.7426761093 0.4268658153 0.2608029082
#>  [41] 0.8198325863 0.0002538419 0.9084104039 0.3369507435 0.6565092273
#>  [46] 0.8714748518 0.2018593028 0.9536871869 0.4051793912 0.2608341607
#>  [51] 0.7383479007 0.8803310172 0.5770751298 0.4419415474 0.4575301168
#>  [56] 0.3893315790 0.5191132148 0.2747535783 0.0303163933 0.5727056840
#>  [61] 0.7317481956 0.9706045627 0.8351634976 0.1479044607 0.9771903002
#>  [66] 0.6356636283 0.6836623179 0.1347600480 0.6566494702 0.9806091023
#>  [71] 0.8802791446 0.0377595484 0.6763959134 0.2774625413 0.9523829905
#>  [76] 0.4107361315 0.6871441744 0.8951433639 0.6537242183 0.9100721998
#>  [81] 0.0175104132 0.7367569526 0.5806499711 0.5816487080 0.1328678955
#>  [86] 0.2486982059 0.1218557719 0.2952999057 0.5528511690 0.3475443877
#>  [91] 0.2819093900 0.6933354730 0.8826332101 0.6216714998 0.5922775816
#>  [96] 0.4528052403 0.3354635380 0.7303193864 0.4995879586 0.3567626640
qvm(c(.25, .5, .75), mean = 90, kappa = 2)
#> [1]  59.65254  90.00000 120.34746